### 2022-03-11 #4241 @editorialbot
[PRE REVIEW]: FlowSieve: A Coarse-Graining Utility for Geophysical Flows on the Sphere
**Submitting author:** @bastorer (<a href="http://orcid.org/0000-0001-5955-2158">Benjamin Storer</a>)
**Repository:** https://github.com/husseinaluie/FlowSieve
**Branch with paper.md** (empty if default branch):
**Version:** v3.0.0
