### 2022-03-11 073dbc2
Creating new RTD landing page file
Previously used READE.md as ReadTheDoc main page. This caused unpleasant
formatting / display on the GitHub page. Previous READE contents are
now in RTD_mainpage.md. New README is a GitHub-targeted file.
