+++ README.md
@@ -1,5 +1,6 @@
-[TOC]
 # FlowSieve
+
+[![Documentation Status](https://readthedocs.org/projects/flowsieve/badge/?version=latest)](https://flowsieve.readthedocs.io/en/latest/?badge=latest)
 
 **FlowSieve** is developed as an open resource by the Complex Flow Group at the University of Rochester, under the sponsorship of the National Science Foundation and the National Aeronautics and Space Administration. 
 Continued support for **FlowSieve** depends on demonstrable evidence of the code’s value to the scientific community. 
@@ -15,122 +16,11 @@ For journal articles, proceedings, etc.., we suggest:
 For presentations, posters, etc.., we suggest acknowledging:
 * **FlowSieve** code from the Complex Flow Group at University of Rochester
 
-
 ---
 
-## Tutorial
-
-A series of [basic tutorials](\ref tutorials1) are provided to outline both various usage cases as well as how to use / process the outputs.
-
+[**FlowSieve** documentation is available here.](https://flowsieve.readthedocs.io/en/latest/) 
 
 ---
-
-## Methods
-
-Some details regarding underlying methods are discussed [on this page](\ref methods1) (warning, math content).
-
-### Helmholtz Decomposition
-
-For notes about the Helmholtz decomposition, [go to this page](\ref helmholtz1).
-
----
-
-## Compilation / Installation
-
-For notes on installation, please see [this page](\ref install1).
-
----
-
-## Input / Output Files and Filetypes
-
-The coarse-graining codebase uses netcdf files for both input and output.
-Dimension orderings are assumed to follow the CF-convention of (time, depth, latitude, longitude).
-
-*scale_factor*,  *offset*, and *fill_value* attributes are applied to output variables following CF-convention usage.
-
-Where possible, units and variables descriptions (which are provided in constants.hpp) are also include as variable attributes in the output files.
-
-Currently, no other filetypes are supported.
-
----
-
-## Postprocessing
-
-Post-processing (such as region-averaging, Okubo-Weiss histogram binning, time-averaging, etc) can be enabled and run on-line
-by setting the **APPLY_POSTPROCESS** flag in constants.hpp to **true**.
-
-This will produce an additional output file for each filtering scale.
-
-Various geographic regions of interest can be provided in a netcdf file.
-
----
-
-## Command-line Arguments
-
-1. `--version`
- * Calling `./coarse_grain.x --version` prints a summary of the constants / variables used when compiling
-
-
-### Specifying Filtering Scales
-
-When specifying filtering scales, consider a wide sweep. It can also be beneficial to use logarithmically-spaced scales, for plotting purposes.
-Python can be helpful for this. For example, `numpy.logspace( np.log10(50e3), np.log10(2000e3), 10 )` would produce 10 logarithmically-spaced
-filter scales between 50km and 2000km.
-
-Hint: to print filter scales to only three significant digits, the `numpy.format_float_scientific` function can help.
-~~~~~~~~~~~~~~{.py}
-import numpy as np
-
-number_of_scales = 10
-
-smallest_scale = 50e3
-largest_scale  = 2000e3
- 
-scales = np.logspace( np.log10(smallest_scale), np.log10(largest_scale), number_of_scales )
- 
-[print( np.format_float_scientific( scale, precision = 2 ), end = ' ' ) for scale in scales]
-~~~~~~~~~~~~~~
-
-If you are using a bash script (e.g. a job-submission script), an easy way to pass the filter scales on to the coarse-graining executable is to define 
-a variable that has the list of scales, and then just pass that to the executable using the **--filter_scales** flag.
-~~~~~~~~~~~~{.sh}
-FILTER_SCALES="1e4 5e4 10e4"
-
---filter_scales "${FILTER_SCALES}"
-~~~~~~~~~~~~
-
----
-
[…]
