+++ .github/workflows/draft-pdf.yml
@@ -12,8 +12,7 @@ jobs:
         with:
           journal: joss
           # This should be the path to the paper within your repo.
-          paper-path: Documentation/JOSS/paper.md
-          bib-path: Documentation/JOSS/paper.bib
+          paper-path: paper.md
       - name: Upload
         uses: actions/upload-artifact@v1
         with:
@@ -21,4 +20,4 @@ jobs:
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: Documentation/JOSS/paper.pdf
+          path: paper.pdf
