### 2022-03-18 #4241 @bastorer
> Hi @bastorer! Thanks for your submission. Your references aren't coming through, in line or at the end of your paper. Can you take a look and try to fix this?

Hi @kthyng Thanks for catching that! I've updated the files and yaml action script and the compiled paper includes the references now :-)
