### 2022-03-21 #4241 @kthyng
@bastorer Can you take a look at this list of reviewer volunteers and suggest people that you think would fit this submission? Please don't use `@`, just list their handles. https://bit.ly/joss-reviewers

Also as implied by @danielskatz, can you take a look at the [paper requirements](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) and make sure you've covered all the necessary information? This will come up during review if it isn't handled now, so might as well deal with it now if possible.
