### 2022-03-21 #4241 @bastorer
Hi @kthyng

NoraLoose would be a good fit topic-wise ( they use similar methodologies in their work ). Comparing domains/topics, ali-ramadhan and ashwinvis look like they could also be a good fit.

I think I covered the paper requirements, but I agree that the paper is rather short. Is it okay if I update the paper tomorrow and send you a reply once that's done before you send it off to reviewers?

Thanks!
