### 2022-03-21 #4241 @kthyng
@bastorer Probably the review process won't get underway so quickly, so please go ahead and update as you suggest. And note that the paper indeed isn't meant to be very long — just long enough to cover what is necessary. I'll reach out to potential reviewers in the meantime.
