### 2022-03-28 #4241 @kris-rowe
> Hi @kris-rowe! Are you interested in reviewing this JOSS submission? We would ask for reviews within 4–6 weeks, but that generally means starting the process sooner since it tends to be iterative between the reviewers and author(s). You can read more about the process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
> Your skills look like a good complement to the other reviewer and for the submission. Thanks for your consideration!

I can be reviewer on this. Sounds like an interesting paper. Happy to help.
