### 2022-03-28 #4277 @editorialbot
[REVIEW]: FlowSieve: A Coarse-Graining Utility for Geophysical Flows on the Sphere
**Submitting author:** @bastorer (<a href="http://orcid.org/0000-0001-5955-2158">Benjamin Storer</a>)
**Repository:** https://github.com/husseinaluie/FlowSieve
**Branch with paper.md** (empty if default branch):
**Version:** v3.3.0
