### 2022-04-12 #4277 @danielskatz
While @kthyng is ooo (or ideally until we finish this review), I'll take over as editor.

It looks like everything is started:
* @NoraLoose needs to generate a checklist (with the command `@editorialbot generate my checklist`) and then get started
* @kris-rowe has already started their review

If you have any questions, please let me know!
