### 2022-04-15 husseinaluie/FlowSieve#5 @NoraLoose
OSI approved license
JOSS requires the LICENSE file to have the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license. I can't find the ODC-BY 1.0 license among the licenses in the OSI approved list.

This issue is part of the JOSS review process over [here](https://github.com/openjournals/joss-reviews/issues/4277).
