### 2022-04-15 husseinaluie/FlowSieve#6 @NoraLoose
Documentation: Methods
There are a few inconsistencies and typos in the method section of the documentation, and it would be helpful to give it a bit more structure (via section headers). Specifically:

- [x] Two different version of phi are used interchangeably. It would make the documentation clearer to stick to only one of the 2 symbols.
- [x] What is "L_pad"?
- [x]  "currently implementation" --> "current implementation"
- [x] "An identical equation":  identical to what? It may be clearer to label the equation for Delta phi_N above, and than you could write: "An identical equation to [labeled equation]".
- [x] Before writing down the large-scale KE budget, it may be helpful to start a new section with a new section header, e.g.,  "Large-scale KE budget". It would also be helpful to refer the user to e.g., Aluie et al. (2018) for the derivation.
- [x] New section header before you introduce the Lamba^m equation.
- [x] It would be helpful to introduce a new section header before the sentence "This section outlines some of the parallelizations that are used.", e.g., "Parallelizing computations"
- [x] reoutines --> routines

This issue is part of the JOSS review process over https://github.com/openjournals/joss-reviews/issues/4277.
