### 2022-04-15 husseinaluie/FlowSieve#7 @NoraLoose
Documentation: Tree structure
There is something funky going on with the documentation tree. The red circled parts don't seem to be actual subsections of the "Installation" section (but rather belong to the License text etc.).

This issue is part of the JOSS review process over https://github.com/openjournals/joss-reviews/issues/4277.
