### 2022-04-15 husseinaluie/FlowSieve#8 @NoraLoose
Community guidelines
JOSS asks to include
> clear community guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.

Could you please add community guidelines to this effect? I think the README would be a great place for that.

This issue is part of the JOSS review process over https://github.com/openjournals/joss-reviews/issues/4277.
