### 2022-04-15 husseinaluie/FlowSieve#9 @NoraLoose
Software paper
JOSS requires the software paper to include the following elements:

- [x] **Summary** of software: While your summary section is a great motivation for why coarse graining is needed, it does not describe the high-level functionality and purpose of the **software** (but rather of the technique coarse graining in general). Could you adapt the summary to be more specific to the software package FlowSieve?
- [ ] **State of the field**: Could you include a discussion on how FlowSieve compares to other commonly-used packages?

This issue is part of the JOSS review process over https://github.com/openjournals/joss-reviews/issues/4277.
