### 2022-05-10 #4277 @danielskatz
👋 @bastorer - note in particular that https://github.com/husseinaluie/FlowSieve/issues/5 is a showstopper here.  The license you have chosen is

1. not a software license but a data license
2. not OSI-approved

JOSS cannot publish papers about software that does not have an OSI-approved license.
