### 2022-06-14 d82c080
Update coarse_grain_helmholtz.cpp
Add assert banning EXTEND_DOMAIN_TO_POLES with CARTESIAN
