### 2022-06-23 a9e0462
Documentation updates, towards JOSS Section headings in documentation don't seem to render correctly. Testing.
