+++ Documentation/METHODS.md
@@ -19,29 +19,32 @@ This is implemented in **Functions/get_lat_bounds.cpp**
 
 ##### Uniform grid spacing
 
-Suppose that we have uniform (spherical) grid spacing \f$\Delta\varphi\f$. 
+Suppose that we have uniform (spherical) grid spacing \f$\Delta\phi\f$. 
 
-The spacing between latitude bands, in *metres*, is then \f$\Delta\varphi_m=\Delta\varphi R_E\f$.
-The number of points that we'll include in either latitude side is then \f$\Delta\phi_N=\mathrm{ceil}\left( L^{pad} (L/2) / \Delta\phi_m\right)\f$
-The scale factor \f$L^{pad}\f$ is a user-specified scaling to indicate how large of an integration region should be used. This is specified as *KernPad* in constants.hpp.
+The spacing between latitude bands, in *metres*, is then \f$\Delta\phi_m=\Delta\phi R_E\f$.
+The number of points that we'll include in either latitude side is then \f$\Delta\phi_N=\mathrm{ceil}\left( L^{pad} (L/2) / \Delta\phi_m\right)\f$.
+The scale factor \f$L^{pad}\f$ is a user-specified scaling to indicate how large of an integration region should be used. 
+This is specified as *KernPad* in constants.hpp.
+That is, the numerical integration to compute the local coarse-grained value only includes points within a distance of \f$ L^{pad}\ell \f$ (in metres) from the target spatial point, so that \f$ L^{pad}=1 \f$ indicates that the integration area has a diameter of exactly \f$ \ell \f$. 
+This is implemented in the functions get_lon_bounds and get_lat_bounds, which determine the physical and logical integration bounds for a given point and filter scale.
 
-The outer loop in the integral is then from \f$I_{\varphi_0} - \Delta\varphi_N\f$ to \f$I_{\varphi_0}+\Delta\varphi_N\f$.
+The outer loop in the integral is then from \f$I_{\phi_0} - \Delta\phi_N\f$ to \f$I_{\phi_0}+\Delta\phi_N\f$.
 
 ##### Non-uniform grid spacing
 
 In this event, we simply use a binary search routine to search the (logical) interval 
-\f$ [0, I_{\varphi_0}-1 ] \f$ for the point whose distance from the reference point is nearest to, but not less than, \f$ L^{pad} L/2  \f$
+\f$ [0, I_{\phi_0}-1 ] \f$ for the point whose distance from the reference point is nearest to, but not less than, \f$ L^{pad} L/2  \f$
 
 #### Restricting the longitudinal (inner) loop
 
 This is implemented in **Functions/get_lon_bounds.cpp**
 
-*The currently implementation requires the longitudinal grid to be uniform.*
+*The current implementation requires the longitudinal grid to be uniform.*
 
 Suppose that we have uniform (spherical) grid spacing \f$\Delta\lambda\f$. 
 
-Next, at each latitude \f$\varphi\f$ within that loop, the same process is applied to compute the bounds for the longitude loop, with
-the spacing between longitude band, in *metres*, given by \f$\Delta\lambda_m=\Delta\lambda R_E \cos(\varphi)\f$.
+Next, at each latitude \f$\phi\f$ within that loop, the same process is applied to compute the bounds for the longitude loop, with
+the spacing between longitude band, in *metres*, given by \f$\Delta\lambda_m=\Delta\lambda R_E \cos(\phi)\f$.
 
[…]
