+++ .readthedocs.yaml
@@ -1,0 +1,15 @@
+version: 2
+
+build:
+  tools:
+    python: "3.7"
+
+sphinx:
+  configuration: docs/sphinx/conf.py
+
+python:
+  install:
+  - requirements: docs/sphinx/requirements.txt
+
+conda:
+  environment: docs/sphinx/environment.yml
