+++ .readthedocs.yaml
@@ -1,8 +1,8 @@
 version: 2
 
-build:
-  tools:
-    python: "3.7"
+#build:
+#  tools:
+#    python: "3.7"
 
 sphinx:
   configuration: docs/sphinx/conf.py
