+++ Documentation/METHODS.md
@@ -1,5 +1,4 @@
-Methods {#methods1}
-===================
+# Methods {#methods1}
 [TOC]
 
 \brief A review of the computational methodologies (warning: math content)
