+++ Documentation/METHODS.md
@@ -1,5 +1,9 @@
-[TOC]
 # Methods {#methods1}
+[TOC]
+
+---
+
+A review of the computational methodologies (warning: math content).
 
 ## Coarse-Grain Filtering
 
