+++ Documentation/METHODS.md
@@ -1,9 +1,8 @@
 # Methods {#methods1}
-[TOC]
-
----
 
 A review of the computational methodologies (warning: math content).
+
+---
 
 ## Coarse-Grain Filtering
 
