+++ licence.md
@@ -1,7 +1,9 @@
 # ODC Attribution License (ODC-By) {#license1}
 [TOC]
 
-\brief The ODC Attribution License (ODC-By)
+The ODC Attribution License (ODC-By)
+
+---
 
 ### Preamble
 
