+++ docs/sphinx/environment.yml
@@ -1,7 +1,7 @@
 name: RTD
-channels:
- - conda-forge
- - defaults
-dependencies:
- - python=3.7
- - doxygen=1.9
+#channels:
+# - conda-forge
+# - defaults
+#dependencies:
+# - python=3.7
+# - doxygen=1.9
