+++ licence.md
@@ -1,7 +1,10 @@
 # ODC Attribution License (ODC-By) {#license1}
 [TOC]
 
+## License
+
 The ODC Attribution License (ODC-By)
+
 
 ---
 
