+++ Documentation/ISSUES.md
@@ -1,9 +1,11 @@
 # Solutions to Common Problems {#issues1}
 [ TOC]
 
----
+# Known Issues
 
 A description of some know issues / bugs, with solutions where available.
+
+---
 
 ## I/O
 
