### 2022-06-23 3c0d99a
Adjusting licence header settings
