### 2022-06-23 fefe5a0
Finalizing some documentation updates
[#7] Tree structure: Thanks for pointing this out! It looks like there were some cascading errors that caused headers to not render properly / get jumbled around.

[#6] Methods: with the header issues resolved, this page should be much easier to follow now. Contents updated in response to feedback. Thanks!
