### 2022-06-23 fe9bdea
Minor documentation shuffling
