### 2022-06-23 c344d5b
Adding community guidelines
[#8] community guidelines added to README and main RTD landing page
