### 2022-06-23 6c9ff52
Changing to Open Software License 3.0 (OSL-3.0)
[#5] changed to OSI-approved license
