+++ licence.md
@@ -1,436 +1,53 @@
-# ODC Attribution License (ODC-By) {#license1}
+# Open Software License 3.0 {#license1}
 
-\brief The ODC Attribution License (ODC-By)
+\brief The Open Software License 3.0 (OSL-3.0)
 
 # License {#license1-0}
-[TOC]
-
 
 ---
 
-## Preamble
-
-The Open Data Commons Attribution License is a license agreement
-intended to allow users to freely share, modify, and use this Database
-subject only to the attribution requirements set out in Section 4.
-
-Databases can contain a wide variety of types of content (images,
-audiovisual material, and sounds all in the same database, for example),
-and so this license only governs the rights over the Database, and not
-the contents of the Database individually. Licensors may therefore wish
-to use this license together with another license for the contents.
-
-Sometimes the contents of a database, or the database itself, can be
-covered by other rights not addressed here (such as private contracts,
-trademark over the name, or privacy rights / data protection rights
-over information in the contents), and so you are advised that you may
-have to consult other documents or clear other rights before doing
-activities not covered by this License.
+This Open Software License (the "License") applies to any original work of authorship (the "Original Work") whose owner (the "Licensor") has placed the following licensing notice adjacent to the copyright notice for the Original Work:
 
-------
+Licensed under the Open Software License version 3.0
 
-The Licensor (as defined below) 
+1) Grant of Copyright License. Licensor grants You a worldwide, royalty-free, non-exclusive, sublicensable license, for the duration of the copyright, to do the following:
 
-and 
+a) to reproduce the Original Work in copies, either alone or as part of a collective work;
 
-You (as defined below) 
+b) to translate, adapt, alter, transform, modify, or arrange the Original Work, thereby creating derivative works ("Derivative Works") based upon the Original Work;
 
-agree as follows: 
+c) to distribute or communicate copies of the Original Work and Derivative Works to the public, with the proviso that copies of Original Work or Derivative Works that You distribute or communicate shall be licensed under this Open Software License;
 
-## 1.0 Definitions of Capitalised Words
+d) to perform the Original Work publicly; and
 
-"Collective Database" - Means this Database in unmodified form as part
-of a collection of independent databases in themselves that together are
-assembled into a collective whole. A work that constitutes a Collective
-Database will not be considered a Derivative Database.
+e) to display the Original Work publicly.
 
-"Convey" - As a verb, means Using the Database, a Derivative Database,
-or the Database as part of a Collective Database in any way that enables
-a Person to make or receive copies of the Database or a Derivative
-Database.  Conveying does not include interaction with a user through a
-computer network, or creating and Using a Produced Work, where no
-transfer of a copy of the Database or a Derivative Database occurs.
-
-"Contents" - The contents of this Database, which includes the
-information, independent works, or other material collected into the
-Database. For example, the contents of the Database could be factual
-data or works such as images, audiovisual material, text, or sounds.
-
-"Database" - A collection of material (the Contents) arranged in a
-systematic or methodical way and individually accessible by electronic
-or other means offered under the terms of this License.
+2) Grant of Patent License. Licensor grants You a worldwide, royalty-free, non-exclusive, sublicensable license, under patent claims owned or controlled by the Licensor that are embodied in the Original Work as furnished by the Licensor, for the duration of the patents, to make, use, sell, offer for sale, have made, and import the Original Work and Derivative Works.
 
-"Database Directive" - Means Directive 96/9/EC of the European
-Parliament and of the Council of 11 March 1996 on the legal protection
-of databases, as amended or succeeded.
[…]
