### 2022-06-23 693ec6b
Updating README to mention correct license.
[#5] updating README reference
