+++ README.md
@@ -7,7 +7,7 @@
 **FlowSieve** is developed as an open resource by the [Complex Flow Group at the University of Rochester](http://www.complexflowgroup.com), under the sponsorship of the National Science Foundation and the National Aeronautics and Space Administration. 
 Continued support for **FlowSieve** depends on demonstrable evidence of the code’s value to the scientific community. 
 We kindly request that you cite the code in your publications and presentations. 
-**FlowSieve** is made available under the Open Data Commons Attribution License (ODC-By) (http://opendatacommons.org/licenses/by/, see [the license file](\ref license1) or the human-readable summary at the end of the README), which means it is open to use, but requires attribution. 
+**FlowSieve** is made available under The Open Software License 3.0 (OSL-3.0) (see [the license file](\ref license1) or the human-readable summary at the end of the README), which means it is open to use, but requires attribution. 
 
 The following citations are suggested:
 
@@ -39,7 +39,7 @@ For presentations, posters, etc.., we suggest acknowledging:
 
 ## Licence
 
-This is a brief human-readable summary of the ODC-BY 1.0 license, and is not the actual licence. 
+This is a brief human-readable summary of the license, and is not the actual licence. 
 See `licence.md` for the full licence details.
 
 You are free:
