### 2022-06-29 #4277 @kris-rowe
@danielskatz Should I continue to work on my review for this, or should I wait for the license issue to be resolved?
