### 2022-07-21 d4e56ac
Some paper additions, testing compilation
