### 2022-07-21 c0a393f
Update paper.md
(Hopefully) fixing some lists
