### 2022-07-21 15d1b68
Update paper.md
Couldn't get lists to build, so changing to numbering in-paragraph.
