### 2022-07-22 6a4aff4
Adding 'Summary of Software' and 'State of the Field' to paper.md
[#9] Adds in the missing sections :-)
