### 2022-07-22 #4277 @bastorer
Hi @danielskatz

I'm very sorry about the slow responses. It also looks like my in-line email responses didn't show up like I thought they would, but I was also very slow to respond.

I've submitted commits addressing the issues posted on the FlowSieve repository, including changing the license to one on the OSI-approved list.
