### 2022-07-22 husseinaluie/FlowSieve#5 @bastorer
License has been changed to the Open Software License 3 (OLS-3.0, https://opensource.org/licenses/OSL-3.0)
