### 2022-07-22 husseinaluie/FlowSieve#9 @bastorer
I've updated the paper with "Summary of Software" and "State of the Field" sections added.
