### 2022-07-22 husseinaluie/FlowSieve#8 @bastorer
Community guidelines are now part of the readme and the documentation landing page
