### 2022-07-25 husseinaluie/FlowSieve#7 @husseinaluie
Hi Ben,

What are the next steps with the JOSS paper? Is everything done and
awaiting approval from the "editor"? Or will it go back to the reviewers
for another round?

Thanks,
Hussein

Hussein Aluie (he/him)
Associate Professor
Department of Mechanical Engineering
University of Rochester
Office: 585-276-7170
Cell: 410-340-3145
http://www.me.rochester.edu/~haluie/

On Fri, Jul 22, 2022 at 3:54 PM Benjamin Storer ***@***.***>
wrote:

> I managed to hunt down the silent cascading error that caused this, so the
> tree structure should now resemble the actual documentation structure!
>
