### 2022-07-25 4cbe175
Removing UiUj decomposition parts from Tutorial
It's somewhat messy, and not strictly necessary, so it's
best covered in a separate tutorial if we want it.
