### 2022-08-22 husseinaluie/FlowSieve#9 @NoraLoose
Thanks for adding these sections! I think they will really help the reader understand better what FlowSieve is all about.

It would be nice if you could add a citation for the GCM-Filters package. :)
