### 2022-08-22 husseinaluie/FlowSieve#5 @NoraLoose
Sorry, will re-open this because I found an inconsistency regarding the license: In the documentation it says you use the ODC-BY 1.0 license. Should be an easy fix to change it to OSL-3.0!
