### 2022-08-23 110cf68
Fixing licence reference
#5 : Corrects licence name reference
