### 2022-08-23 41ad25a
Updating paper with additional references
