+++ paper.bib
@@ -26,3 +26,27 @@
     year = {2019}
 }
 
+@article{Grooms2021,
+    author = {Grooms, I. and Loose, N. and Abernathey, R. and Steinberg, J. M. and Bachman, S. D. and Marques, G. and Guillaumin, A. P. and Yankovsky, E.},
+    doi = {10.1029/2021MS002552},
+    issn = {19422466},
+    journal = {Journal of Advances in Modeling Earth Systems},
+    keywords = {coarse graining,data analysis,spatial filtering},
+    number = {9},
+    pages = {1--24},
+    title = {{Diffusion-Based Smoothers for Spatial Filtering of Gridded Geophysical Data}},
+    volume = {13},
+    year = {2021}
+}
+
+@article{Balwada2022,
+    archivePrefix = {arXiv},
+    arxivId = {2202.08637},
+    author = {Balwada, Dhruv and Xie, Jin-Han and Marino, Raffaele and Feraco, Fabio},
+    eprint = {2202.08637},
+    month = {feb},
+    title = {{Direct observational evidence of an oceanic dual kinetic energy cascade and its seasonality}},
+    url = {http://arxiv.org/abs/2202.08637},
+    year = {2022}
+}
+
