### 2022-08-23 husseinaluie/FlowSieve#9 @bastorer
Absolutely, I'm not sure why I didn't include the reference to begin with, sorry!

I've now included a reference to GCM-Filters, as well as a reference to the structure function approach ( via Dhruv's cascade paper on arxiv ).
