### 2022-08-23 594b1d0
[re]introducing IO checks on read from netcdf version Reading in netcdf classic files causes crash errors. Trying to check if a netcdf file is v3 or v4 can be buggy. Hopefully, the introduced system is a reliable test.
