### 2022-08-23 f6fd723
Updating some print statements
