### 2022-08-23 ff04c16
Adding in particle option for single time point If the velocity field only has a single time, then the particle code will assume a time-constant velocity field and evolve particles accordingly. This essentially allows for streamlines
