### 2022-08-23 e25bc47
Adding in "--help" functionality. Basic tools set up. Not yet incorporated into CaseFiles (next step). Still room for improvement
