### 2022-08-23 413ff4a
Adding basic "--help" functionality, and option to read in wind data
