+++ Case_Files/coarse_grain_helmholtz.cpp
@@ -57,41 +57,48 @@ int main(int argc, char *argv[]) {
         if (wRank == 0) { print_compile_info(NULL); } 
         return 0;
     }
+    const bool asked_help = input.cmdOptionExists("--help");
+    if (asked_help) {
+        fprintf( stdout, "The command-line input arguments [and default values] are:\n" );
+    }
 
     // first argument is the flag, second argument is default value (for when flag is not present)
-    //const std::string   &tor_input_fname   = input.getCmdOption("--toroidal_input_file",        "toroidal_projection.nc"),
-    //                    &pot_input_fname   = input.getCmdOption("--potential_input_file",       "potential_projection.nc"),
-    const std::string   &Helm_input_fname  = input.getCmdOption("--Helmholtz_input_file",       "Helmholtz_projection.nc"),
-                        &vel_input_fname   = input.getCmdOption("--velocity_input_file",        "toroidal_projection.nc"),
-                        &quad_input_fname  = input.getCmdOption("--uiuj_Helmholtz_input_file",  "helmholtz_projection_uiuj.nc");
+    const std::string   &Helm_input_fname  = input.getCmdOption("--Helmholtz_input_file",       "Helmholtz_projection.nc",      asked_help),
+                        &vel_input_fname   = input.getCmdOption("--velocity_input_file",        "vels.nc",                      asked_help),
+                        &wind_input_fname  = input.getCmdOption("--wind_tau_input_file",        "wind_tau_projection.nc",       asked_help),
+                        &quad_input_fname  = input.getCmdOption("--uiuj_Helmholtz_input_file",  "helmholtz_projection_uiuj.nc", asked_help);
 
-    const std::string   &time_dim_name      = input.getCmdOption("--time",        "time"),
-                        &depth_dim_name     = input.getCmdOption("--depth",       "depth"),
-                        &latitude_dim_name  = input.getCmdOption("--latitude",    "latitude"),
-                        &longitude_dim_name = input.getCmdOption("--longitude",   "longitude");
+    const std::string   &time_dim_name      = input.getCmdOption("--time",        "time",       asked_help),
+                        &depth_dim_name     = input.getCmdOption("--depth",       "depth",      asked_help),
+                        &latitude_dim_name  = input.getCmdOption("--latitude",    "latitude",   asked_help),
+                        &longitude_dim_name = input.getCmdOption("--longitude",   "longitude",  asked_help);
 
-    const std::string &latlon_in_degrees  = input.getCmdOption("--is_degrees",   "true");
+    const std::string &latlon_in_degrees  = input.getCmdOption("--is_degrees",   "true", asked_help);
 
-    const std::string   &Nprocs_in_time_string  = input.getCmdOption("--Nprocs_in_time",  "1"),
-                        &Nprocs_in_depth_string = input.getCmdOption("--Nprocs_in_depth", "1");
+    const std::string   &Nprocs_in_time_string  = input.getCmdOption("--Nprocs_in_time",  "1", asked_help),
+                        &Nprocs_in_depth_string = input.getCmdOption("--Nprocs_in_depth", "1", asked_help);
     const int   Nprocs_in_time_input  = stoi(Nprocs_in_time_string),
                 Nprocs_in_depth_input = stoi(Nprocs_in_depth_string);
 
-    const std::string   &tor_field_var_name     = input.getCmdOption("--tor_field",     "Psi"),
-                        &pot_field_var_name     = input.getCmdOption("--pot_field",     "Phi"),
-                        &vel_field_var_name     = input.getCmdOption("--vel_field",     "u_lat"),
[…]
