### 2022-08-23 c6f8784
Cleaning out old print statements
