### 2022-09-05 husseinaluie/FlowSieve#10 @NoraLoose
Installation
I think it would be helpful for many (possibly inexperienced) users if the first step in the installation instructions would be:

- [x] To clone the FlowSieve repository (or a fork) onto the HPC system they want to use. You could also provide a link that explains how to generally clone / fork GitHub repos.

This issue is part of the JOSS review process over https://github.com/openjournals/joss-reviews/issues/4277.
