+++ paper.bib
@@ -39,6 +39,19 @@
     year = {2021}
 }
 
+@article{Loose2022,
+   author = {Nora Loose and Ryan Abernathey and Ian Grooms and Julius Busecke and Arthur Guillaumin and Elizabeth Yankovsky and Gustavo Marques and Jacob Steinberg and Andrew Slavin Ross and Hemant Khatri and Scott Bachman and Laure Zanna and Paige Martin},
+   title = {GCM-Filters: A Python Package for Diffusion-based Spatial Filtering of Gridded Data},
+   journal = {Journal of Open Source Software},
+   volume = {7},
+   number = {70},
+   pages = {3947},
+   doi = {10.21105/joss.03947},
+   url = {https://doi.org/10.21105/joss.03947},
+   year = {2022},
+   publisher = {The Open Journal},
+}
+
 @article{Balwada2022,
     archivePrefix = {arXiv},
     arxivId = {2202.08637},
@@ -50,3 +63,13 @@
     year = {2022}
 }
 
+@book{Frisch1995,
+    author = {Frisch, Uriel},
+    doi = {10.1017/CBO9781139170666},
+    isbn = {9780521451031},
+    month = {nov},
+    publisher = {Cambridge University Press},
+    title = {{Turbulence: The Legacy of A.N. Kolmogorov}},
+    year = {1995}
+}
+        
