### 2022-09-15 2e9a255
Minor improvements to INSTALL documentation
