### 2022-09-15 husseinaluie/FlowSieve#10 @bastorer
Good suggestion! Documentation updated accordingly.
