### 2022-09-16 74e3779
Adding '--help' to some more executables
