+++ Case_Files/coarsen_grid_linear.cpp
@@ -42,28 +42,34 @@ int main(int argc, char *argv[]) {
         if (wRank == 0) { print_compile_info(NULL); } 
         return 0;
     }
+    const bool asked_help = input.cmdOptionExists("--help");
+    if (asked_help) {
+        fprintf( stdout, "The command-line input arguments [and default values] are:\n" );
+    }
 
     // first argument is the flag, second argument is default value (for when flag is not present)
-    const std::string   &coarse_fname   = input.getCmdOption("--coarse_file",   "coarse.nc"),
-                        &fine_fname     = input.getCmdOption("--fine_file",     "fine.nc"),
-                        &output_fname   = input.getCmdOption("--output_file",   "coarse_vel.nc");
+    const std::string   &coarse_fname   = input.getCmdOption("--coarse_file",   "coarse.nc",        asked_help),
+                        &fine_fname     = input.getCmdOption("--fine_file",     "fine.nc",          asked_help),
+                        &output_fname   = input.getCmdOption("--output_file",   "coarse_vel.nc",    asked_help);
 
-    const std::string   &time_dim_name      = input.getCmdOption("--time",        "time"),
-                        &depth_dim_name     = input.getCmdOption("--depth",       "depth"),
-                        &latitude_dim_name  = input.getCmdOption("--latitude",    "latitude"),
-                        &longitude_dim_name = input.getCmdOption("--longitude",   "longitude");
+    const std::string   &time_dim_name      = input.getCmdOption("--time",        "time",       asked_help),
+                        &depth_dim_name     = input.getCmdOption("--depth",       "depth",      asked_help),
+                        &latitude_dim_name  = input.getCmdOption("--latitude",    "latitude",   asked_help),
+                        &longitude_dim_name = input.getCmdOption("--longitude",   "longitude",  asked_help);
 
-    const std::string &latlon_in_degrees  = input.getCmdOption("--is_degrees",   "true");
+    const std::string &latlon_in_degrees  = input.getCmdOption("--is_degrees",   "true", asked_help);
 
-    const std::string   &Nprocs_in_time_string  = input.getCmdOption("--Nprocs_in_time",  "1"),
-                        &Nprocs_in_depth_string = input.getCmdOption("--Nprocs_in_depth", "1");
+    const std::string   &Nprocs_in_time_string  = input.getCmdOption("--Nprocs_in_time",  "1", asked_help),
+                        &Nprocs_in_depth_string = input.getCmdOption("--Nprocs_in_depth", "1", asked_help);
     const int   Nprocs_in_time_input  = stoi(Nprocs_in_time_string),
                 Nprocs_in_depth_input = stoi(Nprocs_in_depth_string);
 
     std::vector< std::string > vars_to_refine, vars_in_output;
-    input.getListofStrings( vars_to_refine, "--input_variables" );
-    input.getListofStrings( vars_in_output, "--output_variables" );
+    input.getListofStrings( vars_to_refine, "--input_variables", "",  asked_help );
+    input.getListofStrings( vars_in_output, "--output_variables", "", asked_help );
     const int Nvars = vars_to_refine.size();
+
+    if (asked_help) { return 0; }
 
     // Print processor assignments
     const int max_threads = omp_get_max_threads();
