+++ Case_Files/Helmholtz_projection.cpp
@@ -47,43 +47,49 @@ int main(int argc, char *argv[]) {
         if (wRank == 0) { print_compile_info(NULL); } 
         return 0;
     }
+    const bool asked_help = input.cmdOptionExists("--help");
+    if (asked_help) {
+        fprintf( stdout, "The command-line input arguments [and default values] are:\n" );
+    }
 
     // first argument is the flag, second argument is default value (for when flag is not present)
-    const std::string   &input_fname      = input.getCmdOption("--input_file",      "input.nc"),
-                        &output_fname     = input.getCmdOption("--output_file",     "projection_Helmholtz.nc"),
-                        &seed_fname       = input.getCmdOption("--seed_file",       "seed.nc");
+    const std::string   &input_fname      = input.getCmdOption("--input_file",      "input.nc",                 asked_help),
+                        &output_fname     = input.getCmdOption("--output_file",     "projection_Helmholtz.nc",  asked_help),
+                        &seed_fname       = input.getCmdOption("--seed_file",       "seed.nc",                  asked_help);
 
-    const std::string   &time_dim_name      = input.getCmdOption("--time",        "time"),
-                        &depth_dim_name     = input.getCmdOption("--depth",       "depth"),
-                        &latitude_dim_name  = input.getCmdOption("--latitude",    "latitude"),
-                        &longitude_dim_name = input.getCmdOption("--longitude",   "longitude");
+    const std::string   &time_dim_name      = input.getCmdOption("--time",        "time",       asked_help),
+                        &depth_dim_name     = input.getCmdOption("--depth",       "depth",      asked_help),
+                        &latitude_dim_name  = input.getCmdOption("--latitude",    "latitude",   asked_help),
+                        &longitude_dim_name = input.getCmdOption("--longitude",   "longitude",  asked_help);
 
-    const std::string &latlon_in_degrees  = input.getCmdOption("--is_degrees",   "true");
+    const std::string &latlon_in_degrees  = input.getCmdOption("--is_degrees",   "true", asked_help);
 
-    const std::string   &Nprocs_in_time_string  = input.getCmdOption("--Nprocs_in_time",  "1"),
-                        &Nprocs_in_depth_string = input.getCmdOption("--Nprocs_in_depth", "1");
+    const std::string   &Nprocs_in_time_string  = input.getCmdOption("--Nprocs_in_time",  "1", asked_help),
+                        &Nprocs_in_depth_string = input.getCmdOption("--Nprocs_in_depth", "1", asked_help);
     const int   Nprocs_in_time_input  = stoi(Nprocs_in_time_string),
                 Nprocs_in_depth_input = stoi(Nprocs_in_depth_string);
 
-    const std::string   &zonal_vel_name    = input.getCmdOption("--zonal_vel",   "uo"),
-                        &merid_vel_name    = input.getCmdOption("--merid_vel",   "vo"),
-                        &tor_seed_name     = input.getCmdOption("--tor_seed",    "Psi_seed"),
-                        &pot_seed_name     = input.getCmdOption("--pot_seed",    "Phi_seed");
+    const std::string   &zonal_vel_name    = input.getCmdOption("--zonal_vel",   "uo",          asked_help),
+                        &merid_vel_name    = input.getCmdOption("--merid_vel",   "vo",          asked_help),
+                        &tor_seed_name     = input.getCmdOption("--tor_seed",    "Psi_seed",    asked_help),
[…]
