### 2022-09-21 d951de5
Adding in a constants.hpp flag for handling wind inputs
