### 2022-09-21 a071c2a
Adding flags to ignore time-means and OkuboWeiss histograms when merging postprocess results
