### 2022-09-22 7c277b0
Narrowing pole masking to 0.5 degrees
