### 2022-09-22 7aac37a
Adding OutputHandling tool to merge timeseries chunks.
