+++ PythonTools/OutputHandling/merge_timeseries.py
@@ -1,0 +1,223 @@
+import numpy as np
+import argparse
+import glob
+from netCDF4 import Dataset
+
+parser = argparse.ArgumentParser(description='Merge timeseries.')
+
+parser.add_argument('--file_pattern', metavar='search_pattern', type=str, nargs=1, required = True,
+        help='Filename pattern for merging. e.g. "postprocess_*.nc"')
+
+parser.add_argument('--output_filename', metavar='output', type=str, nargs=1, required = True,
+        help='Filename pattern for merged set. e.g. "postprocess.nc"')
+
+parser.add_argument('--print_level', metavar='debug', type=int, nargs=1, default = 0,
+        help='String to indicate how much printing to do. Options are 0, 1, 2 [higher value means more printed].')
+
+# Pass --exclude_time to exclude time means
+parser.add_argument('--exclude_time', dest='copy_time_means', action='store_false',
+        help="Pass '--exclude_time' to have time means ignored when merging postprocessing files.")
+parser.set_defaults(copy_time_means=True)
+
+# Pass --exclude_OkuboWeiss to exclude OkuboWeiss histograms
+parser.add_argument('--exclude_OkuboWeiss', dest='copy_OkuboWeiss', action='store_false',
+        help="Pass '--exclude_OkuboWeiss' to have OkuboWeiss histograms ignored when merging postprocessing files.")
+parser.set_defaults(copy_OkuboWeiss=True)
+
+# Pass --exclude_zonal_mean to exclude zonal means
+parser.add_argument('--exclude_zonal_means', dest='copy_zonal_means', action='store_false',
+        help="Pass '--exclude_zonal_means' to have zonal means ignored when merging postprocessing files.")
+parser.set_defaults(copy_zonal_means=True)
+
+# Now actually parse in command-line flags
+args = parser.parse_args()
+
+if type(args.print_level) == type(0):
+    print_level = args.print_level
+elif type(args.print_level) == type([0,]):
+    print_level = args.print_level[0]
+
+print("Attempting to merge all files matching pattern {0} into output file {1}.".format( args.file_pattern[0], args.output_filename[0] ))
+
+DO_NOT_COPY_TIME_MEANS = not( args.copy_time_means )
+if DO_NOT_COPY_TIME_MEANS:
+    print("  Will not merge time means maps.")
+
+DO_NOT_COPY_OKUBOWEISS = not( args.copy_OkuboWeiss )
+if DO_NOT_COPY_OKUBOWEISS:
+    print("  Will not merge OkuboWeiss histograms.")
+
+DO_NOT_COPY_ZONAL_MEANS = not( args.copy_zonal_means )
+if DO_NOT_COPY_ZONAL_MEANS:
+    print("  Will not merge zonal means.")
+
+
+
+## First, find all files matching the requested pattern
+all_fps = glob.glob( args.file_pattern[0] )
+Nfps = len(all_fps)
+print("  Identified {0:g} files for merging.".format(Nfps), flush = True)
+
+
+## Get a list of the dimensions and variables
+with Dataset( all_fps[0], 'r' ) as dset:
+    Nvars = len( dset.variables.copy() )
+    Ndims = len( dset.dimensions.copy() )
+
+print("  Identified {0:g} dimensions and {1:g} variables to copy into merged file.".format( Ndims, Nvars ), flush = True)
+
+
+## Get the time from each file, and sort them in ascending order
+#   for the sorting, we will assume that the it's enough to sort
+#   the files, and not that we need to sort within or between files.
+##
+t0s = np.zeros( Nfps )
[…]
