### 2022-09-22 0b810a1
Adding flag to also ignore zonal means when merging
