### 2022-10-05 9d8e73e
Including bash for loop syntax in merger demo
