+++ PythonTools/OutputHandling/merge_postprocess_files.sh
@@ -3,6 +3,14 @@
 #   a directory called 'outputs' in the current working directory.
 # Structure assume Helmholtz-type outputs.
 
-python merge_postprocess_results.py --file_pattern "outputs/postprocess_full_*nc" --output_filename "postprocess_full.nc"
-python merge_postprocess_results.py --file_pattern "outputs/postprocess_toroidal_*nc" --output_filename "postprocess_toroidal.nc"
-python merge_postprocess_results.py --file_pattern "outputs/postprocess_potential_*nc" --output_filename "postprocess_potential.nc"
+KINDS=("full" "toroidal" "potential")
+
+for KIND in "${KINDS[@]}"
+do
+    python merge_postprocess_results.py \
+        --file_pattern "outputs/postprocess_${KIND}_*nc" \
+        --output_filename "postprocess_${KIND}.nc" \
+        --exclude_time_means \
+        --exclude_OkuboWeiss \
+        --exclude_zonal_means
+done
