### 2022-10-10 #4277 @kthyng
Hello! I have just looked through the page, looks like we are a bit stuck. What can we do to keep moving forward?

@kris-rowe Looks like you need to pick up your review after the license situation has been worked out.

@NoraLoose I see you've opened some issues — they all look like the author has responded, so please check that out and see if the issues still need to be open. And otherwise can you carry on with your review?

@bastorer What is the status on your end? Are you waiting for reviews or are you still working on anything?

Thanks all.
