### 2022-10-11 #4277 @bastorer
Hello all,

Sorry about the slow responses. Commits have been submitted responding to
the issues posted on the github page, except for the paper changes, which I
will have by early next week.

Regards
Ben
