### 2022-10-18 7981254
Adding functionality to include radial velocity
