+++ Functions/Helmholtz/filtering_helmholtz.cpp
@@ -45,11 +45,15 @@ void filtering_helmholtz(
                                 &dAreas     = source_data.areas;
 
     const std::vector<double>   &F_potential    = source_data.variables.at("F_potential"),
-                                &F_toroidal     = source_data.variables.at("F_toroidal");
+                                &F_toroidal     = source_data.variables.at("F_toroidal"),
+                                &u_r            = source_data.variables.at("u_r");
 
     const std::vector<double>   &uiuj_F_r   = ( constants::COMP_PI_HELMHOLTZ ) ? source_data.variables.at("uiuj_F_r")   : zero_vector,
                                 &uiuj_F_Phi = ( constants::COMP_PI_HELMHOLTZ ) ? source_data.variables.at("uiuj_F_Phi") : zero_vector,
                                 &uiuj_F_Psi = ( constants::COMP_PI_HELMHOLTZ ) ? source_data.variables.at("uiuj_F_Psi") : zero_vector;
+
+    const std::vector<double>   &wind_tau_Psi = ( constants::COMP_WIND_FORCE ) ? source_data.variables.at("wind_tau_Psi") : zero_vector,
+                                &wind_tau_Phi = ( constants::COMP_WIND_FORCE ) ? source_data.variables.at("wind_tau_Phi") : zero_vector;
 
     const std::vector<bool> &mask = source_data.mask;
 
@@ -96,6 +100,7 @@ void filtering_helmholtz(
         // Arrays to store filtered Phi and Psi fields (potential, pseudo-potential)
         coarse_F_tor(   num_pts, 0. ),
         coarse_F_pot(   num_pts, 0. ),
+        u_r_coarse(     num_pts, 0. ),
 
         coarse_uiuj_F_r(   num_pts, 0. ),
         coarse_uiuj_F_Phi( num_pts, 0. ),
@@ -201,9 +206,11 @@ void filtering_helmholtz(
         //
         //// Spherical velocity components
         //
+        zero_array( num_pts, 0.),
 
-        // Spherical - radial velocities (just set to zero)
-        u_r_zero( num_pts, 0.),
+        // Spherical - radial velocity
+        //  by definition this is potential-only since toroidal is incompressible
+        //u_r( num_pts, 0.),
 
         // Spherical - zonal velocities
         u_lon_tor( num_pts, 0. ),
@@ -251,6 +258,23 @@ void filtering_helmholtz(
         Z_pot( num_pts, 0. ),
         Z_tot( num_pts, 0. );
 
+    std::vector<double> tau_wind_x_tor( constants::COMP_WIND_FORCE ? num_pts : 1, 0. ),
+                        tau_wind_y_tor( constants::COMP_WIND_FORCE ? num_pts : 1, 0. ),
+                        tau_wind_x_pot( constants::COMP_WIND_FORCE ? num_pts : 1, 0. ),
+                        tau_wind_y_pot( constants::COMP_WIND_FORCE ? num_pts : 1, 0. ),
+                        tau_wind_dot_u_tor( constants::COMP_WIND_FORCE ? num_pts : 1, 0. ),
+                        tau_wind_dot_u_pot( constants::COMP_WIND_FORCE ? num_pts : 1, 0. ),
+                        local_wind_forcing_tor( constants::COMP_WIND_FORCE ? num_pts : 1, 0. ),
+                        local_wind_forcing_pot( constants::COMP_WIND_FORCE ? num_pts : 1, 0. ),
[…]
