+++ Functions/Differentiation_Tools/spher_derivative_at_point.cpp
@@ -37,22 +37,27 @@ void spher_derivative_at_point(
     // Check which derivative we're taking
     int index, Iref;
     const int Nref = grid.size();
+    const bool do_dep = (dim == "depth");
     const bool do_lat = (dim == "lat");
     const bool do_lon = (dim == "lon");
-    assert( do_lat ^ do_lon ); // xor
+    assert( do_dep ^ (do_lat ^ do_lon) ); // ^ = xor
 
-    // Verify input
-    if      (do_lon) { Iref = Ilon; }
-    else if (do_lat) { Iref = Ilat; }
-    else { 
-        fprintf(stderr, "Illegal dimension provided! %s given to %s\n", 
-                dim.c_str(), __FILE__);
-        assert(false);
+    int Iref = do_dep ? Idepth :
+               do_lat ? Ilat :
+               do_lon ? Ilon : -1;
+    const int Nref = grid.size();
+
+    // If it's a singleton dimension, just return zeros (we zeroed out earlier)
+    //      this if for the case of no actual depth values
+    if (Nref == 1) {
+        return;
     }
 
     // Determine lowest lower bound (LLB) and upperest upper bound (UUB)
     //   for the integration region. This essentially just depends on periodicity.
-    const bool periodic = do_lat ? constants::PERIODIC_Y : constants::PERIODIC_X ;
+    const bool periodic = do_dep ? false : 
+                          do_lat ? constants::PERIODIC_Y : 
+                          do_lon ? constants::PERIODIC_X : false;
     const int LLB = periodic ? Iref - Nref : 0 ;
     const int UUB = periodic ? Iref + Nref : Nref - 1 ;
 
@@ -74,8 +79,8 @@ void spher_derivative_at_point(
        
         // Check if the next point would be land
         lb = ( ( LB - 1 ) % Nref + Nref ) % Nref;
-        if (do_lon) { index = Index(Itime, Idepth, Ilat, lb,   Ntime, Ndepth, Nlat, Nlon); }
-        else        { index = Index(Itime, Idepth, lb,   Ilon, Ntime, Ndepth, Nlat, Nlon); }
+        index = Index( Itime, do_dep ? lb : Idepth, do_lat ? lb : Ilat, do_lon ? lb : Ilon,
+                       Ntime, Ndepth,               Nlat,               Nlon );
         
         if ( mask.at(index) )   { LB--;  }  // If next point is still water, extend stencil over it
         else                    { break; }  // Otherwise, halt [ without extending stencil ]
@@ -90,8 +95,8 @@ void spher_derivative_at_point(
        
         // Check if the next point would be land
         ub = ( ( UB + 1 ) % Nref + Nref ) % Nref;
-        if (do_lon) { index = Index(Itime, Idepth, Ilat, ub,   Ntime, Ndepth, Nlat, Nlon); }
-        else        { index = Index(Itime, Idepth, ub,   Ilon, Ntime, Ndepth, Nlat, Nlon); }
+        index = Index( Itime, do_dep ? ub : Idepth, do_lat ? ub : Ilat, do_lon ? ub : Ilon,
+                       Ntime, Ndepth,               Nlat,               Nlon );
 
         if ( mask.at(index) )   { UB++;  }  // If next point is still water, extend stencil over it
         else                    { break; }  // Otherwise, halt [ without extending stencil ]
@@ -116,7 +121,7 @@ void spher_derivative_at_point(
     int ind; 
     if (UB - LB + 1 == num_deriv_pts) {
         // If we have enough cells for differentiation, do it
-        if ( do_lon or (constants::UNIFORM_LAT_GRID)) {
[…]

[…]
