### 2022-10-18 a3a5873
Introducing MPI subcommunicators One is for all ranks with same time indices Another is for all ranks with the same depth indices
These should reduce the cost of various gather and reduce calls.

These have been put into the time averaging routines.
