### 2022-10-25 husseinaluie/FlowSieve#13 @NoraLoose
Automated tests
I see that the package contains tests (in `Tests/`). I have a few questions:
* Are the tests automated?
* If the tests rely on manual testing, are there instructions for how the user can run these tests?

This issue is part of the JOSS review process over at https://github.com/openjournals/joss-reviews/issues/4277.
