### 2022-10-25 husseinaluie/FlowSieve#14 @NoraLoose
Running the basic tutorial
- [x] Starting from the "default" `constants.hpp`, I modified variables as in the documentation, but in addition needed to make the following modification:

```
EXTEND_DOMAIN_TO_POLES = false
```
I think it would be useful to include this in the docs too.

- [ ] It would be nice if the tutorial included the required python dependencies, e.g., for example by providing an `environment.yml` file to set up a conda environment.
- [ ] How about including a jupyter notebook that executes the steps done with python? Could be a great way to make the tutorial more interactive, and give the user some nice figures to look at and reproduce! It would only take a few lines of code to get started, e.g.,
```
import xarray as xr
ds = xr.open_dataset('/glade/u/home/noraloose/FlowSieve/Tutorial/Basic/velocity_sample.nc')
ds.uo.plot()
```
but you could also consider to make the 2 python scripts more interactive and step through them in the notebook.

This issue is part of the JOSS review process over at https://github.com/openjournals/joss-reviews/issues/4277.
