### 2022-10-25 husseinaluie/FlowSieve#15 @NoraLoose
Statement of need in documentation
JOSS asks for a statement of need in the documentation. This statement should state what problems the software is designed to solve and who the target audience is.

https://github.com/openjournals/joss-reviews/issues/4277
