### 2022-10-25 husseinaluie/FlowSieve#6 @NoraLoose
I agree - this looks much more user-friendly now! :)

While reading through the documentation, I spotted a few further minor typos (not all in the Methods sections):

- [ ] consants --> constants on https://flowsieve.readthedocs.io/en/latest/constants_8hpp.html
- [ ] reach --> reached on https://flowsieve.readthedocs.io/en/latest/helmholtz1.html
- [ ] "Coarsen the to" --> "Coarsen to" on https://flowsieve.readthedocs.io/en/latest/helmholtz1.html
