### 2022-10-25 husseinaluie/FlowSieve#16 @NoraLoose
What variables can be coarse grained by FlowSieve?
Is FlowSieve specialized in coarse-graining velocity fields? Or can you also coarse-grain tracer fields?

I think these are questions that many potential users will have. It would be great if 1-2 clarifying sentences can be included in the paper & documentation.

https://github.com/openjournals/joss-reviews/issues/4277
