### 2022-10-25 husseinaluie/FlowSieve#14 @bastorer
Making notebook versions of the tutorial python scripts is a good idea!

I'm not very familiar with conda, but agree that giving explicit library dependencies would be helpful. I'll look into those this week.
