### 2022-10-25 husseinaluie/FlowSieve#17 @NoraLoose
Running the tutorial on the sphere
- [ ] Is there a chance that you could include a second submission script for folks that have to use PBS (rather than slurm) schedulers? I think this could lower the startup hurdle for new users.
- [x] Can this tutorial be simplified? I imagine that most users do not want to wait for 3 hours until the scripts have finished.
- [ ] Independent of whether the previous 2 points can be done, including a jupyter notebook for analysis would be a great. The notebook with included figures would help users to get an idea what the scripts and the package does, even before they run the full tutorial.

https://github.com/openjournals/joss-reviews/issues/4277
