### 2022-10-25 husseinaluie/FlowSieve#16 @bastorer
FlowSieve can coarse-grain any scalar fields, and can nominally process any vector field as well, provided that it first be passed through the Helmholtz projection step.

A lot / most of the diagnostic outputs are built around filtering velocity, but we've also filtered density fields using the `coarse_grain_scalar.x` case file.

Good suggestion about the clarifying sentences! I'll add those.
