### 2022-10-25 husseinaluie/FlowSieve#13 @bastorer
The tests aren't automated. They were originally built during the earlier development phase where we were testing several components. Unfortunately they've been rather neglected lately, and many likely don't work at the moment [although modifying them to use the new function shouldn't be too difficult. I'll add it to the to-do list]. I don't know how to automate the tests, but is something that I'll look into.

The Tutorials are the closest thing that we have for users to test that their FlowSieve installation is up and running.
