### 2022-10-25 husseinaluie/FlowSieve#16 @NoraLoose
Gotcha! Could be nice to illustrate the filtering of a scalar field as part of the basic tutorial as well. I know tons of people who want to filter scalar fields - so having an example for scalar fields in the tutorial could attract more users! 😃
