### 2022-10-25 husseinaluie/FlowSieve#17 @bastorer
1. I'd be happy to include PBS submit scripts, but I'm afraid I'm not very familiar with it. Could you point me to a documentation page or sample script?
2. That's a very fair point, three hours is a bit long to wait just to find out if something works. I propose keeping that one as it is, with the different scales of eddies, and add in a second spherical tutorial that runs on very low resolution (e.g. 2-degree grid) in order to give results very quickly.
3. Agreed, notebooks would be nice. I originally just used .py scripts so that everything could be wrapped up in the submission script, but having a notebook that shows users how to interact with the outputs would be very beneficial.
