### 2022-10-25 husseinaluie/FlowSieve#13 @NoraLoose
I think it will be worth resurrecting the testing framework! And it's not just me who is a big fan of tests - JOSS also requires us reviewers to check the following box:

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
