### 2022-10-26 husseinaluie/FlowSieve#17 @NoraLoose
Okay, maybe 1. (the PBS submission script) is not really necessary. If your HPC cluster does not support PBS, it would be hard for you test it out. The user can hopefully figure out how to translate from slurm to PBS if necessary.

2. and 3. sound good!
