### 2022-10-26 8fbd745
Only print "Hello world" from each thread if DEBUG >= 3
