+++ Functions/Interface_Tools/print_header_info.cpp
@@ -33,7 +33,7 @@ void print_header_info( ) {
     MPI_Barrier(MPI_COMM_WORLD);
 
     // Print processor assignments
-    #if DEBUG >= 2
+    #if DEBUG >= 3
     int tid, nthreads;
     #pragma omp parallel default(none) private(tid, nthreads) \
         shared(stdout) firstprivate(wRank, wSize)
