### 2022-10-26 ef0568d
Increasing OMP thread chunk-size
