### 2022-10-26 e258e51
Fixing a time-ordering problem in merge_timeseries
