### 2022-10-26 e2f0ac8
Depth-integrator handles increasing/decreasing depth grids, and depth/elevation
