### 2022-10-26 80bb1d9
Updating compute_radial_vel casefile to new function formats
