### 2022-10-26 7366b87
Time-means now only communicate across MPI subcommunicators
