### 2022-10-26 f500982
Fixing typos in issue #6
