### 2022-10-26 e53ba60
Adding reference to '--help' in docs
