### 2022-10-26 8058aed
Adding smaller tutorial case for sphere. Runs in ~5 mins. Partial response to issue #17
