### 2022-10-26 8d9caee
Adding summary of feature and statement of need to ReadTheDocs Following issues #15 and #16 This text is copied verbatim from the paper.md file
