### 2022-10-26 f269c00
Revamping the Scalars tutorial (see Issue #16)
