### 2022-10-28 husseinaluie/FlowSieve#13 @bastorer
**Re: tutorials**

I'm just about done updating the tutorials. There's now four in total. In addition to SLURM submit scripts, I also include regular bash scripts so that the user can just run the tutorial directly. Except for the largest one, they all run in under 5 minutes on one processor.

1. "Basic" just a Cartesian doubly-periodic box. Runs in ~1 minute on one processor.
2. "Scalar" working on a sphere, filters an artificial 'density' field at a couple filter scales. Runs in ~1 min on one processor
3. "Low Resolution Helmholtz" is essentially a lower resolution version of the old Helmholtz tutorial. Runs in ~5 minutes on one processor
4. "Helmholtz" is the previous Helmholtz tutorial. On one processor, the three parts take a couple hours to run. With a couple processors in OpenMP ( there's a line in the bash and SLURM scripts to set the number of processors ), it would be 30min - 1 hour. It doesn't show particularly new functionality, so the other tutorials illustrate usage, but gives a clearer sense of what a "production level" run might look like.

I've added Jupyter notebooks for analyzing the results for most of them, just finishing up the last one now (will finish some time tomorrow).

**Re: testing**

I have some ideas for resurrecting the testing framework, but building up a proper framework is going to take a decent bit of time. With the tutorial Jupyter notebooks including the figures from when I generated them, would those qualify as "manual steps described so that the functionality of the software can be verified?"

Have a nice weekend!
