### 2022-11-03 644bb6a
Improving "--help" command by including argument descriptions
Now with fancy formatting!
