+++ Case_Files/coarse_grain_helmholtz.cpp
@@ -64,45 +64,100 @@ int main(int argc, char *argv[]) {
     }
     const bool asked_help = input.cmdOptionExists("--help");
     if (asked_help) {
-        fprintf( stdout, "The command-line input arguments [and default values] are:\n" );
+        fprintf( stdout, "\033[1;4mThe command-line input arguments [and default values] are:\033[0m\n" );
     }
 
     // first argument is the flag, second argument is default value (for when flag is not present)
-    const std::string   &Helm_input_fname  = input.getCmdOption("--Helmholtz_input_file",       "Helmholtz_projection.nc",      asked_help),
-                        &vel_input_fname   = input.getCmdOption("--velocity_input_file",        "vels.nc",                      asked_help),
-                        &u_r_input_fname   = input.getCmdOption("--radial_velocity_input_file", "NONE",                         asked_help),
-                        &wind_input_fname  = input.getCmdOption("--wind_tau_input_file",        "wind_tau_projection.nc",       asked_help),
-                        &quad_input_fname  = input.getCmdOption("--uiuj_Helmholtz_input_file",  "helmholtz_projection_uiuj.nc", asked_help);
+    const std::string   &Helm_input_fname  = input.getCmdOption("--Helmholtz_input_file",       
+                                                                "Helmholtz_projection.nc",      
+                                                                asked_help,
+                                                                "netCDF file containing streamfunction and potential function."),
+                        &vel_input_fname   = input.getCmdOption("--velocity_input_file",        
+                                                                "vels.nc",                      
+                                                                asked_help,
+                                                                "netCDF file containing velocity field (only used to get land mask)"),
+                        &u_r_input_fname   = input.getCmdOption("--radial_velocity_input_file", 
+                                                                "NONE",                         
+                                                                asked_help,
+                                                                "netCDF file containing radial/vertical velocity, if applicable. \nNONE if no radial velocity."),
+                        &wind_input_fname  = constants::COMP_WIND_FORCE ?
+                                                input.getCmdOption("--wind_tau_input_file",        
+                                                                   "wind_tau_projection.nc",       
+                                                                   asked_help,
+                                                                   "netCDF file containing surface wind stress Helmholtz Psi and Phi")
+                                                : "",
+                        &quad_input_fname  = constants::COMP_PI_HELMHOLTZ ? 
+                                                input.getCmdOption("--uiuj_Helmholtz_input_file",  
+                                                                   "helmholtz_projection_uiuj.nc", 
+                                                                   asked_help,
+                                                                   "netCDF file containing Helmholtz components of u*u tensor")
+                                                : "";
 
-    const std::string   &time_dim_name      = input.getCmdOption("--time",        "time",       asked_help),
-                        &depth_dim_name     = input.getCmdOption("--depth",       "depth",      asked_help),
-                        &latitude_dim_name  = input.getCmdOption("--latitude",    "latitude",   asked_help),
-                        &longitude_dim_name = input.getCmdOption("--longitude",   "longitude",  asked_help);
+    const std::string   &time_dim_name      = input.getCmdOption("--time",        
+                                                                 "time",       
+                                                                 asked_help,
+                                                                 "Name of 'time' dimension in netCDF input file."),
+                        &depth_dim_name     = input.getCmdOption("--depth",       
+                                                                 "depth",      
[…]
