+++ functions.hpp
@@ -538,7 +538,8 @@ class InputParser {
         const std::string getCmdOption(
                 const std::string &option,
                 const std::string &default_value = "",
-                const bool help = false
+                const bool help = false,
+                const std::string &description = ""
                 ) const;
 
         /*!
@@ -563,7 +564,8 @@ class InputParser {
         void getListofStrings( 
                 std::vector<std::string> &list_of_strings, 
                 const std::string &argname,
-                const bool help = false
+                const bool help = false,
+                const std::string &description = ""
                 ) const;
 
     private:
