### 2022-11-03 6282267
Adding more "--help" descriptors
