+++ Case_Files/coarsen_grid_linear.cpp
@@ -44,29 +44,61 @@ int main(int argc, char *argv[]) {
     }
     const bool asked_help = input.cmdOptionExists("--help");
     if (asked_help) {
-        fprintf( stdout, "The command-line input arguments [and default values] are:\n" );
+        fprintf( stdout, "\033[1;4mThe command-line input arguments [and default values] are:\033[0m\n" );
     }
 
     // first argument is the flag, second argument is default value (for when flag is not present)
-    const std::string   &coarse_fname   = input.getCmdOption("--coarse_file",   "coarse.nc",        asked_help),
-                        &fine_fname     = input.getCmdOption("--fine_file",     "fine.nc",          asked_help),
-                        &output_fname   = input.getCmdOption("--output_file",   "coarse_vel.nc",    asked_help);
+    const std::string   &coarse_fname   = input.getCmdOption("--coarse_file",   
+                                                             "coarse.nc",        
+                                                             asked_help,
+                                                             "netCDF file containing the grid onto which you want to downsample / coarsen."),
+                        &fine_fname     = input.getCmdOption("--fine_file",     
+                                                             "fine.nc",          
+                                                             asked_help,
+                                                             "netCDf file containing the variables that you want to downsample / coarsen."),
+                        &output_fname   = input.getCmdOption("--output_file",   
+                                                             "coarse_vel.nc",    
+                                                             asked_help,
+                                                             "Filename for where the downsampled variables will be stored.");
 
-    const std::string   &time_dim_name      = input.getCmdOption("--time",        "time",       asked_help),
-                        &depth_dim_name     = input.getCmdOption("--depth",       "depth",      asked_help),
-                        &latitude_dim_name  = input.getCmdOption("--latitude",    "latitude",   asked_help),
-                        &longitude_dim_name = input.getCmdOption("--longitude",   "longitude",  asked_help);
+    const std::string   &time_dim_name      = input.getCmdOption("--time",        
+                                                                 "time",       
+                                                                 asked_help,
+                                                                 "Name of 'time' dimension in netCDF input file."),
+                        &depth_dim_name     = input.getCmdOption("--depth",       
+                                                                 "depth",      
+                                                                 asked_help,
+                                                                 "Name of 'depth' dimension in netCDF input file."),
+                        &latitude_dim_name  = input.getCmdOption("--latitude",    
+                                                                 "latitude",   
+                                                                 asked_help,
+                                                                 "Name of 'latitude' dimension in netCDF input file."),
+                        &longitude_dim_name = input.getCmdOption("--longitude",   
+                                                                 "longitude",  
+                                                                 asked_help,
+                                                                 "Name of 'longitude' dimension in netCDF input file.");
 
-    const std::string &latlon_in_degrees  = input.getCmdOption("--is_degrees",   "true", asked_help);
+    const std::string &latlon_in_degrees  = input.getCmdOption("--is_degrees",   
+                                                               "true", 
+                                                               asked_help,
+                                                               "Boolean (true/false) indicating if the grid is in degrees (true) or radians (false).");
 
-    const std::string   &Nprocs_in_time_string  = input.getCmdOption("--Nprocs_in_time",  "1", asked_help),
-                        &Nprocs_in_depth_string = input.getCmdOption("--Nprocs_in_depth", "1", asked_help);
+    const std::string   &Nprocs_in_time_string  = input.getCmdOption("--Nprocs_in_time",  
+                                                                     "1", 
+                                                                     asked_help,
[…]
