+++ Case_Files/Helmholtz_projection.cpp
@@ -49,44 +49,98 @@ int main(int argc, char *argv[]) {
     }
     const bool asked_help = input.cmdOptionExists("--help");
     if (asked_help) {
-        fprintf( stdout, "The command-line input arguments [and default values] are:\n" );
+        fprintf( stdout, "\033[1;4mThe command-line input arguments [and default values] are:\033[0m\n" );
     }
 
     // first argument is the flag, second argument is default value (for when flag is not present)
-    const std::string   &input_fname      = input.getCmdOption("--input_file",      "input.nc",                 asked_help),
-                        &output_fname     = input.getCmdOption("--output_file",     "projection_Helmholtz.nc",  asked_help),
-                        &seed_fname       = input.getCmdOption("--seed_file",       "seed.nc",                  asked_help);
+    const std::string   &input_fname      = input.getCmdOption("--input_file",      
+                                                               "input.nc",                 
+                                                               asked_help,
+                                                               "netCDF file containing vector field to Helmholtz decompose"),
+                        &output_fname     = input.getCmdOption("--output_file",     
+                                                               "projection_Helmholtz.nc",  
+                                                               asked_help,
+                                                               "Filename for the output (netCDF file of Helmholtz scalars)"),
+                        &seed_fname       = input.getCmdOption("--seed_file",       
+                                                               "zero",                  
+                                                               asked_help,
+                                                               "netCDF file containing initial guesses for Helmholtz scalars.\nUse 'zero' (the default) for no seed.");
 
-    const std::string   &time_dim_name      = input.getCmdOption("--time",        "time",       asked_help),
-                        &depth_dim_name     = input.getCmdOption("--depth",       "depth",      asked_help),
-                        &latitude_dim_name  = input.getCmdOption("--latitude",    "latitude",   asked_help),
-                        &longitude_dim_name = input.getCmdOption("--longitude",   "longitude",  asked_help);
+    const std::string   &time_dim_name      = input.getCmdOption("--time",        
+                                                                 "time",       
+                                                                 asked_help,
+                                                                 "Name of 'time' dimension in netCDF input file."),
+                        &depth_dim_name     = input.getCmdOption("--depth",       
+                                                                 "depth",      
+                                                                 asked_help,
+                                                                 "Name of 'depth' dimension in netCDF input file."),
+                        &latitude_dim_name  = input.getCmdOption("--latitude",    
+                                                                 "latitude",   
+                                                                 asked_help,
+                                                                 "Name of 'latitude' dimension in netCDF input file."),
+                        &longitude_dim_name = input.getCmdOption("--longitude",   
+                                                                 "longitude",  
+                                                                 asked_help,
+                                                                 "Name of 'longitude' dimension in netCDF input file.");
 
-    const std::string &latlon_in_degrees  = input.getCmdOption("--is_degrees",   "true", asked_help);
+    const std::string &latlon_in_degrees  = input.getCmdOption("--is_degrees",   
+                                                               "true", 
+                                                               asked_help,
+                                                               "Boolean (true/false) indicating if the grid is in degrees (true) or radians (false).");
 
-    const std::string   &Nprocs_in_time_string  = input.getCmdOption("--Nprocs_in_time",  "1", asked_help),
-                        &Nprocs_in_depth_string = input.getCmdOption("--Nprocs_in_depth", "1", asked_help);
+    const std::string   &Nprocs_in_time_string  = input.getCmdOption("--Nprocs_in_time",  
+                                                                     "1", 
+                                                                     asked_help,
[…]
