### 2022-11-03 ca9946b
Last round of updates for high-resolution spherical tutorial
See ISSUES #17 and #14
