+++ Tutorial/Spherical_Demo/Helmholtz/postprocessing/generate_data_sphere.py
@@ -1,122 +1,0 @@
-import numpy as np
-from netCDF4 import Dataset
-import FiniteDiff
-
-##
-#
-# The generation script builds a full-sphere data set
-#   at 0.5 degree resolution. It consists of a large
-#   collection of eddies of differing sizes, randomly
-#   distributed around the globe.
-#
-##
-
-# Make grid
-Nlon, Nlat = int(360*2), int(180*2)
-
-Llon, Llat = 2 * np.pi, np.pi
-
-dlon = Llon / Nlon
-dlat = Llat / Nlat
-
-lon = np.arange( dlon/2, Llon, dlon )       - Llon/2
-lat = np.arange( dlat/2, Llat, dlat )[1:-1] - Llat/2  # remove the actual poles
-Nlat -= 2
-
-LON, LAT = np.meshgrid(lon, lat)
-
-# Define some eddy parameters
-eddy_scales = [ 250e3, 750e3, 3500e3 ]
-eddy_counts = [ 1e3,   3e2,   5e1 ]
-eddy_velocs = [ 0.2,   0.4,   0.1 ]
-
-
-# Function to compute distances on sphere
-def dist( lon0, lat0, LON = LON, LAT = LAT, R = 6371e3 ):
-
-    del_LON = np.abs( LON - lon0 )
-    del_LAT = np.abs( LAT - lat0 )
-
-    numer  = ( np.cos( LAT ) * np.sin( del_LON ) )**2
-    numer += ( np.cos( lat0 ) * np.sin( LAT ) - np.sin( lat0 ) * np.cos( LAT ) * np.cos( del_LON) )**2
-    numer  = np.sqrt(numer)
-
-    denom = np.sin( lat0 ) * np.sin( LAT ) + np.cos( lat0 ) * np.cos( LAT ) * np.cos( del_LON )
-
-    #return R * np.arctan( numer / denom )
-    return R * np.arctan2( numer, denom )
-
-
-# Build streamfunction
-Psi   = np.zeros( (Nlat,Nlon) )
-u_lon = np.zeros( (Nlat,Nlon) )
-u_lat = np.zeros( (Nlat,Nlon) )
-
-for scale, count, veloc in zip( eddy_scales, eddy_counts, eddy_velocs ):
-
-    for ii in range(int(count)):
-        W = ( scale * (1 + 0.05 * np.random.randn() ) ) / 2
-
-        lon0 = np.random.choice( lon )  # choose longitude at random
-        lat0 = np.random.choice( lat, p = np.cos(lat) / np.sum( np.cos(lat) ) )   # choose latitude, weighting based on area
-
-        D = dist( lon0, lat0 )
-
-        sign = np.random.choice( (-1,1) )
-
-        Psi += sign * veloc * W * np.exp( -(D/W)**2 )
-
-# Get velocity from streamfunction
-ddlon = FiniteDiff.FiniteDiff( lon, 8, Uniform = True, Periodic = True,  Sparse = True )
-ddlat = FiniteDiff.FiniteDiff( lat, 8, Uniform = True, Periodic = False, Sparse = True )
-
-u_lon = - ddlat.dot( Psi   )   /   6371e3
-u_lat =   ddlon.dot( Psi.T ).T / ( 6371e3 * np.cos(LAT) )
-
-# Save flow to a file
-dtype_dim = np.float64
-dtype = np.float32
-
-dims = ('time','depth','latitude','longitude')
-
-fill_value = -1e10
-
-with Dataset('velocity_sample.nc', 'w', format='NETCDF4') as fp:
-
-    # time
-    dim = 'time'
-    t_dim = fp.createDimension(dim, 1)
[…]
