### 2022-11-03 c4d60dd
Adding pole extension to (mothballed) uiuj projector
