### 2022-11-03 8eb5cc5
Minor tweaks to basic tutorial
