### 2022-11-03 husseinaluie/FlowSieve#17 @bastorer
The smaller tutorial on the sphere has been added that runs in just a couple minutes on one processor, with a bash script to run all of the steps directly (without submitting to a scheduler).

The Scalar and small helmholtz, and large helmholtz have had jupyter notebooks added to them to show using the output files.

I also integrated in some of the newer user-friendliness tools that make reading the outputs easier when using post-processing.
