### 2022-11-03 husseinaluie/FlowSieve#6 @bastorer
Also, I've added a `--help` flag to the executables that gives users more details about the command-line arguments that can be passed at runtime.
