### 2022-11-04 #4277 @bastorer
Hi @kthyng I've submitted a couple commits and text responses to the GitHub issues that @NoraLoose posted on the FlowSieve page (https://github.com/husseinaluie/FlowSieve/issues).
