### 2022-11-04 husseinaluie/FlowSieve#13 @NoraLoose
> Re: tutorials

Great, thanks for updating the tutorials! I think users will be really grateful for those.

> Re: testing
>
> I have some ideas for resurrecting the testing framework, but building up a proper framework is going to take a decent bit of time. With the tutorial Jupyter notebooks including the figures from when I generated them, would those qualify as "manual steps described so that the functionality of the software can be verified?"

I am curious what our editor @kthyng thinks. How strict is JOSS about having a proper testing framework, Kristen?
