### 2022-11-04 husseinaluie/FlowSieve#13 @bastorer
That's correct, "Tests" is mostly non-functional at the moment. Rebuilding it into a proper testing framework is on the to-do list, and I have some specific tests in mind, but it's unfortunately necessarily a lower priority on the list right now.

I'll add in a readme right now to avoid confusion, and will transfer them into a new `Testing-development` branch after taking stock of what all is there / what state they're in.

I'll also add a note near the top of the ReadTheDocs landing page that mentions that the tutorials provide a way to verify that the code is working as expected.
