+++ RTD_mainpage.md
@@ -40,6 +40,12 @@ following the mathematical results in Aluie 2019.
 efficiently analyse large oceanic and atmospheric datasets, and extract scientifically meaningful
 diagnostics, including scale-wise energy content and energy transfer.
 
+## Verifying Installation
+
+The tutorials, in addition to providing introductory instruction to using FlowSieve, also provide a way to verify that your installation is working as expected. 
+The provided Jupyter notebooks include the figures that were generated by the developers, and provide a reference.
+As always, feel free to contact the developers for assistance (see Community Guidelines below).
+
 ---
 
 ## Community Guidelines
