+++ Tests/README.md
@@ -1,0 +1,3 @@
+## Tests
+
+**NOTE** The testing framework is currently obsolete / non-functional. Please use the Tutorials to verify that FlowSieve is working as expected.
