### 2022-11-07 2e67d28
Fixing a fatal error that accidentally came in from a 'help' flag setting.
