### 2022-11-10 husseinaluie/FlowSieve#16 @bastorer
@NoraLoose the scalars tutorial has been added and the documentation and paper includes the line

> computes coarse-grained scalar and vector fields for arbitrary filter scales, in both Cartesian and spherical coordinates

in the core features  section.
