### 2022-11-11 husseinaluie/FlowSieve#15 @NoraLoose
Looks great, thanks for adding this. I'm closing this issue.
