### 2022-11-11 husseinaluie/FlowSieve#17 @NoraLoose
Thanks for making these changes. I will go through the updated tutorials today.

Two quick thoughts:

1. Would it be possible to change the tutorial structure in the documentation to reflect different sections for the different tutorials? Right now, it looks like they are all under the section "Basic Tutorial":
![Screen Shot 2022-11-11 at 8 21 10 AM](https://user-images.githubusercontent.com/23617395/201370700-190833bd-e913-4816-95b5-7e772279de0c.png)

2. I wonder if it is possible to integrate the notebooks (statically) into the documentation, so users can see the figures when cruising through the documentation. No worries if this is too hard to do! But could potentially be a nice addition.
