### 2022-11-11 husseinaluie/FlowSieve#17 @bastorer
1. Oh, hmm, I'll try to fix that. Doxygen table of contents are always more finicky than I expect haha, but that would be a good thing to have fixed! I'll investigate :-)
2. Integrating the notebooks statically probably won't work, but saving the figures as pngs statically and embedding them in the documentation shouldn't be an issue. I'll try to see how to do that.
