### 2022-11-16 husseinaluie/FlowSieve#13 @NoraLoose
I also suggest to remove the "Unit testing" section from the documentation, as long as the tests are not functional:
https://flowsieve.readthedocs.io/en/latest/unittests1.html#autotoc_md74

This could otherwise lead to confusion.
