### 2022-11-16 husseinaluie/FlowSieve#17 @NoraLoose
When I try to compile `coarse_grain_helmholtz.x` in the small Helmholtz tutorial, I get the following error:

```
Functions/Helmholtz/filtering_helmholtz.cpp(1191): error: namespace "constants" has no member "COMP_WIND_FORCE"
          if ( constants::COMP_WIND_FORCE ){
                          ^
```
