### 2022-11-16 husseinaluie/FlowSieve#19 @NoraLoose
Scalar Tutorial
Just two quick questions regarding the "Scalar Tutorial":

- [x]  Is [FiniteDiff.py](https://github.com/husseinaluie/FlowSieve/blob/master/Tutorial/Scalars/FiniteDiff.py) needed for the tutorial?
- [x]  Is [submit_filtering.sh](https://github.com/husseinaluie/FlowSieve/blob/master/Tutorial/Scalars/submit_filtering.sh) up to date? It specifies to filter a file `velocity_sample.nc`.

Tagging https://github.com/openjournals/joss-reviews/issues/4277
