### 2022-12-01 f848aaf
Updating files for Scalar tutorial
Re Issue #19
