+++ Tutorial/Scalars/submit_filtering.sh
@@ -1,11 +1,13 @@
 #!/bin/bash
 #SBATCH --output=sim-%j.log
 #SBATCH --error=sim-%j.err
-#SBATCH --time=00-00:10:00         # time (DD-HH:MM:SS)
+#SBATCH --time=00-00:15:00         # time (DD-HH:MM:SS)
 #SBATCH --ntasks=1
 #SBATCH --cpus-per-task=1
 #SBATCH --mem-per-cpu=4G
 #SBATCH --job-name="test"
+
+set -ex
 
 echo "${SLURM_NTASKS} MPI processors with ${SLURM_CPUS_PER_TASK} threads each"
 
@@ -13,9 +15,16 @@ export OMP_NUM_THREADS=${SLURM_CPUS_PER_TASK}
 export KMP_AFFINITY="compact"
 export I_MPI_PIN_DOMAIN="auto"
 
-FILTER_SCALES="100e3 250e3"
+##
+## Generate sample data
+##
+python generate_data_sphere.py
 
-mpirun -n ${SLURM_NTASKS} ./coarse_grain_scalars.x \
-        --input_file ./velocity_sample.nc \
-        --filter_scales "${FILTER_SCALES}" \
-        --variables "uo vo psi"
+##
+##  Apply coarse-graining
+##
+./coarse_grain_scalars.x \
+    --input_file ./density_sample.nc \
+    --filter_scales "100e3 500e3 2000e3" \
+    --variables "rho"
+
