+++ Tutorial/Scalars/FiniteDiff.py
@@ -1,197 +1,0 @@
-import numpy as np
-import numpy.linalg as nlg
-import scipy.sparse as sp
-try:
-    from scipy.misc import factorial
-except:
-    from scipy.special import factorial
-
-## FiniteDiff computes FD matrices
-## ------
-## Dr = FiniteDiff(x, Nx, kwargs)
-## ------
-#
-# kwargs: [default]
-#   Sparse = {[True],False}      # Whether or not to return sparse form
-#   Uniform = {[True], False} # If domain is Uniform
-#   DiffOrd = integer [2]       # Which derivative (i.e. 1 = 1st, 2 = 2nd)
-#   Periodic = {True, [False]} # If domain is periodic
-#
-## Notes
-#
-#   Periodic option is only available for Uniform domains.
-def FiniteDiff(x, n, Sparse=True, Uniform=True, DiffOrd = 1, Periodic=False):
-
-    #FiniteDiff : Create a finite difference matrix of arbitrary order for an
-    #             arbitrary grid.
-    n = int(n)
-    if len(x) == 3:
-        # Using a length of 3 is shorthand. x = [a,b,c] is interpreted as
-        # x = linspace(a,b,c). The advantage is that we don't actually need
-        # to generate x, so we can save on memory when we want large grids.
-        Nx = x[2]
-    else:
-        Nx = len(x)
-
-    # Adjust for higher order derivatives
-    n = n + (DiffOrd - 1)
-    if DiffOrd == 0:
-        Dx = np.eye((Nx,Nx))
-        return Dx
-
-    if Sparse:
-        Dx = sp.lil_matrix((Nx, Nx))
-    else:
-        Dx = np.zeros([Nx, Nx])
-
-    n2 = n//2
-    cn2 = int(np.ceil(n/2.0))
-    fn2 = int(np.floor(n/2.0))
-
-    if Uniform:
-        if len(x) == 3:
-            # Using a length of 3 is shorthand. x = [a,b,c] is interpreted as
-            # x = linspace(a,b,c). The advantage is that we don't actually need
-            # to generate x, so we can save on memory when we want large grids.
-            dx = (x[1] - x[0])/float((x[2]-1))
-        else:
-            dx = x[1] - x[0];
-        
-        # Deal with boundary issues
-        if not(Periodic):
-            for i in range(0,cn2):
-                A = np.zeros([n+1,n+1])
-                for j in range(0,n+1):
-                    A[:,j] = np.power(((j-i)*dx)*np.ones([1,n+1]),range(0,n+1))/factorial(range(0,n+1))
-                b = np.zeros(n+1)
-                b[DiffOrd] = 1
-                coeff = nlg.solve(A,b)
-                coeff = coeff.conj().transpose()
-                Dx[i, 0:n+1] = coeff
-        
-            for i in range(Nx-cn2,Nx):
-                A = np.zeros([n+1,n+1])
-                for j in range(Nx-n-1,Nx):
-                    A[:,j-Nx+n+1] = np.power(((j-i)*dx)*np.ones([1,n+1]),range(0,n+1))/factorial(range(0,n+1))
-                b = np.zeros(n+1)
-                b[DiffOrd] = 1
-                coeff = nlg.solve(A,b)
-                coeff = coeff.conj().transpose()
-                Dx[i, Nx-n-1:Nx] = coeff
-        
-        # Now do the internals.
-        A = np.zeros([n+1,n+1])
-        if n % 2 == 0: # If even...
-            for j in range(-n//2-1, n//2):
-                A[:,j+n//2+1] = np.power(((j+1)*dx)*np.ones([1,n+1]),range(0,n+1))/factorial(range(0,n+1))
[…]
