### 2022-12-01 husseinaluie/FlowSieve#19 @bastorer
Thanks for catching those. FiniteDiff was not needed (and has been removed) and the submit script has been updated to match the run_all_steps script :-)
