### 2022-12-01 husseinaluie/FlowSieve#17 @bastorer
COMP_WIND_FORCE should be included in the constants file on the main branch (as of commit d951de57f9dcd7ebcbf69977223367858a12076c on Sept 21), so that shouldn't be coming up as an issue, but if you're still getting it, please let me know!
