### 2022-12-07 husseinaluie/FlowSieve#17 @NoraLoose
Ok, great. Compilation works fine now. But when running `./run_all_steps.sh`, I get stuck after
```
Start-up time  = 0.02928596735001
Filtering time = 270.4323036373
   (clock resolution = 1e-09)
```
Is this normal?
