### 2022-12-07 husseinaluie/FlowSieve#14 @NoraLoose
Let's try to work toward closing this issue.

- [ ] I'll leave it up to you if you want to include a jupyter notebook for this tutorial or not.
- [ ] But I think providing an `environment.yml` for the required python dependencies (for all tutorials together, so I would place it [here](https://github.com/husseinaluie/FlowSieve/tree/master/Tutorial) with some instructions) would be very helpful - then I can also tick off the following reviewer box :)
```
Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
```
