### 2022-12-08 ef321a2
Adding environment file for tutorial python libraries
Re Issue #14
