### 2022-12-08 fca25bf
Adding notebook to view Basic Tutorial results
Re Issue #14
