+++ Tutorial/Spherical_Demo/Helmholtz/project/run_all_Helmholtz_steps.sh
@@ -80,3 +80,6 @@ echo "\n\n FINE HELMHOLTZ Ui\n\n"
 ###
 
 python analyze_projection.py
+
+##
+echo "Process successfully completed!"
