+++ Tutorial/Spherical_Demo/Low_Resolution/run_all_steps.sh
@@ -60,3 +60,7 @@ do
         --file_pattern "outputs/postprocess_${KIND}_*nc" \
         --output_filename "RESULTS_${KIND}.nc"
 done
+
+##
+
+echo "Process successfully completed!"
