### 2022-12-08 b71f19d
Adding figure of unfiltered density to notebook
Re Issue #19
