### 2022-12-08 husseinaluie/FlowSieve#19 @bastorer
Good idea, I've added the figure :-)
