### 2022-12-08 husseinaluie/FlowSieve#17 @bastorer
The `Filtering time = ...` is printed when the coarse-grain routine finishes. This should then be followed with a bash for-loop that merges the files together, which should print something like `Attempting to merge all files matching pattern outputs/postprocess_full_*nc into output file RESULTS_full.nc.`

If after running you have files `RESULTS_full.nc`, `RESULTS_toroidal.nc`, and `RESULTS_potential.nc`, then the process ran successfully. [I've added an echo statement to the end of the "run all steps" routines to announce successful completion.]

When I run `./run_all_steps.sh`, the full output that I get is:
```
++ python generate_data_sphere.py
++ python define_geographic_regions.py
++ ./Helmholtz_projection.x --input_file ./velocity_sample.nc --seed_file zero --output_file projection_ui.nc --max_iterations 1500 --tolerance 1e-18 --Tikhov_Laplace 1
 Commandline flag "--input_file" got value "./velocity_sample.nc"
 Commandline flag "--output_file" got value "projection_ui.nc"
 Commandline flag "--seed_file" got value "zero"
 Commandline flag "--time" received no value - will use default "time"
 Commandline flag "--depth" received no value - will use default "depth"
 Commandline flag "--latitude" received no value - will use default "latitude"
 Commandline flag "--longitude" received no value - will use default "longitude"
 Commandline flag "--is_degrees" received no value - will use default "true"
 Commandline flag "--Nprocs_in_time" received no value - will use default "1"
 Commandline flag "--Nprocs_in_depth" received no value - will use default "1"
 Commandline flag "--zonal_vel" received no value - will use default "uo"
 Commandline flag "--merid_vel" received no value - will use default "vo"
 Commandline flag "--tor_seed" received no value - will use default "Psi_seed"
 Commandline flag "--pot_seed" received no value - will use default "Phi_seed"
 Commandline flag "--tolerance" got value "1e-18"
 Commandline flag "--max_iterations" got value "1500"
 Commandline flag "--Tikhov_Laplace" got value "1"
 Commandline flag "--use_mask" received no value - will use default "false"
 Commandline flag "--use_area_weight" received no value - will use default "true"

Compiled at 17:04:51 on Dec  8 2022.
  Version 3.1.1

Using spherical coordinates.
 Nproc(time, depth) = (1, 1)

deriv_scale_factor = 8.59437

Termination counts: 0 from absolute tolerance
                    0 from relative tolerance
                    1 from iteration maximum
                    0 from rounding errors
                    0 from other causes

Process completed.

++ FILTER_SCALES='
1.e5 1.29e5 1.67e5 2.15e5 2.78e5 3.59e5 4.64e5 5.99e5 7.74e5
1.e6 1.29e6 1.67e6 2.15e6 2.78e6 3.59e6 4.64e6 5.99e6 7.74e6
1.e7 1.29e7 1.67e7 2.15e7 2.78e7 3.59e7 4.64e7 5.99e7 7.74e7
'
++ mkdir -p outputs
++ cd outputs
++ ../coarse_grain_helmholtz.x --Helmholtz_input_file ../projection_ui.nc --velocity_input_file ../velocity_sample.nc --tor_field Psi --pot_field Phi --vel_field uo --region_definitions_file ../region_definitions.nc --filter_scales '
1.e5 1.29e5 1.67e5 2.15e5 2.78e5 3.59e5 4.64e5 5.99e5 7.74e5
1.e6 1.29e6 1.67e6 2.15e6 2.78e6 3.59e6 4.64e6 5.99e6 7.74e6
1.e7 1.29e7 1.67e7 2.15e7 2.78e7 3.59e7 4.64e7 5.99e7 7.74e7
'
 Commandline flag "--Helmholtz_input_file" got value "../projection_ui.nc"
 Commandline flag "--velocity_input_file" got value "../velocity_sample.nc"
 Commandline flag "--radial_velocity_input_file" received no value - will use default "NONE"
 Commandline flag "--time" received no value - will use default "time"
 Commandline flag "--depth" received no value - will use default "depth"
 Commandline flag "--latitude" received no value - will use default "latitude"
 Commandline flag "--longitude" received no value - will use default "longitude"
 Commandline flag "--is_degrees" received no value - will use default "true"
 Commandline flag "--use_depth_derivs" received no value - will use default "false"
 Commandline flag "--depth_is_elevation" received no value - will use default "false"
 Commandline flag "--Nprocs_in_time" received no value - will use default "1"
 Commandline flag "--Nprocs_in_depth" received no value - will use default "1"
 Commandline flag "--tor_field" got value "Psi"
 Commandline flag "--pot_field" got value "Phi"
 Commandline flag "--vel_field" got value "uo"
 Commandline flag "--u_r_field" received no value - will use default "u_r"
 Commandline flag "--region_definitions_file" got value "../region_definitions.nc"
 Commandline flag "--region_definitions_dim" received no value - will use default "region"
 Commandline flag "--region_definitions_var" received no value - will use default "region_definition"
 Commandline flag "--filter_scales" got value "
1.e5 1.29e5 1.67e5 2.15e5 2.78e5 3.59e5 4.64e5 5.99e5 7.74e5
1.e6 1.29e6 1.67e6 2.15e6 2.78e6 3.59e6 4.64e6 5.99e6 7.74e6
1.e7 1.29e7 1.67e7 2.15e7 2.78e7 3.59e7 4.64e7 5.99e7 7.74e7
"
Filter scales (27) are:  100km,  129km,  167km,  215km,  278km,  359km,  464km,  599km,  774km,  1000km,  1290km,  1670km,  2150km,  2780km,  3590km,  4640km,  5990km,  7740km,  10000km,  12900km,  16700km,  21500km,  27800km,  35900km,  46400km,  59900km,  77400km,

Compiled at 17:04:51 on Dec  8 2022.
  Version 3.1.1

Using spherical coordinates.
 Nproc(time, depth) = (1, 1)

Scale 1 of 27 (100 km)
....|....|....|....|

Scale 2 of 27 (129 km)
....|....|....|....|

Scale 3 of 27 (167 km)
....|....|....|....|

Scale 4 of 27 (215 km)
....|....|....|....|

Scale 5 of 27 (278 km)
....|....|....|....|

Scale 6 of 27 (359 km)
....|....|....|....|

Scale 7 of 27 (464 km)
....|....|....|....|

Scale 8 of 27 (599 km)
....|....|....|....|

Scale 9 of 27 (774 km)
....|....|....|....|

Scale 10 of 27 (1000 km)
....|....|....|....|

Scale 11 of 27 (1290 km)
....|....|....|....|

Scale 12 of 27 (1670 km)
....|....|....|....|

Scale 13 of 27 (2150 km)
....|....|....|....|

Scale 14 of 27 (2780 km)
....|....|....|....|

Scale 15 of 27 (3590 km)
....|....|....|....|

Scale 16 of 27 (4640 km)
....|....|....|....|

Scale 17 of 27 (5990 km)
....|....|....|....|

Scale 18 of 27 (7740 km)
....|....|....|....|

Scale 19 of 27 (10000 km)
....|....|....|....|

Scale 20 of 27 (12900 km)
....|....|....|....|

Scale 21 of 27 (16700 km)
....|....|....|....|

Scale 22 of 27 (21500 km)
....|....|....|....|

Scale 23 of 27 (27800 km)
....|....|....|....|

Scale 24 of 27 (35900 km)
....|....|....|....|

Scale 25 of 27 (46400 km)
....|....|....|....|

Scale 26 of 27 (59900 km)
....|....|....|....|

Scale 27 of 27 (77400 km)
....|....|....|....|

Process completed.

Start-up time  = 0.08102444186807
Filtering time = 272.5161228143
   (clock resolution = 3.84612007634e-10)
++ cd ..
++ KINDS=("full" "toroidal" "potential")
++ for KIND in '"${KINDS[@]}"'
++ python ../../../PythonTools/OutputHandling/merge_postprocess_results.py --file_pattern 'outputs/postprocess_full_*nc' --output_filename RESULTS_full.nc
Attempting to merge all files matching pattern outputs/postprocess_full_*nc into output file RESULTS_full.nc.
  Identified 27 files for merging.
  Identified 5 dimensions and 29 variables to copy into merged file.
  .. copying dimension time
  .. copying dimension depth
  .. copying dimension latitude
  .. copying dimension longitude
  .. copying dimension region
  Preparing to copy variables...
  .... | .... | .... | ....
Done.
++ for KIND in '"${KINDS[@]}"'
++ python ../../../PythonTools/OutputHandling/merge_postprocess_results.py --file_pattern 'outputs/postprocess_toroidal_*nc' --output_filename RESULTS_toroidal.nc
Attempting to merge all files matching pattern outputs/postprocess_toroidal_*nc into output file RESULTS_toroidal.nc.
  Identified 27 files for merging.
  Identified 5 dimensions and 29 variables to copy into merged file.
  .. copying dimension time
  .. copying dimension depth
  .. copying dimension latitude
  .. copying dimension longitude
  .. copying dimension region
  Preparing to copy variables...
  .... | .... | .... | ....
Done.
++ for KIND in '"${KINDS[@]}"'
++ python ../../../PythonTools/OutputHandling/merge_postprocess_results.py --file_pattern 'outputs/postprocess_potential_*nc' --output_filename RESULTS_potential.nc
Attempting to merge all files matching pattern outputs/postprocess_potential_*nc into output file RESULTS_potential.nc.
  Identified 27 files for merging.
  Identified 5 dimensions and 29 variables to copy into merged file.
  .. copying dimension time
  .. copying dimension depth
  .. copying dimension latitude
  .. copying dimension longitude
  .. copying dimension region
  Preparing to copy variables...
  .... | .... | .... | ....
Done.
```
