### 2022-12-08 husseinaluie/FlowSieve#14 @bastorer
Notebook added :-)

I've added an environment file listing the versions that I used why I ran the tutorials. I'm not very familiar with .yml files, but I *think* I've got it formatted correctly.
