### 2022-12-08 f4753fb
Adding the pre-generated figures to the notebook
Re Issue #14
