### 2022-12-09 husseinaluie/FlowSieve#17 @yguo20ncsu
Hi Ben,

When I run the small tutorial on the sphere, I always get stuck at:
Start-up time  = 0.03308053314686
Filtering time = 270.9834141545
   (clock resolution = 1e-09)

I (of course) can't find the files RESULTS_FULL.nc etc. Do you have any insights on this? It seems Nora had the same issue, I am not sure if she got it resolved. Thank you very much!

Thanks,
Yiming
