### 2022-12-09 husseinaluie/FlowSieve#17 @bastorer
That's very peculiar! It sounds like the bash `for` loop might not be running.

Could you
1) confirm if you have an `outputs` directory that was created with a bunch of files that look like `postprocess_..._##km.nc`
2) try manually running one of the python merging routines to make the `RESULTS_full.nc` file?

The command for 2 should be
```
python ../../../PythonTools/OutputHandling/merge_postprocess_results.py --file_pattern "outputs/postprocess_full_*nc" --output_filename "RESULTS_full.nc"
```
