### 2022-12-09 husseinaluie/FlowSieve#17 @yguo20ncsu
1. the outputs directory with postprocess_..._##km.nc exists.
2. maybe the error comes from missing this keyword when create variables.

Traceback (most recent call last):
  File "../../../PythonTools/OutputHandling/merge_postprocess_results.py", line 144, in <module>
    nc_var_objs[varname].setncattr(attr, all_vars[varname].getncattr( attr ) )
  File "netCDF4/_netCDF4.pyx", line 4123, in netCDF4._netCDF4.Variable.setncattr
AttributeError: _FillValue attribute must be set when variable is created (using fill_value keyword to createVariable)
