### 2022-12-12 #4277 @kthyng
@bastorer I know it's not ideal but I think it's unlikely we'll find someone with the time to review your software before potentially breaking for holidays. However, it's great that we have a second reviewer lined up for the new year now!

Also, I was just looking at your repo. I think it'd be good for your software to have more information in the readme about what it is for — currently it is about citations, etc, which is good info to have there too. But I'd suggest prefacing that information with a brief description of what it does.
