### 2022-12-12 husseinaluie/FlowSieve#14 @NoraLoose
You can check whether your conda environment works correctly by following the steps [here](https://conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html#creating-an-environment-from-an-environment-yml-file).
* A quick test tells me that `cmocean` is missing in the `environment.yml` file. There may be other packages missing. Could you check if all tutorials are running after activating the `FlowSieve-tutorial-env` environment? Thanks!
* It would be great if you could add instructions to https://github.com/husseinaluie/FlowSieve/blob/master/Tutorial/TUTORIAL.md on how to create and activate the conda environment before running any of the tutorials.
