### 2022-12-12 #4277 @NoraLoose
Hi @kthyng, here is a summary of the issues that @bastorer and I have worked on over the past few months:

* https://github.com/husseinaluie/FlowSieve/issues/5
* https://github.com/husseinaluie/FlowSieve/issues/6
* https://github.com/husseinaluie/FlowSieve/issues/7
* https://github.com/husseinaluie/FlowSieve/issues/8
* https://github.com/husseinaluie/FlowSieve/issues/9
* https://github.com/husseinaluie/FlowSieve/issues/10
* https://github.com/husseinaluie/FlowSieve/issues/13
* https://github.com/husseinaluie/FlowSieve/issues/15
* https://github.com/husseinaluie/FlowSieve/issues/16
* https://github.com/husseinaluie/FlowSieve/issues/19
* https://github.com/husseinaluie/FlowSieve/issues/14
* https://github.com/husseinaluie/FlowSieve/issues/17

I think we are pretty close. I am just waiting on some last fixes from Ben concerning the remaining 2 open issues. I'm happy if we can wrap it up before the break.
