### 2022-12-12 husseinaluie/FlowSieve#17 @bastorer
Okay, it sounds like the coarse-graining part itself runs, which is good, it's just the output-handling that fails.

Could you tell me what version of netcdf and netCDF4 (the python library) you have? I would have hoped it wasn't that sensitive to those, but maybe?
