### 2022-12-12 husseinaluie/FlowSieve#17 @bastorer
Thanks! I'm at AGU this week but will try digging into this tonight.
