### 2022-12-13 husseinaluie/FlowSieve#17 @bastorer
Okay, it looks like somewhere in the change to v1.5 they required fill_value to be set at the time of variable creation.

I've updated the merging routines and tested them on v1.5.8 (what I had installed on my laptop), and they run as expected. When you get a chance, can you try pulling the latest commits and running the merge step again?
