### 2022-12-13 husseinaluie/FlowSieve#17 @yguo20ncsu
Hi Ben,

Thanks for updating the package. Unfortunately, when running ./run_all_steps.sh it is still stuck at the same place. But when manually run the python merging routine, I can get the merged data file this time.

Yiming
