### 2022-12-14 husseinaluie/FlowSieve#17 @bastorer
@yguo20ncsu okay, great! I'll probably just replace the bash for loop with the three explicit python merge commands.

If you're able to get the merged data files though then it sounds like everything is running :-) !
