### 2023-01-03 husseinaluie/FlowSieve#17 @bastorer
@yguo20ncsu to test if it is just the loop not working, could you try running a bash script with the following contents:

```
KINDS=("full" "toroidal" "potential")

for KIND in "${KINDS[@]}"
do
    echo ${KIND}
done
```

Have you run into any other problems running the code / tutorials?

Thanks!
