### 2023-01-03 #4277 @bastorer
@NoraLoose That's very exciting, congratulations! Absolutely, I'm working through the conda environments right now :-)
