### 2023-01-03 aa04c19
Adding anaconda environment file for tutorials
Re Issue #14

Added environment.yml to the main tutorial directory, which should improve reproducibility and
portability. Tested using conda v4.8

Also removing some old and unused PlotTools scripts
