+++ PythonTools/PlotTools/__init__.py
@@ -1,10 +1,5 @@
 
-from .SignedLogPlot import SignedLogPlot
-from .SignedLogPlot_onMap import SignedLogPlot_onMap
 from .ScientificCbar import ScientificCbar
-from .MapProjection import MapProjection
-from .SignedLogScatter import SignedLogScatter
-from .SignedLogScatter_hist import SignedLogScatter_hist
 from .merge_to_mp4 import merge_to_mp4
 from .AddParallels_and_Meridians import AddParallels_and_Meridians
 from .LabelTimeAxis import LabelTimeAxis
