+++ Tutorial/Spherical_Demo/Low_Resolution/analyse_results.py
@@ -1,0 +1,139 @@
+import numpy as np
+import matplotlib.pyplot as plt
+from netCDF4 import Dataset
+from FiniteDiff import FiniteDiff
+
+with Dataset( 'velocity_sample.nc', 'r' ) as dset:
+
+    lon = dset['longitude'][:]
+    lat = dset['latitude'][:]
+
+    uo = dset['uo'][0,0,:,:]
+    vo = dset['vo'][0,0,:,:]
+
+    Phi = dset['phi'][0,0,:,:]
+    Psi = dset['psi'][0,0,:,:]
+
+# Plot original / unfiltered fields
+fig, axes = plt.subplots( 2, 2, sharex = True, sharey = True, figsize = (7,5) )
+
+axes[0,0].pcolormesh( lon, lat, uo,  vmin = -0.1, vmax = 0.1, rasterized = True )
+axes[1,0].pcolormesh( lon, lat, vo,  vmin = -0.1, vmax = 0.1, rasterized = True )
+
+cv = np.max(np.abs(Phi))
+axes[0,1].pcolormesh( lon, lat, Phi, vmin = -cv, vmax = cv, rasterized = True )
+axes[1,1].pcolormesh( lon, lat, Psi, vmin = -cv, vmax = cv, rasterized = True )
+
+axes[0,0].set_title( "A) Zonal Velocity",      loc = 'left', pad = 3 )
+axes[1,0].set_title( "B) Meridional Velocity", loc = 'left', pad = 3 )
+axes[0,1].set_title( "C) Potential Function",  loc = 'left', pad = 3 )
+axes[1,1].set_title( "D) Streamfunction",      loc = 'left', pad = 3 )
+
+plt.savefig('original_fields.pdf', dpi = 250)
+plt.close()
+
+
+with Dataset( 'projection_ui.nc', 'r' ) as dset:
+
+    Phi_proj = dset['Phi'][0,0,:,:]
+    Psi_proj = dset['Psi'][0,0,:,:]
+
+
+# Plot the Helmholtz Projection with Errors
+fig, axes = plt.subplots( 2, 3, sharex = True, sharey = True, figsize = (10,5) )
+
+cv = np.max(np.abs(Phi))
+
+qm = axes[0,0].pcolormesh( lon, lat, Phi, vmin = -cv, vmax = cv, rasterized = True )
+axes[1,0].pcolormesh(      lon, lat, Psi, vmin = -cv, vmax = cv, rasterized = True )
+
+axes[0,1].pcolormesh( lon, lat, Phi_proj, vmin = -cv, vmax = cv, rasterized = True )
+axes[1,1].pcolormesh( lon, lat, Psi_proj, vmin = -cv, vmax = cv, rasterized = True )
+
+qm2 = axes[0,2].pcolormesh( lon, lat,  Phi                 -  Phi_proj,                      vmin = -cv / 1e2, vmax = cv / 1e2, rasterized = True )
+axes[1,2].pcolormesh(       lon, lat, (Psi - np.mean(Psi)) - (Psi_proj - np.mean(Psi_proj)), vmin = -cv / 1e3, vmax = cv / 1e3, rasterized = True )
+
+plt.colorbar( qm,  ax = axes[:,:2] )
+plt.colorbar( qm2, ax = axes[:,2] )
+
+axes[0,0].set_title( "A) Original $\Phi$", loc = 'left', pad = 3 )
+axes[1,0].set_title( "B) Original $\Psi$", loc = 'left', pad = 3 )
+
+axes[0,1].set_title( "C) Projected $\Phi$", loc = 'left', pad = 3 )
+axes[1,1].set_title( "D) Projected $\Psi$", loc = 'left', pad = 3 )
+
[…]
