+++ Tutorial/Spherical_Demo/Low_Resolution/run_all_steps.sh
@@ -51,15 +51,15 @@ cd ..
 ##      Now merge the postprocess outputs into single files
 ###
 
-
-KINDS=("full" "toroidal" "potential")
-
-for KIND in "${KINDS[@]}"
-do
-    python ../../../PythonTools/OutputHandling/merge_postprocess_results.py \
-        --file_pattern "outputs/postprocess_${KIND}_*nc" \
-        --output_filename "RESULTS_${KIND}.nc"
-done
+python ../../../PythonTools/OutputHandling/merge_postprocess_results.py \
+        --file_pattern "outputs/postprocess_full_*nc" \
+        --output_filename "RESULTS_full.nc"
+python ../../../PythonTools/OutputHandling/merge_postprocess_results.py \
+        --file_pattern "outputs/postprocess_toroidal_*nc" \
+        --output_filename "RESULTS_toroidal.nc"
+python ../../../PythonTools/OutputHandling/merge_postprocess_results.py \
+        --file_pattern "outputs/postprocess_potential_*nc" \
+        --output_filename "RESULTS_potential.nc"
 
 ##
 

[…]
