### 2023-01-03 husseinaluie/FlowSieve#14 @bastorer
I've got the conda environment working on the other server :-) I'm having some trouble getting it to play nice when I try to activate the environment inside of submitted jobs, but otherwise running in command-line in the activated environment reproduces the tutorial figures
