### 2023-01-03 28dd058
Removing bash for loops in tutorial
