+++ Tutorial/Spherical_Demo/Helmholtz/postprocessing/submit_filter_for_postprocessing.sh
@@ -47,12 +47,12 @@ mpirun -n ${SLURM_NTASKS} ./coarse_grain_helmholtz.x \
 ##      Now merge the postprocess outputs into single files
 ###
 
-
-KINDS=("full" "toroidal" "potential")
-
-for KIND in "${KINDS[@]}"
-do
-    python ../../../../PythonTools/OutputHandling/merge_postprocess_results.py \
-        --file_pattern "postprocess_${KIND}_*nc" \
-        --output_filename "RESULTS_${KIND}.nc"
-done
+python ../../../../PythonTools/OutputHandling/merge_postprocess_results.py \
+    --file_pattern "postprocess_full_*nc" \
+    --output_filename "RESULTS_full.nc"
+python ../../../../PythonTools/OutputHandling/merge_postprocess_results.py \
+    --file_pattern "postprocess_toroidal_*nc" \
+    --output_filename "RESULTS_toroidal.nc"
+python ../../../../PythonTools/OutputHandling/merge_postprocess_results.py \
+    --file_pattern "postprocess_potential_*nc" \
+    --output_filename "RESULTS_potential.nc"
