### 2023-01-03 husseinaluie/FlowSieve#17 @bastorer
@NoraLoose I've removed the bash for loops and just replaced them with manual repetitions of the command, since the loops don't seem to be as portable as I'd hoped!
