### 2023-01-04 husseinaluie/FlowSieve#17 @yguo20ncsu
@bastorer
Hi Ben, actually the loop works good in this test bash script. I don't have any other problems running the code so far (I also have replaced the loop with all commands manually). Thank you so much for your help!
Thanks!
