### 2023-01-05 husseinaluie/FlowSieve#17 @NoraLoose
@bastorer I tried running `./Tutorial/Spherical_Demo/Low_Resolution/run_all_steps.sh` again, but still have the same problem. Does this file still need to be updated?
