### 2023-01-05 husseinaluie/FlowSieve#17 @bastorer
The bash loop has been replaced by just directly calling the three python commands. Are you getting any specific errors about the python crashing?
