### 2023-01-05 husseinaluie/FlowSieve#17 @NoraLoose
No crashing, I simply get stuck after:

```
Scale 27 of 27 (77400 km)
....|....|....|....|

Process completed.

Start-up time  = 0.02932177111506
Filtering time = 272.1048071757
   (clock resolution = 1e-09)
```
