### 2023-01-05 husseinaluie/FlowSieve#17 @NoraLoose
@yguo20ncsu, does this still happen to you too, when running `./Tutorial/Spherical_Demo/Low_Resolution/run_all_steps.sh`?
