### 2023-01-05 husseinaluie/FlowSieve#17 @yguo20ncsu
Yes, I also still get stuck at the same place when running run_all_steps.sh. Ben mentioned that it is probably something wrong with the bash loop where all nc files in "outputs" directory are merged into RESULTS_*.nc. Below is how to manually solve it as Ben suggested.

1. I commented the loop in the run_all_steps.sh as:
##KINDS=("full" "toroidal" "potential")
###for KIND in "${KINDS[@]}"
###do
###        --output_filename "RESULTS_${KIND}.nc"
###done

2. Instead of using bash loop, manually run the three python merging routines as:
python ../../../PythonTools/OutputHandling/merge_postprocess_results.py --file_pattern "outputs/postprocess_full_*nc" --output_filename "RESULTS_full.nc"
python ../../../PythonTools/OutputHandling/merge_postprocess_results.py --file_pattern "outputs/postprocess_toroidal_*nc" --output_filename "RESULTS_toroidal.nc"
python ../../../PythonTools/OutputHandling/merge_postprocess_results.py --file_pattern "outputs/postprocess_potential_*nc" --output_filename "RESULTS_potential.nc"

3. then you will see the merged nc files (RESULTS_*.nc)
