### 2023-01-05 husseinaluie/FlowSieve#17 @bastorer
After a bit of reading, I suspect it might be `set -e` that's the problem. Ideally that would halt the script whenever an error occurs. However, it sounds like it can be a little over zealous in what it classifies as an 'error', and those classifications can change between version of bash.

If you agree (and if replacing `set -ex` with `set -x` fixes the problem for you), I'll replace the `set -ex` with `set -x` in the bash scripts in the tutorial.
