### 2023-01-05 husseinaluie/FlowSieve#17 @NoraLoose
Hmm, I replaced `set -ex` with `set -x` in line 5 of `./Tutorial/Spherical_Demo/Low_Resolution/run_all_steps.sh`, but it did not fix the problem. :/
