### 2023-01-12 b56fee1
Adding new kernel option. Similar to the tanh profile, but second order so that it can capture steeper spectral slopes.
