### 2023-01-12 c659d66
Fixing a bug in get_lon_bounds that incorrectly computed the optimized kernel bounds.
The new formulation is faster, more accurate, and easier to read /
understand.

Now v3.2.0
