+++ Functions/get_lon_bounds.cpp
@@ -2,6 +2,7 @@
 #include <math.h>    
 #include <algorithm>
 #include <vector>
+#include <cassert>
 #include "../constants.hpp"
 #include "../functions.hpp"
 
@@ -30,6 +31,11 @@ void get_lon_bounds(
     const double dlon    = longitude.at( 1) - longitude.at( 0);
     const double KernPad = constants::KernPad;
     const int    Nlon    = (int) longitude.size();
+
+    // assumes uniform lon grid
+    static_assert( constants::UNIFORM_LON_GRID, "Currently required uniform lon grid." );
+
+    int dlon_N;
     
     if (KernPad < 0) {
         // KernPad < 0 means we use the entire domain, so don't bother
@@ -57,43 +63,54 @@ void get_lon_bounds(
             const double delta_lat = fabs( centre_lat - curr_lat ); 
             const double square_diff = pow(padded_scale, 2.) - pow(delta_lat, 2.);
             local_scale = sqrt( fmax( square_diff, 0. ) );
+
+            dlon_N = int( std::min( (double) Nlon, ceil( ( local_scale / dlon_m ) / 2.) ) ); 
         }
         else { 
-            // Spherical law of cosines, simplified because we know
-            //   that lines of longitude are perpendicular to
-            //   lines of latitude
-            const double delta_lat = constants::R_earth * fabs( centre_lat - curr_lat ); 
-            if ( delta_lat < padded_scale ) { 
-                // Ratios subtended by the padded filter scale, and the current latitudinal extent
-                const double    padded_angle = padded_scale / constants::R_earth,
-                                lat_angle    = delta_lat / constants::R_earth;
 
-                if ( (padded_angle >= M_PI / 2) or (lat_angle >= M_PI / 2) ) {
-                    // If the padded angle exceeds pi / 2, then the arccos breaks.
-                    // For now, just make the whole sphere the integration range in this case.
-                    // This is more expensive than it needs to be! But need to sort out the
-                    // trig / spherical law of cosines for other quadrants.
-                    local_scale = 2 * M_PI * constants::R_earth;
-                } else {
-                    const double cosrat = cos( padded_angle ) / cos( lat_angle );
-                    local_scale = acos(cosrat) * constants::R_earth;
-                }
+            // lambda0, phi0 are coordinates for centre of kernel "centre"
+            // phi1 is latitude of current integration area "curr"
 
+            
+            // This is the distance between the centre point (centre_lat, centre_lon)
+            // and the point at this latitude on the other side (curr_lat, centre_lon + pi)
+            // dist( ( lambda0, phi0 ), ( lambda0 + pi, phi1 ) )
+            const double dist_on_other_side = distance( longitude.at(Ilon), centre_lat,
+                                                        longitude.at(Ilon) + M_PI, curr_lat );
+
+            // dist( ( lambda0, phi0 ), ( lambda0, phi1 ) )
+            const double dist_to_lat1 = distance( longitude.at(Ilon), centre_lat,
+                                                  longitude.at(Ilon), curr_lat );
+
+            if ( dist_on_other_side <= padded_scale / 2. ) {
+                // Use all points on line of latitude phi1
+                dlon_N = Nlon;
+            } else if ( dist_to_lat1 >= padded_scale / 2 ) {
+                // Do not use any points on line of latitude phi1
+                dlon_N = 1;
             } else {
-                local_scale = 0.;
+                // Else, there must be some del_lon in (0,pi) such that
+                // dist( (lambda0, phi0), (lambda0+del_lon, phi1) ) = ell/2
[…]
