### 2023-01-18 husseinaluie/FlowSieve#14 @bastorer
Hi Nora,

With the environment.yml and jupyter notebooks set up, is there anything else that you think needs to be addressed for this ticket?

Thanks!
