### 2023-01-18 husseinaluie/FlowSieve#17 @bastorer
Hi Nora,

As far as I can tell, this is an issue with bash scripting that I'm honestly not sure how to solve. All of the parts themselves run correctly, the final python scripts just need to be called separately.

If I add a note in the tutorial documentation highlighting this (essentially saying that if the `RESULTS_*.nc` files are missing, run `<insert python code here>`), would that be sufficient?

It's not ideal, but I'm not sure what else to do for the bash script.

Thanks!
