### 2023-01-18 930469e
Adding comments outlining that python postprocessing does not always run in the tutorial scripts.
It's not clear why this happens, other than seemingly being a bash
issue.

The simplest fix is to simply run those three python scripts manually in
terminal (i.e. just copy from the script into temrinal).

Re Issue #17
