### 2023-01-20 husseinaluie/FlowSieve#14 @NoraLoose
Cool, I just tested the conda environment, and am able to run all python scripts. Closing this issue!
