### 2023-01-20 #4277 @NoraLoose
Hi @kthyng, @bastorer!

All issues have been addressed, and I checked the remaining boxes in the review list. I am happy for the software to be published in JOSS. :tada:
