### 2023-01-21 #4277 @ashwinvis
> Hi @ashwinvis
>
> Just checking if this is an in-progress checklist, or if the unchecked items are ones where you have concerns about FlowSieve.
>
> Thanks!

Working on it. I did notice that the installation docs are insufficient. It does not list a list of external dependencies and libraries required (for eg. Netcdf, I dont know if there are more).
