### 2023-01-23 2984dcd
Reducing memory access cost for derivatives when DEBUG <= 0
