### 2023-01-23 a1f293c
Improving threading scheduler setting for diagnostics
