+++ Functions/compute_vorticity.cpp
@@ -77,7 +77,7 @@ void compute_vorticity(
     private( Itime, Idepth, Ilat, Ilon, index, vort_r_tmp, vort_lon_tmp, vort_lat_tmp, \
              div_tmp, OkuboWeiss_tmp)
     {
-        #pragma omp for collapse(1) schedule(dynamic, OMP_chunksize)
+        #pragma omp for collapse(1) schedule(guided)
         for (index = 0; index < Npts; index++) {
 
             vort_r_tmp   = 0.; 
