### 2023-01-23 39ebac3
v3.2.1 Greatly reducing synchronization cost for postprocessing
Computing region averages previously had large synchronization costs
for OpenMP parallelization. This has been greatly reduced by changing
the reduction type to allow a 5-deep loop collapse with threading
