### 2023-01-23 3392b48
Adding safety checks for zero kernel area in quadratics filtering
