### 2023-01-23 0ff8e9b
Extending mask to pole now parallelized
Minor performance improvement, since this happens once at the beggining,
but reduces initialization cost
