+++ Functions/extend_mask_to_poles.cpp
@@ -2,6 +2,7 @@
 #include <algorithm>
 #include <vector>
 #include <cassert>
+#include <omp.h>
 #include "../functions.hpp"
 #include "../constants.hpp"
 
@@ -30,15 +31,22 @@ void extend_mask_to_poles(
     std::vector<bool> extended_mask( extended_size, extend_val );
 
     int Itime, Idepth, Ilat, Ilon;
-    size_t extended_index;
+    size_t extended_index, index;
 
     // Copy from the original mask into a padded one
-    for ( size_t index = 0; index < size_to_extend; index++ ) {
-        Index1to4( index, Itime, Idepth, Ilat, Ilon, Ntime, Ndepth, Nlat, Nlon );
+    #pragma omp parallel \
+    default(none) \
+    shared( extended_mask, mask_to_extend ) \
+    private( index, Itime, Idepth, Ilat, Ilon, extended_index )
+    {
+        #pragma omp for collapse(1) schedule(static)
+        for ( index = 0; index < size_to_extend; index++ ) {
+            Index1to4( index, Itime, Idepth, Ilat, Ilon, Ntime, Ndepth, Nlat, Nlon );
 
-        extended_index = Index( Itime, Idepth, Ilat + Ilat_start, Ilon, Ntime, Ndepth, extended_Nlat, Nlon );
+            extended_index = Index( Itime, Idepth, Ilat + Ilat_start, Ilon, Ntime, Ndepth, extended_Nlat, Nlon );
 
-        extended_mask.at( extended_index ) = mask_to_extend.at( index );
+            extended_mask.at( extended_index ) = mask_to_extend.at( index );
+        }
     }
 
     // Now swap out the contents of the extended mask 
