### 2023-01-23 c25ac74
Chaing OpenMP scheduler type for computing areas
