### 2023-01-23 d168151
v3.2.2 Improvide threading performance of coarse_grain_helmholtz
Changed scheduler for loops other than the main filtering loop
to reduce synchronization cost

Change loop structure of main filtering loop to improve load balancing.
Loop now split across lat and lon, not just lat. Manual load-splitting
used to reduce the cost of repeated kernel calculations.
