+++ Functions/Helmholtz/filtering_helmholtz.cpp
@@ -77,8 +77,6 @@ void filtering_helmholtz(
     if (wRank == 0) { fprintf( stdout, "\nPreparing to apply %d filters to data with (MPI-local) sizes (%'d - %'d - %'d - %'d) \n", Nscales, Ntime, Ndepth, Nlat, Nlon ); }
     #endif
 
-    const int OMP_chunksize = get_omp_chunksize(Nlat,Nlon);
-
     char fname [50];
     
     const int ndims = 4;
@@ -293,7 +291,7 @@ void filtering_helmholtz(
             u_r, u_lon_tor, u_lat_tor, u_lon_pot, u_lat_pot, u_lon_tot, u_lat_tot) \
     private( index )
     {
-        #pragma omp for collapse(1) schedule(dynamic, OMP_chunksize)
+        #pragma omp for collapse(1) schedule(guided)
         for (index = 0; index < u_lon_tor.size(); ++index) {
             u_lon_tot.at(index) = u_lon_tor.at(index) + u_lon_pot.at(index);
             u_lat_tot.at(index) = u_lat_tor.at(index) + u_lat_pot.at(index);
@@ -327,7 +325,7 @@ void filtering_helmholtz(
                 u_lon_tor, u_lat_tor, u_lon_pot, u_lat_pot ) \
         private( index )
         {
-            #pragma omp for collapse(1) schedule(dynamic, OMP_chunksize)
+            #pragma omp for collapse(1) schedule(guided)
             for (index = 0; index < u_lon_tor.size(); ++index) {
 
                 if ( mask.at(index) ) { 
@@ -464,7 +462,8 @@ void filtering_helmholtz(
 
     // Now prepare to filter
     double scale;
-    int Itime, Idepth, Ilat, Ilon, tid;
+    int Itime, Idepth, Ilat, Ilon, Ilatlon, thread_id, num_threads, prev_Ilat = -1;
+    const bool can_roll_in_longitude = ( (constants::PERIODIC_X) and (constants::UNIFORM_LON_GRID) and (constants::FULL_LON_SPAN) );
 
     int perc_base = 5;
     int perc, perc_count=0;
@@ -667,9 +666,9 @@ void filtering_helmholtz(
                 coarse_wind_tau_Psi, coarse_wind_tau_Phi, \
                 coarse_tau_wind_dot_u_tor, coarse_tau_wind_dot_u_pot, coarse_tau_wind_dot_u_tot \
                 ) \
-        private(Itime, Idepth, Ilat, Ilon, index, \
+        private(Itime, Idepth, Ilat, Ilon, index, prev_Ilat, Ilatlon, \
                 F_tor_tmp, F_pot_tmp, u_r_tmp, uxux_tmp, uxuy_tmp, uxuz_tmp, uyuy_tmp, uyuz_tmp, uzuz_tmp, \
-                vort_ux_tmp, vort_uy_tmp, vort_uz_tmp, LAT_lb, LAT_ub, tid, filtered_vals, \
+                vort_ux_tmp, vort_uy_tmp, vort_uz_tmp, LAT_lb, LAT_ub, thread_id, num_threads, filtered_vals, \
                 uiuj_F_r_tmp, uiuj_F_Phi_tmp, uiuj_F_Psi_tmp, \
                 wind_tau_Psi_tmp, wind_tau_Phi_tmp, tau_wind_dot_u_tor_tmp, tau_wind_dot_u_pot_tmp ) \
         firstprivate(perc, wRank, local_kernel, perc_count)
@@ -697,161 +696,174 @@ void filtering_helmholtz(
                 filtered_vals.push_back( &tau_wind_dot_u_pot_tmp );
             }
 
-            tid = omp_get_thread_num();
[…]
