### 2023-01-25 d80cf25
Fixing error with dimensions in merging resolutions
