### 2023-01-25 5158f76
Minor documentation and openmp changes
