### 2023-01-25 f680415
Coarsen-grid now handles land-avoidance (i.e. land cells are fully ignored)
