### 2023-01-30 husseinaluie/FlowSieve#23 @ashwinvis
Specify minimum requirements for installation in documentation
The ones I have found so far...

- mpi
- gcc
- netcdf-mpi
- hdf5
