### 2023-01-30 husseinaluie/FlowSieve#24 @ashwinvis
Installation errors with gcc 10
With gcc 10.3.0, I get

```c
mpic++   -c -DDEBUG=0  -O3    -o NETCDF_IO/add_attr_to_file.o NETCDF_IO/add_attr_to_file.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
NETCDF_IO/add_attr_to_file.cpp: In function 'void add_attr_to_file(const char*, double, const char*, MPI_Comm)':
NETCDF_IO/add_attr_to_file.cpp:23:38: error: format not a string literal and no format arguments [-Werror=format-security]
   23 |         snprintf(buffer, 50, filename);
      |                                      ^
```

If I do `CFLAGS := -w` to suppress warnings, it goes a bit further, but errors again

```c
mpic++   -c -DDEBUG=0 -w -O3    -o NETCDF_IO/add_attr_to_file.o NETCDF_IO/add_attr_to_file.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -O3    -o NETCDF_IO/add_var_to_file.o NETCDF_IO/add_var_to_file.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -O3    -o NETCDF_IO/check_file_existence.o NETCDF_IO/check_file_existence.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -O3    -o NETCDF_IO/initialize_output_file.o NETCDF_IO/initialize_output_file.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -O3    -o NETCDF_IO/initialize_particle_file.o NETCDF_IO/initialize_particle_file.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -O3    -o NETCDF_IO/initialize_postprocess_file.o NETCDF_IO/initialize_postprocess_file.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -O3    -o NETCDF_IO/initialize_projected_particle_file.o NETCDF_IO/initialize_projected_particle_file.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -O3    -o NETCDF_IO/initialize_regions_file.o NETCDF_IO/initialize_regions_file.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -O3    -o NETCDF_IO/initialize_subset_file.o NETCDF_IO/initialize_subset_file.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -O3    -o NETCDF_IO/load_region_definitions.o NETCDF_IO/load_region_definitions.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -O3    -o NETCDF_IO/nc_err.o NETCDF_IO/nc_err.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -O3    -o NETCDF_IO/package_field.o NETCDF_IO/package_field.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
NETCDF_IO/package_field.cpp: In function 'void package_field(std::vector<short int>&, double&, double&, const std::vector<double>&, const std::vector<bool>*, MPI_Comm)':
NETCDF_IO/package_field.cpp:59:54: error: 'fmiddle' not specified in encerrorlosing 'parallel'
   59 |                 local_double = (original.at(index) - fmiddle) / frange;
      |                                                      ^~~~~~~
NETCDF_IO/package_field.cpp:49:13: error: enclosing 'parallel'
   49 |     #pragma omp parallel \
      |             ^~~
NETCDF_IO/package_field.cpp:59:65: error: 'frange' not specified in enclosing 'parallel'
   59 |                 local_double = (original.at(index) - fmiddle) / frange;
      |                                                                 ^~~~~~
NETCDF_IO/package_field.cpp:49:13: error: enclosing 'parallel'
   49 |     #pragma omp parallel \
      |             ^~~
make: *** [Makefile:78: NETCDF_IO/package_field.o] Error 1
```
