### 2023-01-30 husseinaluie/FlowSieve#23 @bastorer
Expanding on the list a bit before I add it into the documentation

- netcdf4 (tested on 4.7 and 4.8)
  - requires parallel installation
  - hdf5 required
- mpi (openmpi or intel mpi, not mpich)
- compilers: either gcc or icpc (intel compiler)
  - intended to be compiled with c++14 standard (this is specified in the compiler flags in the `system.mk` file)

Not required for FlowSieve, but required for the Tutorials (specified in `Tutorials/environment.yml`). May not require these specific version, but tested with these versions.
 - python=3.7
 - numpy=1.17
 - matplotlib=3.1
 - netCDF4=1.4
 - scipy=1.3
 - pip
