### 2023-01-30 husseinaluie/FlowSieve#24 @bastorer
Oh, it looks like gcc wants `const`s to be specified as openmp-shared (whereas intel auto-imposes shared `const`s).

Thanks! I've mostly been compiling with intel lately, but will rebuild a gcc version to hunt for other syntax mis-matches.
