### 2023-02-01 husseinaluie/FlowSieve#24 @bastorer
Ah, okay, it looks like GCC 9 and onwards changed how `const`-declared variables are handle in OpenMP clauses (https://gcc.gnu.org/gcc-9/porting_to.html) in a way the lightly breaks back-compatibility. I'm resolving the (shockingly many) warnings that pop up with GCC (but not intel, weirdly enough), but resolving the OpenMP predetermined errors shouldn't be too bad. It looks like it only hits small variables, so setting them as `firstprivate` should do it. I'm testing on GCCv7 right now, but will do a 9+ test afterwards (just need to see which of the servers have a GCCv9+ installation of the modules...).

Will hopefully be done later today.
