### 2023-02-02 6672768
Resolving the const-declared in OMP issues for GCC
