### 2023-02-02 a97a526
Even with serial, using mpirun -n 1 seems to sidestep some gcc errors
