+++ Tutorial/Basic/run_all_steps.sh
@@ -10,7 +10,7 @@ python generate_data.py
 ## Run coarse-graining
 ##
 
-./coarse_grain.x --input_file velocity_sample.nc --filter_scales "1e3 15e3 50e3 100e3"
+mpirun -n 1 ./coarse_grain.x --input_file velocity_sample.nc --filter_scales "1e3 15e3 50e3 100e3"
 
 ##
 ## Make figures from the outputs
