### 2023-02-02 husseinaluie/FlowSieve#24 @bastorer
I've updated the code base and it now compiles for me without warnings or errors with GCCv8 and GCCv9 ( I didn't have access to a GCCv10 module set that had the libraries that I needed, but v9 is on the same side of the openmp behaviour change so with a little luck v10 should work as well  )
