### 2023-02-02 755fade
Adding some print statements for DEBUG>=2
