### 2023-02-02 2192280
Fixing a bug that prevented zonal averages in atmosphere-like data
