### 2023-02-02 20cb342
Adding in some nan-catching
