### 2023-02-02 e7b9f5f
Updating outputs in SymTensor projection
