+++ Preprocess/Apply_Helmholtz_Projection_SymTensor.cpp
@@ -146,7 +146,7 @@ void Apply_Helmholtz_Projection_SymTensor(
 
             // Now handle longitude derivatives
             LB = - 2 * Nlon;
-            get_diff_vector(diff_vec, LB, longitude, "lon", Itime, Idepth, Ilat, Ilon, Ntime, Ndepth, Nlat, Nlon, mask, 1, constants::DiffOrd);
+            get_diff_vector(diff_vec, LB, longitude, "lon", Itime, Idepth, Ilat, Ilon, Ntime, Ndepth, Nlat, Nlon, unmask, 1, constants::DiffOrd);
             Ndiff = diff_vec.size();
 
             if ( LB != - 2 * Nlon) {
@@ -173,7 +173,7 @@ void Apply_Helmholtz_Projection_SymTensor(
 
             // Now handle latitude derivatives
             LB = - 2 * Nlat;
-            get_diff_vector(diff_vec, LB, latitude, "lat", Itime, Idepth, Ilat, Ilon, Ntime, Ndepth, Nlat, Nlon, mask, 1, constants::DiffOrd);
+            get_diff_vector(diff_vec, LB, latitude, "lat", Itime, Idepth, Ilat, Ilon, Ntime, Ndepth, Nlat, Nlon, unmask, 1, constants::DiffOrd);
             Ndiff = diff_vec.size();
 
             if ( LB != - 2 * Nlat) {
@@ -247,8 +247,6 @@ void Apply_Helmholtz_Projection_SymTensor(
     size_t lower_count = alglib::sparsegetlowercount( LHS_matr );
     size_t upper_count = alglib::sparsegetuppercount( LHS_matr );
 
-    fprintf( stdout, "counts: %'zu %'zu\n", lower_count, upper_count );
-
     alglib::sparseconverttocrs(LHS_matr);
 
     #if DEBUG >= 1
@@ -287,7 +285,7 @@ void Apply_Helmholtz_Projection_SymTensor(
             // Get velocity from seed
             alglib::sparsemv( LHS_matr, lhs_seed, rhs_seed );
 
-            #if DEBUG >= 1
+            #if DEBUG >= 2
             if ( (wRank == 0) and (Itime == 0) ) {
                 fprintf(stdout, "Building the RHS of the least squares problem.\n");
                 fflush(stdout);
@@ -324,7 +322,7 @@ void Apply_Helmholtz_Projection_SymTensor(
             //
             //// Now apply the least-squares solver
             //
-            #if DEBUG >= 1
+            #if DEBUG >= 2
             if ( (wRank == 0) and (Itime == 0) ) {
                 fprintf(stdout, "Solving the least squares problem.\n");
                 fflush(stdout);
@@ -356,7 +354,7 @@ void Apply_Helmholtz_Projection_SymTensor(
                 * NMV countains number of matrix-vector calculations
             */
 
-            #if DEBUG >= 1
+            #if DEBUG >= 2
             if ( (wRank == 0) and (Itime == 0) ) {
                 fprintf(stdout, " Done solving the least squares problem.\n");
                 fflush(stdout);
@@ -366,11 +364,10 @@ void Apply_Helmholtz_Projection_SymTensor(
             // Extract the solution and add the seed back in
             double *LHS_ptr = F_alglib.getcontent();
             std::vector<double> F_array( LHS_ptr, LHS_ptr + 3 * Npts);
-            //std::copy ( LHS_ptr, LHS_ptr + 3 * Npts, F_array.begin() );
             for (size_t ii = 0; ii < Npts; ++ii) { F_array.at(ii) += LHS_seed.at(ii); }
 
             // Get velocity associated to computed F field
-            #if DEBUG >= 1
+            #if DEBUG >= 2
             if ( (wRank == 0) and (Itime == 0) ) {
[…]
