### 2023-02-02 a1d6189
Minor comment and printing changes to uiuj projector
