### 2023-02-02 1df853e
Improving land-handling in downsampling routine
