### 2023-02-02 4c90cec
Adding in post-processing to coarse_grain_scalars
As well as functionality for getting PE/KE conversions
