+++ Case_Files/coarse_grain_scalars.cpp
@@ -12,6 +12,7 @@
 #include "../netcdf_io.hpp"
 #include "../functions.hpp"
 #include "../constants.hpp"
+#include "../postprocess.hpp"
 
 int main(int argc, char *argv[]) {
     
@@ -37,6 +38,10 @@ int main(int argc, char *argv[]) {
     static_assert( not( (constants::EXTEND_DOMAIN_TO_POLES) and (constants::CARTESIAN) ),
             "Cartesian implies that there are no poles, so cannot extend to poles."
             "Please update constants.hpp accordingly.");
+
+    // Cannot have OkuboWeiss and postprocessing turned on
+    static_assert( not( (constants::DO_OKUBOWEISS_ANALYSIS) and (constants::APPLY_POSTPROCESS) ),
+           "" );
 
     // Enable all floating point exceptions but FE_INEXACT
     //feenableexcept( FE_ALL_EXCEPT & ~FE_INEXACT & ~FE_UNDERFLOW );
@@ -77,13 +82,28 @@ int main(int argc, char *argv[]) {
 
     const std::string &latlon_in_degrees  = input.getCmdOption("--is_degrees",   "true", asked_help);
 
+    const std::string &compute_PEKE_conv  = input.getCmdOption("--do_PEKE_conversion", "false", asked_help);
+
     const std::string   &Nprocs_in_time_string  = input.getCmdOption("--Nprocs_in_time",  "1", asked_help),
                         &Nprocs_in_depth_string = input.getCmdOption("--Nprocs_in_depth", "1", asked_help);
     const int   Nprocs_in_time_input  = stoi(Nprocs_in_time_string),
                 Nprocs_in_depth_input = stoi(Nprocs_in_depth_string);
 
+    const std::string   &region_defs_fname    = input.getCmdOption("--region_definitions_file",    
+                                                                   "region_definitions.nc", 
+                                                                   asked_help,
+                                                                   "netCDF file containing user-specified region definitions."),
+                        &region_defs_dim_name = input.getCmdOption("--region_definitions_dim",     
+                                                                   "region",                
+                                                                   asked_help,
+                                                                   "Name of the region dimension in the regions file."),
+                        &region_defs_var_name = input.getCmdOption("--region_definitions_var",     
+                                                                   "region_definition",     
+                                                                   asked_help,
+                                                                   "Name of the variable in the regions file that provides the region definitions.");
+
     // Also read in the fields to be filtered from commandline
-    //   e.g. --filter_scales "rho salinity p" (names must match with input netcdf file)
+    //   e.g. --variables "rho salinity p" (names must match with input netcdf file)
     std::vector< std::string > vars_to_filter;
     input.getListofStrings( vars_to_filter, "--variables", asked_help );
     const int Nvars = vars_to_filter.size();
@@ -101,6 +121,8 @@ int main(int argc, char *argv[]) {
 
     // Print some header info, depending on debug level
     print_header_info();
+
+    Timing_Records timing_records;
 
     // Initialize dataset class instance
     dataset source_data;
@@ -129,13 +151,13 @@ int main(int argc, char *argv[]) {
 
     // Compute the area of each 'cell' which will be necessary for integration
     #if DEBUG >= 2
[…]
