### 2023-02-04 husseinaluie/FlowSieve#24 @ashwinvis
It goes a bit further now, but errors in a different file

```cpp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o NETCDF_IO/add_attr_to_file.o NETCDF_IO/add_attr_to_file.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o NETCDF_IO/add_var_to_file.o NETCDF_IO/add_var_to_file.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o NETCDF_IO/check_file_existence.o NETCDF_IO/check_file_existence.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o NETCDF_IO/initialize_output_file.o NETCDF_IO/initialize_output_file.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o NETCDF_IO/initialize_particle_file.o NETCDF_IO/initialize_particle_file.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o NETCDF_IO/initialize_postprocess_file.o NETCDF_IO/initialize_postprocess_file.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o NETCDF_IO/initialize_projected_particle_file.o NETCDF_IO/initialize_projected_particle_file.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o NETCDF_IO/initialize_regions_file.o NETCDF_IO/initialize_regions_file.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o NETCDF_IO/initialize_subset_file.o NETCDF_IO/initialize_subset_file.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o NETCDF_IO/load_region_definitions.o NETCDF_IO/load_region_definitions.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o NETCDF_IO/nc_err.o NETCDF_IO/nc_err.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o NETCDF_IO/package_field.o NETCDF_IO/package_field.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o NETCDF_IO/read_attr_from_file.o NETCDF_IO/read_attr_from_file.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o NETCDF_IO/read_mask_from_file.o NETCDF_IO/read_mask_from_file.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o NETCDF_IO/read_var_from_file.o NETCDF_IO/read_var_from_file.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o NETCDF_IO/write_field_to_output.o NETCDF_IO/write_field_to_output.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o NETCDF_IO/write_integral_to_post.o NETCDF_IO/write_integral_to_post.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o NETCDF_IO/write_regions_to_post.o NETCDF_IO/write_regions_to_post.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o NETCDF_IO/write_time_average_to_post.o NETCDF_IO/write_time_average_to_post.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o Functions/Index.o Functions/Index.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o Functions/Index1to4.o Functions/Index1to4.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o Functions/KE_from_vels.o Functions/KE_from_vels.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o Functions/apply_filter_at_point.o Functions/apply_filter_at_point.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o Functions/apply_filter_at_point_for_quadratics.o Functions/apply_filter_at_point_for_quadratics.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o Functions/compute_Pi.o Functions/compute_Pi.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o Functions/compute_Pi_Helmholtz.o Functions/compute_Pi_Helmholtz.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o Functions/compute_Pi_shift_deriv.o Functions/compute_Pi_shift_deriv.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o Functions/compute_Z.o Functions/compute_Z.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o Functions/compute_areas.o Functions/compute_areas.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
mpic++   -c -DDEBUG=0 -w -std=c++14 -O3    -o Functions/compute_div_transport.o Functions/compute_div_transport.cpp -lnetcdf -lhdf5_hl -lhdf5 -lz -lcurl -fopenmp
In file included from Functions/compute_div_transport.cpp:1:
Functions/../constants.hpp:346:26: error: 'string' is not a member of 'std'
  346 |     const std::map< std::string, std::string > variable_descriptions = {
      |                          ^~~~~~
In file included from Functions/compute_div_transport.cpp:1:
Functions/../constants.hpp:5:1: note: 'std::string' is defined in header '<string>'; did you forget to '#include <string>'?
    4 | #include <map>
  +++ |+#include <string>
    5 |
In file included from Functions/compute_div_transport.cpp:1:
Functions/../constants.hpp:346:26: error: 'string' is not a member of 'std'
  346 |     const std::map< std::string, std::string > variable_descriptions = {
      |                          ^~~~~~
Functions/../constants.hpp:346:26: note: 'std::string' is defined in header '<string>'; did you forget to '#include <string>'?
Functions/../constants.hpp:346:39: error: 'string' is not a member of 'std'
  346 |     const std::map< std::string, std::string > variable_descriptions = {
      |                                       ^~~~~~
Functions/../constants.hpp:346:39: note: 'std::string' is defined in header '<string>'; did you forget to '#include <string>'?
Functions/../constants.hpp:346:39: error: 'string' is not a member of 'std'
Functions/../constants.hpp:346:39: note: 'std::string' is defined in header '<string>'; did you forget to '#include <string>'?
Functions/../constants.hpp:346:46: error: template argument 1 is invalid
  346 |     const std::map< std::string, std::string > variable_descriptions = {
      |                                              ^
Functions/../constants.hpp:346:46: error: template argument 2 is invalid
Functions/../constants.hpp:346:46: error: template argument 3 is invalid
Functions/../constants.hpp:346:46: error: template argument 4 is invalid
Functions/../constants.hpp:346:48: error: scalar object 'constants::variable_descriptions' requires one element in initializer
  346 |     const std::map< std::string, std::string > variable_descriptions = {
      |                                                ^~~~~~~~~~~~~~~~~~~~~
Functions/../constants.hpp:367:26: error: 'string' is not a member of 'std'
  367 |     const std::map< std::string, std::string > variable_units = {
      |                          ^~~~~~
Functions/../constants.hpp:367:26: note: 'std::string' is defined in header '<string>'; did you forget to '#include <string>'?
Functions/../constants.hpp:367:26: error: 'string' is not a member of 'std'
Functions/../constants.hpp:367:26: note: 'std::string' is defined in header '<string>'; did you forget to '#include <string>'?
Functions/../constants.hpp:367:39: error: 'string' is not a member of 'std'
  367 |     const std::map< std::string, std::string > variable_units = {
      |                                       ^~~~~~
Functions/../constants.hpp:367:39: note: 'std::string' is defined in header '<string>'; did you forget to '#include <string>'?
Functions/../constants.hpp:367:39: error: 'string' is not a member of 'std'
Functions/../constants.hpp:367:39: note: 'std::string' is defined in header '<string>'; did you forget to '#include <string>'?
Functions/../constants.hpp:367:46: error: template argument 1 is invalid
  367 |     const std::map< std::string, std::string > variable_units = {
      |                                              ^
Functions/../constants.hpp:367:46: error: template argument 2 is invalid
Functions/../constants.hpp:367:46: error: template argument 3 is invalid
Functions/../constants.hpp:367:46: error: template argument 4 is invalid
Functions/../constants.hpp:367:48: error: scalar object 'constants::variable_units' requires one element in initializer
  367 |     const std::map< std::string, std::string > variable_units = {
      |                                                ^~~~~~~~~~~~~~
Functions/../constants.hpp:389:16: error: 'string' in namespace 'std' does not name a type
  389 |     const std::string spatial_average_description       = "The lat/lon average computed over each defined region (see region dimension).";
      |                ^~~~~~
Functions/../constants.hpp:389:11: note: 'std::string' is defined in header '<string>'; did you forget to '#include <string>'?
  389 |     const std::string spatial_average_description       = "The lat/lon average computed over each defined region (see region dimension).";
      |           ^~~
Functions/../constants.hpp:397:16: error: 'string' in namespace 'std' does not name a type
  397 |     const std::string zonal_average_description         = "The zonal (longitudinal) average computed at each latitude.";
      |                ^~~~~~
Functions/../constants.hpp:397:11: note: 'std::string' is defined in header '<string>'; did you forget to '#include <string>'?
  397 |     const std::string zonal_average_description         = "The zonal (longitudinal) average computed at each latitude.";
      |           ^~~
Functions/../constants.hpp:405:16: error: 'string' in namespace 'std' does not name a type
  405 |     const std::string time_average_description          = "Time average over the entire provided dataset.";
      |                ^~~~~~
Functions/../constants.hpp:405:11: note: 'std::string' is defined in header '<string>'; did you forget to '#include <string>'?
  405 |     const std::string time_average_description          = "Time average over the entire provided dataset.";
      |           ^~~
Functions/../constants.hpp:415:16: error: 'string' in namespace 'std' does not name a type
  415 |     const std::string OkuboWeiss_average_description    = "Variable binned by Okubo-Weiss parameter (i.e. histogram). "
      |                ^~~~~~
Functions/../constants.hpp:415:11: note: 'std::string' is defined in header '<string>'; did you forget to '#include <string>'?
  415 |     const std::string OkuboWeiss_average_description    = "Variable binned by Okubo-Weiss parameter (i.e. histogram). "
      |           ^~~
make: *** [Makefile:78: Functions/compute_div_transport.o] Error 1

````
