### 2023-02-04 husseinaluie/FlowSieve#23 @ashwinvis
Seems like

- curl
- hdf5-mpi (not just hdf5)

is also required to compile. Here is the Nix file I used to compile

```nix
❯ cat default.nix
# Use `nix-shell --pure` to verify if the listed dependencies are sufficient or not
{ pkgs ? import <nixpkgs> {} }:

let
  shell = pkgs.mkShell {
    name = "flowsieve-nix-dev";

    packages = [
      pkgs.mpi
      pkgs.pkg-config
      pkgs.netcdf-mpi
      pkgs.hdf5-mpi
      pkgs.gcc9
      pkgs.curl
    ];

    # cp -f ${builtins.getEnv "PWD"}/Systems/Graham.mk system.mk
    shellHook = ''
      make clean
      OMPI_CXX=$(command -v g++) make -j Case_Files/coarse_grain.x
    '';
  };
in shell
```
