### 2023-02-06 6b264f8
Add library requirements to Installation guide
Re #23
