### 2023-02-06 husseinaluie/FlowSieve#23 @bastorer
I've added a section to the `Install` documentation outlining these library requirement. Thanks for the suggestion!

I'm not familiar with Nix (or other library image tools), but I'll look into them a bit. I tend to shy away from compiling libraries mostly because it seems daunting, but the improved portability would be nice.
