### 2023-02-08 husseinaluie/FlowSieve#24 @ashwinvis
Isn't it `-Wall` and `-Werror`?
