### 2023-02-08 husseinaluie/FlowSieve#23 @ashwinvis
I dont expect you to use Nix, I simply shared it for future reference. It fetches precompiled builds of libraries.and gives an isolated shell. It was handy for me to track down dependencies.
