### 2023-02-08 husseinaluie/FlowSieve#24 @bastorer
Yep! Sorry, typo :-) I have `-Wall` on by default, but only added `-Werror` while I was hunting down the GCC warnings. If you remove `-w`, do you still get warnings?
