### 2023-02-17 a6503cf
Updating coarsen/refine routines to recognize and handle decreasing grids
Previously implicitly assumed that lat/lon grids were increasing
