+++ Case_Files/coarsen_grid_linear.cpp
@@ -133,6 +133,25 @@ int main(int argc, char *argv[]) {
     fine_data.load_latitude(  latitude_dim_name,  fine_fname );
     fine_data.load_longitude( longitude_dim_name, fine_fname );
 
+    const bool  COARSE_LAT_GRID_INCREASING = ( coarse_data.latitude[1]  > coarse_data.latitude[0]  ) ? true : false,
+                COARSE_LON_GRID_INCREASING = ( coarse_data.longitude[1] > coarse_data.longitude[0] ) ? true : false;
+
+    const bool  FINE_LAT_GRID_INCREASING = ( fine_data.latitude[1]  > fine_data.latitude[0]  ) ? true : false,
+                FINE_LON_GRID_INCREASING = ( fine_data.longitude[1] > fine_data.longitude[0] ) ? true : false;
+
+    #if DEBUG >= 0
+    if (wRank == 0) {
+        fprintf( stdout, "Coarse latitude grid is %s and longitude grid is %s\n",
+                COARSE_LAT_GRID_INCREASING ? "increasing" : "decreasing",
+                COARSE_LON_GRID_INCREASING ? "increasing" : "decreasing"
+               );
+        fprintf( stdout, "Fine latitude grid is %s and longitude grid is %s\n",
+                FINE_LAT_GRID_INCREASING ? "increasing" : "decreasing",
+                FINE_LON_GRID_INCREASING ? "increasing" : "decreasing"
+               );
+    }
+    #endif
+
     // Apply some cleaning to the processor allotments if necessary. 
     coarse_data.check_processor_divisions( Nprocs_in_time_input, Nprocs_in_depth_input );
     fine_data.Nprocs_in_time = coarse_data.Nprocs_in_time;
@@ -202,7 +221,8 @@ int main(int argc, char *argv[]) {
         shared( coarse_data, fine_data, var_coarse, mask_coarse, vars_to_refine, stdout ) \
         private( target_lat, target_lon, Itime, Idepth, II_coarse, Ilat_coarse, Ilon_coarse, \
                  RIGHT, LEFT, BOT, TOP, II_fine, Ilat_fine, Ilon_fine, cnt, interp_val, land_cnt ) \
-        firstprivate( Npts_coarse, Nlon_coarse, Nlat_coarse, Nlon_fine, Nlat_fine, Ntime, Ndepth )
+        firstprivate( Npts_coarse, Nlon_coarse, Nlat_coarse, Nlon_fine, Nlat_fine, Ntime, Ndepth, \
+                      COARSE_LAT_GRID_INCREASING, COARSE_LON_GRID_INCREASING, FINE_LAT_GRID_INCREASING, FINE_LON_GRID_INCREASING )
         {
             #pragma omp for collapse(1) schedule(guided)
             for (II_coarse = 0; II_coarse < Npts_coarse; ++II_coarse) {
@@ -215,47 +235,116 @@ int main(int argc, char *argv[]) {
                 //
 
                 // bottom
-                if ( Ilat_coarse == 0 ) {
-                    target_lat = coarse_data.latitude.at(Ilat_coarse);
+                if ( COARSE_LAT_GRID_INCREASING ) {
+                    if ( Ilat_coarse == 0 ) {
+                        target_lat = coarse_data.latitude.at(Ilat_coarse);
+                    } else {
+                        target_lat = 0.5 * ( coarse_data.latitude.at(Ilat_coarse) + coarse_data.latitude.at(Ilat_coarse - 1) );
+                    }
                 } else {
-                    target_lat = 0.5 * ( coarse_data.latitude.at(Ilat_coarse) + coarse_data.latitude.at(Ilat_coarse - 1) );
[…]
