### 2023-02-17 cc0cc8d
Turning PEKE computations on/off with a flag in coarse_grain_scalars
