### 2023-02-20 #4277 @bastorer
Hi @kthyng I don't have access to test / build with gcc10, but it sounds like @ashwinvis was able to compile with gcc9 (I've tested with 8 and 9 on my end, and they worked as expected).

@ashwinvis did you run into any other problems with the software? Until I can find a gcc10 environment to work in, I may just need to add to the minimum requirements ([husseinaluie/FlowSieve#23](https://github.com/husseinaluie/FlowSieve/issues/23)) that gcc10+ are currently not supported.
