### 2023-02-23 #4277 @kthyng
@bastorer and @ashwinvis It does seem like if the code can compiled with gcc8 and gcc9, then gcc10 can be left for future work. @ashwinvis do you agree? What are the next steps in your review?
