### 2023-02-26 husseinaluie/FlowSieve#24 @ashwinvis
An update: I tried to build using conda so I get a full environment based on gcc 9 and there are no warnings, even with `-Wall`. Here is the `environment.yaml` file

```yaml
name: flowsieve-build-env
channels:
 - conda-forge

dependencies:
 - python=3.7
 - numpy=1.17
 - matplotlib=3.*
 - netCDF4=*=mpi*
 - scipy=1.*
 - pip
 - hdf5=*=mpi*
 - openmpi
 - openmpi-mpicxx
 - gcc=9*
 - pip:
   - cmocean
   - pyproj
```
