### 2023-02-26 #4277 @ashwinvis
@kthyng I agree, as long as it is explicitly mentioned in the docs that FlowSieve only supports upto gcc9 (as @bastorer says he will) we can move forward. I noted now in https://github.com/husseinaluie/FlowSieve/issues/24 that it works as expected with gcc9.
