### 2023-02-26 #4277 @ashwinvis
@bastorer I have the following annotations from the manuscript

“careful coarse-graining” ([Storer and Aluie, 2022, p. 1](zotero://select/groups/1699532/items/SS9H8YE9)) ([pdf](zotero://open-pdf/groups/1699532/items/ELS8ID4U?page=1&annotation=MQBYRVNU)) I wonder if this adjective mean anything in this context?

“python user-friendliness” ([Storer and Aluie, 2022, p. 1](zotero://select/groups/1699532/items/SS9H8YE9)) ([pdf](zotero://open-pdf/groups/1699532/items/ELS8ID4U?page=1&annotation=E6CPL4E8)) reword as "user-friendly Python"

“extend this functionality” ([Storer and Aluie, 2022, p. 1](zotero://select/groups/1699532/items/SS9H8YE9)) ([pdf](zotero://open-pdf/groups/1699532/items/ELS8ID4U?page=1&annotation=9KM4NV7H)) to what kind of grids?

“State of the Field” ([Storer and Aluie, 2022, p. 1](zotero://select/groups/1699532/items/SS9H8YE9)) ([pdf](zotero://open-pdf/groups/1699532/items/ELS8ID4U?page=1&annotation=LZFE4VNC)) The state of the field section only mentions what other packages exists. However to the reader it remains unclear what advantages or disadvantages FlowSieve bring in compared to other coarse-graining approaches/

“python” ([Storer and Aluie, 2022, p. 2](zotero://select/groups/1699532/items/SS9H8YE9)) ([pdf](zotero://open-pdf/groups/1699532/items/ELS8ID4U?page=2&annotation=FRN52WKI)) capitalize "Python"

“Another approach applies structure functions to Lagrangian trajectories to extract spectral diagnostics, such as power spectra and inter-scale energy transfers (Frisch (1995), Balwada et al. (2022)).” ([Storer and Aluie, 2022, p. 2](zotero://select/groups/1699532/items/SS9H8YE9)) ([pdf](zotero://open-pdf/groups/1699532/items/ELS8ID4U?page=2&annotation=NDJBHKWB)) How does computation of spectra and transfer functions relate to coarse-graining? To my experience these are statistical quantities. One cannot build a vector field from say a K.E. spectra, only vice-versa is possible. This statement seems out-of-context here, so perhaps move it to another paragraph.

“follows the rigorous underlying mathematical framework to preserve physical properties of the data” ([Storer and Aluie, 2022, p. 2](zotero://select/groups/1699532/items/SS9H8YE9)) ([pdf](zotero://open-pdf/groups/1699532/items/ELS8ID4U?page=2&annotation=ACNBZ57A)) Which mathematical framework? Which physical properties? Please specify and / or cite.
