### 2023-02-28 401a6b5
Update INSTALL.md
Adding note to inform users that FlowSieve currently does not support gcc10+.

Re Issue #24
