+++ Documentation/INSTALL.md
@@ -14,9 +14,10 @@
   * hdf5-mpi required
 * mpi (openmpi or intel mpi, not mpich)
 * compilers: either gcc or icpc (intel compiler)
-  * For gcc, tested on gcc8 and gcc9. When tested, FlowSieve compiled _without_ warnings or errors.
-  * Intended to be compiled with c++14 standard (this is specified in the compiler flags in the `system.mk` file)
+  * For gcc, tested on gcc8 and gcc9. When tested, FlowSieve compiled _without_ warnings or errors. 
+    * Note that FlowSieve currently **does not support for gcc10+**.
   * If Intel compilers are available, they are recommended, since the additional compiler optimizations provide a minor performance improvement. 
+  * Intended to be compiled with c++14 standard (this is specified in the compiler flags in the `system.mk` file)
 * curl
 
 The following, while not required to run FlowSieve itself, are required to run the tutorial. These requirements are outline in `Tutorials/environment.yml` to facility create an anaconda environment. Note that these _exact_ package versions may not be required, but they are sufficient. That is, other package versions may work, but have not been tested.
