### 2023-02-28 husseinaluie/FlowSieve#24 @bastorer
Excellent! Glad to hear that it compiles without warnings :-)

I've updated the documentation to note that gcc10+ is currently not supported.
