### 2023-02-28 #4277 @bastorer
Thanks for these notes! I've updated the paper files and pushed them to the repo :-)

> “careful coarse-graining” (Storer and Aluie, 2022, p. 1) (pdf) I wonder if this adjective mean anything in this context?

"Careful" does have some meaning, since 'direct' filtering of the velocity fields without Helmholtz decomposition [whether with coarse-graining or block averaging (which is also sometimes called coarse-graining)] does not guarantee commutativity with differential operators. In that sense, by including the Helmholtz decomposition step, we are being more careful in our scale decomposition.

> “python user-friendliness” (Storer and Aluie, 2022, p. 1) (pdf) reword as "user-friendly Python"

Text updated.

> “extend this functionality” (Storer and Aluie, 2022, p. 1) (pdf) to what kind of grids?

Text updated to specify "(e.g. to include unstructured grids)"

> “State of the Field” (Storer and Aluie, 2022, p. 1) (pdf) The state of the field section only mentions what other packages exists. However to the reader it remains unclear what advantages or disadvantages FlowSieve bring in compared to other coarse-graining approaches

That's a fair point. The main advantages of FlowSieve are
1) designed to handle spherical geometries [i.e. can filter full global flow fields]
2) computationally / numerically stable at all filter scales, so can coarse-grain at all scales from grid-scale to domain-scale [e.g. [Storer et al., 2022](https://doi.org/10.1038/s41467-022-33031-3) filters from 10s of km up to 40,000km (the equatorial circumference of the Earth)]
3) heavily parallelized to utilize HPC environments
4) on-line post-processing and diagnostic calculations [e.g. across-scale energy transfers].

I will update the text accordingly!

> “python” (Storer and Aluie, 2022, p. 2) (pdf) capitalize "Python"

Text updated.

> “Another approach applies structure functions to Lagrangian trajectories to extract spectral diagnostics, such as power spectra and inter-scale energy transfers (Frisch (1995), Balwada et al. (2022)).” (Storer and Aluie, 2022, p. 2) (pdf) How does computation of spectra and transfer functions relate to coarse-graining? To my experience these are statistical quantities. One cannot build a vector field from say a K.E. spectra, only vice-versa is possible. This statement seems out-of-context here, so perhaps move it to another paragraph.

Computing power spectra and transfer functions are actually our primary applications for coarse-graining. Coarse-graining is applied to the physical-space vector fields to compute the amount of energy contained in scales larger than a given ell. By then performing a 'scan' over a wide range of filter scales ell, we can determine the contribution to the total energy from each scale, and thereby compute the KE spectrum. In [Storer et al., 2022](https://doi.org/10.1038/s41467-022-33031-3) we used this method to compute power spectra on the entire globe. [Aluie et al. 2018](https://doi.org/10.1175/JPO-D-17-0100.1), among others, used coarse-graining to compute the across-scale kinetic energy transfer.

This statement is included to acknowledge other techniques that have been implemented to try and measure similar flow properties (spectra, energy transfers, etc) as coarse-graining.

> “follows the rigorous underlying mathematical framework to preserve physical properties of the data” (Storer and Aluie, 2022, p. 2) (pdf) Which mathematical framework? Which physical properties? Please specify and / or cite.

Because of the methodology, we preserve commutativity with differential operators. For physical properties, that means that a flow that is initially non-divergent will remain non-divergent. Also that computing the vorticity of a filtered field is equivalent to computing the filtered vorticity. Those properties are not guaranteed with direct 'block averaging' or filtering with using Helmholtz decomposition.

For the 'mathematical framework', I'm referring to [Aluie 2019](https://doi.org/10.1007/s13137-019-0123-9) (citation added to text).
