### 2023-03-01 6a41bf8
Re-introduce region std-devs to postprocess outputs
