+++ Postprocess/compute_region_avg_and_std.cpp
@@ -91,64 +91,60 @@ void compute_region_avg_and_std(
         }
     }
 
-    /*
-     *  We've silenced standard deviation outputs for now, so don't waste time computing them
-     *
-     
-    // Now that we have region averages, get region standard deviations
+    //
+    //// Repeat the process on standard deviations
+    //
+
+
     #if DEBUG >= 1
     if (wRank == 0) { fprintf(stdout, "  Computing region standard deviations\n"); }
     #endif
     for (Ifield = 0; Ifield < num_fields; ++Ifield) {
         #if DEBUG >= 2
-        if (wRank == 0) { fprintf(stdout, "    processing field %d of %d standard deviations\n", Ifield + 1, num_fields); }
+        if (wRank == 0) { fprintf(stdout, "    processing field %d of %d std devs\n", Ifield + 1, num_fields); }
         #endif
-        for (Iregion = 0; Iregion < num_regions; ++Iregion) {
-            for (Itime = 0; Itime < Ntime; ++Itime) {
-                for (Idepth = 0; Idepth < Ndepth; ++Idepth) {
 
-                    int_index = Index(0, Itime, Idepth, Iregion, 1, Ntime, Ndepth, num_regions);
-                    reg_area = source_data.region_areas.at(int_index);
+        std::vector<double> field_integrals(Ntime*Ndepth*num_regions,0.);
 
-                    int_val = 0.;
-
-                    #pragma omp parallel default(none)\
-                    private(Ilat, Ilon, index, dA, area_index )\
-                    shared( source_data, Ifield, Iregion, Itime, Idepth, int_index,\
-                            postprocess_fields, field_averages) \
-                    reduction(+ : int_val)
-                    { 
-                        #pragma omp for collapse(2) schedule(dynamic)
+        #pragma omp parallel default(none)\
+        private(Ilat, Ilon, index, dA, area_index, increment, int_index, \
+                Idepth, Itime, Iregion )\
+        shared( source_data, Ifield, postprocess_fields, field_averages ) \
+        firstprivate( Nlon, Nlat, Ndepth, Ntime, num_regions ) \
+        reduction(vec_double_plus : field_integrals)
+        { 
+            #pragma omp for collapse(5) schedule(static)
+            for (Iregion = 0; Iregion < num_regions; ++Iregion) {
+                for (Itime = 0; Itime < Ntime; ++Itime) {
+                    for (Idepth = 0; Idepth < Ndepth; ++Idepth) {
                         for (Ilat = 0; Ilat < Nlat; ++Ilat) {
                             for (Ilon = 0; Ilon < Nlon; ++Ilon) {
-
+                                int_index = Index(0, Itime, Idepth, Iregion, 1, Ntime, Ndepth, num_regions);
                                 index = Index(Itime, Idepth, Ilat, Ilon, Ntime, Ndepth, Nlat, Nlon);
 
-                                if ( source_data.mask.at(index) ) { // Skip land areas
+                                if ( constants::FILTER_OVER_LAND or source_data.mask.at(index) ) { // Skip land areas
 
                                     area_index = Index(0, 0, Ilat, Ilon, 1, 1, Nlat, Nlon);
-
                                     dA = source_data.areas.at(area_index);
+                                    increment = pow(     field_averages.at(Ifield).at(int_index)
+                                                       - postprocess_fields.at(Ifield)->at(index),
+                                                    2.) * dA;
 
                                     if ( source_data.regions.at( source_data.region_names.at(Iregion) ).at(area_index) )
                                     {
-                                        int_val +=
-                                            pow(   field_averages.at(Ifield).at(int_index)
-                                                 - postprocess_fields.at(Ifield)->at(index),
-                                                2.) * dA;
[…]
