### 2023-03-01 23a5a97
Compute areas now handles decreasing grids Previously introduced a sign error
