+++ Case_Files/LLC_interpolator_simple.cpp
@@ -1,0 +1,349 @@
+#include <fenv.h>
+#include <stdio.h>
+#include <string.h>
+#include <stdlib.h>
+#include <algorithm>
+#include <math.h>
+#include <vector>
+#include <mpi.h>
+#include <omp.h>
+#include <cassert>
+
+#include "../ALGLIB/stdafx.h"
+#include "../ALGLIB/interpolation.h"
+
+#include "../netcdf_io.hpp"
+#include "../functions.hpp"
+#include "../constants.hpp"
+#include "../preprocess.hpp"
+
+
+// Snyder, J. P. (1987). Map Projections: A Working Manual. https://doi.org/10.3133/pp1395
+// page 173
+void near_sided_project( 
+       double & proj_x,
+       double & proj_y,
+       const double & pt_lon,
+       const double & pt_lat,
+       const double & centre_lon,
+       const double & centre_lat,
+       const double & viewing_elev
+        ) {
+
+    const double    cos_phi = cos( pt_lat ),
+                    sin_phi = sin( pt_lat ),
+                    cos_phi0 = cos( centre_lat ),
+                    sin_phi0 = sin( centre_lat ),
+                    sin_lam = sin( pt_lon - centre_lon ),
+                    cos_lam = cos( pt_lon - centre_lon );
+
+    const double    P = 1 + viewing_elev / constants::R_earth;
+
+    const double    cos_c = sin_phi0 * sin_phi + cos_phi0 * cos_phi * cos_lam;
+
+    if (cos_c < 1/P ) {
+        proj_x = 1e100;
+        proj_y = 1e100;
+        return;
+    }
+    
+    const double k_prime = ( P - 1 ) / ( P - cos_c );
+
+    proj_x = constants::R_earth * k_prime * cos_phi * sin_lam;
+    proj_y = constants::R_earth * k_prime * ( cos_phi0 * sin_phi - sin_phi0 * cos_phi * cos_lam );
+
+};
+
+
+int main(int argc, char *argv[]) {
+    
+    // PERIODIC_Y implies UNIFORM_LAT_GRID
+    static_assert( (constants::UNIFORM_LAT_GRID) or (not(constants::PERIODIC_Y)),
+            "PERIODIC_Y requires UNIFORM_LAT_GRID.\n"
+            "Please update constants.hpp accordingly.\n");
+    static_assert( not(constants::CARTESIAN),
+            "Toroidal projection now set to handle Cartesian coordinates.\n"
+            );
+
+    // Specify the number of OpenMP threads
+    //   and initialize the MPI world
+    int thread_safety_provided;
+    MPI_Init_thread(&argc, &argv, MPI_THREAD_MULTIPLE, &thread_safety_provided);
+    //MPI_Comm_set_errhandler(MPI_COMM_WORLD, MPI::ERRORS_THROW_EXCEPTIONS);
+
+    int wRank=-1, wSize=-1;
+    MPI_Comm_rank( MPI_COMM_WORLD, &wRank );
+    MPI_Comm_size( MPI_COMM_WORLD, &wSize );
+
+    //
+    //// Parse command-line arguments
+    //
+    InputParser input(argc, argv);
+    if(input.cmdOptionExists("--version")){
+        if (wRank == 0) { print_compile_info(NULL); } 
+        return 0;
+    }
+
+    // first argument is the flag, second argument is default value (for when flag is not present)
+    const std::string   &input_fname    = input.getCmdOption("--input_file",    "input.nc"),
+                        &grid_fname     = input.getCmdOption("--grid_file",     "grid.nc"),
+                        &output_fname   = input.getCmdOption("--output_file",   "output.nc");
+
+    const std::string   &time_dim_name      = input.getCmdOption("--time",        "time"),
[…]
