+++ NETCDF_IO/read_LLC_latlon_from_file.cpp
@@ -1,0 +1,215 @@
+
+#include "../netcdf_io.hpp"
+#include "../constants.hpp"
+#include "../functions.hpp"
+#include <cassert>
+#include <math.h>
+
+/*!
+ *  \brief Read a specific variable from a specific file.
+ *
+ *  Accounts for variable attributes 'scale_factor' and 'add_offset'.
+ *
+ *  If mask != NULL, then determine the mask based on variable attribute '_FillValue'
+ *
+ *  @param[in,out]  var                 vector into which to store the loaded variable
+ *  @param[in]      var_name            name of the variable to be read
+ *  @param[in]      filename            name of the file from which to load the variable
+ *  @param[in]      comm                the MPI communicator world
+ *
+ */
+
+void read_LLC_latlon_from_file(
+        std::vector<double> &var,
+        const std::string & var_name,
+        const std::string & filename,
+        const MPI_Comm comm
+        ) {
+
+    assert( check_file_existence( filename.c_str() ) );
+
+    int wRank, wSize, Nprocs_in_dim, Iproc_in_dim;
+    MPI_Comm_rank( comm, &wRank );
+    MPI_Comm_size( comm, &wSize );
+
+    // Open the NETCDF file
+    const int str_len = 100;
+    int FLAG = NC_NETCDF4 | NC_MPIIO;
+    int ncid=0, retval;
+    char buffer [str_len];
+    snprintf(buffer, str_len, filename.c_str());
+
+    #if DEBUG >= 1
+    if (wRank == 0) {
+        fprintf(stdout, "Attempting to read %s from %s\n", var_name.c_str(), buffer);
+        fflush(stdout);
+    }
+    #endif
+
+    retval = nc_open_par(buffer, FLAG, comm, MPI_INFO_NULL, &ncid);
+    if (retval != NC_NOERR ) { NC_ERR(retval, __LINE__, __FILE__); }
+
+    // Check if netcdf-4 format
+    int input_nc_format;
+    retval = nc_inq_format( ncid, &input_nc_format );
+    if (retval != NC_NOERR ) { NC_ERR(retval, __LINE__, __FILE__); }
+    #if DEBUG >= 1
+    // NC_FORMAT_CLASSIC, NC_FORMAT_64BIT_OFFSET, NC_FORMAT_CDF5, NC_FORMAT_NETCDF4, NC_FORMAT_NETCDF4_CLASSIC
+    if (wRank == 0) {
+        fprintf( stdout, "Detected file format (%d) is ", input_nc_format );
+        if ( input_nc_format == NC_FORMAT_CLASSIC )         { fprintf( stdout, "NC_FORMAT_CLASSIC (%d)", NC_FORMAT_CLASSIC ); }
+        if ( input_nc_format == NC_FORMAT_64BIT_OFFSET )    { fprintf( stdout, "NC_FORMAT_64BIT_OFFSET (%d)", NC_FORMAT_64BIT_OFFSET ); }
+        if ( input_nc_format == NC_FORMAT_CDF5 )            { fprintf( stdout, "NC_FORMAT_CDF5 (%d)", NC_FORMAT_CDF5 ); }
+        if ( input_nc_format == NC_FORMAT_NETCDF4 )         { fprintf( stdout, "NC_FORMAT_NETCDF4 (%d)", NC_FORMAT_NETCDF4 ); }
+        if ( input_nc_format == NC_FORMAT_NETCDF4_CLASSIC ) { fprintf( stdout, "NC_FORMAT_NETCDF4_CLASSIC (%d)", NC_FORMAT_NETCDF4_CLASSIC ); }
+        fprintf( stdout, " \n" );
+        fflush(stdout);
+    }
+    #endif
[…]
