### 2023-03-03 eee951f
Logging an earlier version of 'build-adjacency' before updating
