+++ Functions/build_adjacency.cpp
@@ -1,0 +1,331 @@
+#include <fenv.h>
+#include <stdio.h>
+#include <string.h>
+#include <stdlib.h>
+#include <algorithm>
+#include <math.h>
+#include <vector>
+#include <mpi.h>
+#include <omp.h>
+#include <cassert>
+
+#include "../functions.hpp"
+#include "../constants.hpp"
+
+// Method of the dataset class
+// i.e. source_data.build_adjacency();
+void dataset::build_adjacency(
+        const MPI_Comm comm
+        ) {
+
+    #if DEBUG >= 1
+    int wRank, wSize;
+    MPI_Comm_rank( comm, &wRank );
+    MPI_Comm_size( comm, &wSize );
+
+    if (wRank == 0) {
+        fprintf( stdout, "Building adjacency matrix.\n" );
+        fflush(stdout);
+    }
+    #endif
+
+    size_t pt_index, search_index;
+    const size_t num_pts = longitude.size();
+
+    adjacency_indices.resize( num_pts );
+    adjacency_projected_x.resize( num_pts );
+    adjacency_projected_y.resize( num_pts );
+    adjacency_distances.resize( num_pts );
+
+    adjacency_ddlon_weights.resize( num_pts );
+    adjacency_ddlat_weights.resize( num_pts );
+
+    // Be a litlle sneaky. We want to make this search faster if possible.
+    // To do that, it would be helpful to have a coarse estimate of how far apart points
+    // are. 
+    // Note: surface area of a sphere is 4 * pi * R^2
+    //       surface area / number of points ~ area per grid point
+    //       sqrt( area per grid point ) ~ side length of grid, i.e. spacing between points
+    // Is this perfect? Nope. But it'll get us in the ballpark
+    const double    typical_spacing = sqrt( 4 * M_PI / num_pts ) * constants::R_earth,
+                    max_distance = 10 * typical_spacing;
+
+    double UR_x, UR_y, UR_dist, UL_x, UL_y, UL_dist, LR_x, LR_y, LR_dist, LL_x, LL_y, LL_dist;
+    size_t UR_ind, UL_ind, LR_ind, LL_ind;
+
+    double pt_lon, pt_lat, search_lon, search_lat, poleward_lat, del_lon_lim,
+           proj_x, proj_y, local_dist, denom, stencil_count;
+    bool near_pole;
+    int Istencil, II, JJ;
+
+    std::vector< double > xi, yi, ddx_coeffs, ddy_coeffs;
+
+    #pragma omp parallel \
+    default(none) \
+    shared( longitude, latitude, adjacency_indices, adjacency_projected_x, adjacency_projected_y, \
+            adjacency_distances, adjacency_ddlon_weights, adjacency_ddlat_weights ) \
+    private( pt_index, search_index, pt_lon, pt_lat, search_lon, search_lat, \
+             proj_x, proj_y, near_pole, poleward_lat, del_lon_lim, local_dist, \
+             LL_ind, LL_x, LL_y, LL_dist, LR_ind, LR_x, LR_y, LR_dist, \
+             UL_ind, UL_x, UL_y, UL_dist, UR_ind, UR_x, UR_y, UR_dist, \
+             Istencil, II, JJ, xi, yi, denom, stencil_count, ddx_coeffs, ddy_coeffs ) \
+    firstprivate( num_pts, max_distance )
+    {
+        #pragma omp for collapse(1) schedule(guided)
+        for ( pt_index = 0; pt_index < num_pts; pt_index++ ) {
+
+            search_index = 0;
+
+            UR_dist = 1e100;
+            UL_dist = 1e100;
+            LR_dist = 1e100;
+            LL_dist = 1e100;
+
+            #if DEBUG > 0
[…]
