### 2023-03-06 e649d72
Storing next version of build_adjacency
This one uses all 126 4-stencils from the 8 nearest neighbours + centre
point to avoid ill-conditioned systems
