+++ Functions/build_adjacency.cpp
@@ -8,6 +8,7 @@
 #include <mpi.h>
 #include <omp.h>
 #include <cassert>
+#include <bitset>
 
 #include "../functions.hpp"
 #include "../constants.hpp"
@@ -29,8 +30,9 @@ void dataset::build_adjacency(
     }
     #endif
 
-    size_t pt_index, search_index;
-    const size_t num_pts = longitude.size();
+    size_t pt_index, search_index, Ineighbour;
+    const size_t num_pts = longitude.size(),
+                 num_neighbours = 8;
 
     adjacency_indices.resize( num_pts );
     adjacency_projected_x.resize( num_pts );
@@ -50,13 +52,15 @@ void dataset::build_adjacency(
     const double    typical_spacing = sqrt( 4 * M_PI / num_pts ) * constants::R_earth,
                     max_distance = 10 * typical_spacing;
 
-    double UR_x, UR_y, UR_dist, UL_x, UL_y, UL_dist, LR_x, LR_y, LR_dist, LL_x, LL_y, LL_dist;
-    size_t UR_ind, UL_ind, LR_ind, LL_ind;
+    std::vector<double> neighbour_x, neighbour_y, neighbour_dist;
+    std::vector<size_t> neighbour_ind;
+    std::bitset<16> stencil_set;
 
     double pt_lon, pt_lat, search_lon, search_lat, poleward_lat, del_lon_lim,
-           proj_x, proj_y, local_dist, denom, stencil_count;
+           proj_x, proj_y, proj_theta, local_dist, denom, denom_sign, stencil_count,
+           stencil_min_x, stencil_min_y, stencil_max_x, stencil_max_y;
     bool near_pole;
-    int Istencil, II, JJ;
+    int Istencil, II, JJ, angle_ind;
 
     std::vector< double > xi, yi, ddx_coeffs, ddy_coeffs;
 
@@ -64,22 +68,29 @@ void dataset::build_adjacency(
     default(none) \
     shared( longitude, latitude, adjacency_indices, adjacency_projected_x, adjacency_projected_y, \
             adjacency_distances, adjacency_ddlon_weights, adjacency_ddlat_weights ) \
-    private( pt_index, search_index, pt_lon, pt_lat, search_lon, search_lat, \
-             proj_x, proj_y, near_pole, poleward_lat, del_lon_lim, local_dist, \
-             LL_ind, LL_x, LL_y, LL_dist, LR_ind, LR_x, LR_y, LR_dist, \
-             UL_ind, UL_x, UL_y, UL_dist, UR_ind, UR_x, UR_y, UR_dist, \
-             Istencil, II, JJ, xi, yi, denom, stencil_count, ddx_coeffs, ddy_coeffs ) \
-    firstprivate( num_pts, max_distance )
+    private( pt_index, search_index, pt_lon, pt_lat, search_lon, search_lat, angle_ind, \
+             proj_x, proj_y, proj_theta, near_pole, poleward_lat, del_lon_lim, local_dist, \
+             Ineighbour, neighbour_x, neighbour_y, neighbour_dist, neighbour_ind, stencil_set, \
+             Istencil, II, JJ, xi, yi, denom, denom_sign, stencil_count, ddx_coeffs, ddy_coeffs, \
+             stencil_min_x, stencil_min_y, stencil_max_x, stencil_max_y ) \
+    firstprivate( num_pts, max_distance, stdout )
     {
         #pragma omp for collapse(1) schedule(guided)
         for ( pt_index = 0; pt_index < num_pts; pt_index++ ) {
 
+            //fprintf( stdout, "Starting point %'zu\n", pt_index );
+
             search_index = 0;
 
-            UR_dist = 1e100;
-            UL_dist = 1e100;
[…]
