### 2023-03-11 #4277 @ashwinvis
Thanks @bastorer it is much clearer. Some minor comments regarding the language. I could not overlook it since it was so obvious and they (JOSS) always put the name of reviewer on the article :)

---

_“The unique contributions of FlowSieve to the field are:” ([Storer and Aluie, p. 2](zotero://select/groups/1699532/items/V94P4DI2)) ([pdf](zotero://open-pdf/groups/1699532/items/X9ZX2ANX?page=2&annotation=QJT9PB2H))_

Now it is clear how FlowSieve sets itself apart from other solutions. Please fix the grammatical structure of this statement though. Right now it conveys "to the field are + Designed for" (also make it lowercase **d**) and  "to the field are + can apply any", which sounds wrong. Also replace periods between each items with a comma.

_“cirfumerence” ([Storer and Aluie, p. 2](zotero://select/groups/1699532/items/V94P4DI2)) ([pdf](zotero://open-pdf/groups/1699532/items/X9ZX2ANX?page=2&annotation=4AGZNFQE))_

 typo

_“to 51 preserve physical properties of the data (e.g. non-divergence of flow) and to accurately filter 52 flows that not divergence-free.” ([Storer and Aluie, p. 2](zotero://select/groups/1699532/items/V94P4DI2)) ([pdf](zotero://open-pdf/groups/1699532/items/X9ZX2ANX?page=2&annotation=967U92ZB))_

Thanks for clarifying, but repeats "non-divergence"
