### 2023-03-13 husseinaluie/FlowSieve#23 @ashwinvis
I see that it updated in https://flowsieve.readthedocs.io/en/latest/install1.html, so this issue can be closed.
