### 2023-03-13 husseinaluie/FlowSieve#29 @ashwinvis
Example of usage with MPI parallelism
While in the article you state that

> In particular, MPI is used to divide time and depth

I did not find an example which demonstrates this in the Tutorials. We only see OpenMP being used and `SLURM_NTASKS` is always set as 1. Would it be possible to construct a simple example which shows MPI parallelism. This is necessary to check out from https://github.com/openjournals/joss-reviews/issues/4277#issuecomment-1383020819:

> Functionality: Have the functional claims of the software been confirmed?
