### 2023-03-14 #4277 @bastorer
> Thanks @bastorer it is much clearer. Some minor comments regarding the language. I could not overlook it since it was so obvious and they (JOSS) always put the name of reviewer on the article :)

Absolutely! I'm always happy to hear suggestions on improved wording :-)

I've updated the text in the most recent commit (fbdc32bb), but copied the update text below as well.

> The unique contributions of `FlowSieve` to the field are:
> 1) analysis on full spherical geometries, allowing the processing of global data,
> 2) arbitrary filtering scales spanning from sub-grid to domain-size (e.g. > @Storer2022 extracts global power spectra for scales spanning 10s of km to 40000 km - the equatorial circumference of the Earth),
> 3) on-line diagnostic calculations [e.g. across-scale energy transfers, large-scale vorticity and divergence],
> 4) on-line post-processing to reduce output file sizes [e.g. averages over user-specified regions, zonal averages],
> 5) rigorous underlying mathematical framework of @Aluie2019 to preserve physical properties of the data (e.g. non-divergence of flow) to accurately filter realistic flows and ensure commutativity between filtering and differential operators, and
> 6) heavy parallelization to utilize HPC environments efficiently.
