### 2023-03-14 #4277 @bastorer
> it occured to me that it is unclear, both from the article and the documentation, what the core functionality is. In other words which is the public API for others? Is it a C++ library or a selection of command line utilities that you have put together under `Case_Files`? If it is the latter, as it appears from the tutorials, it should be documented

I'm not entirely sure that I understand the difference. The use-case that I imagine is that someone would install FlowSieve, compile the appropriate executables (e.g. `Helmholtz_projection.x` and `coarse_grain_Helmholtz.x` from `Case_Files`), and run those executables on whatever dataset of interest they have.

Depending on the diagnostics that they want to study, they would possible add those routines to `Functions` [although admittedly that would require a fair bit of understanding of the code structure, and so this might more realistically involve the user reaching out to us to help them set up their new diagnostics].

So, it sounds like
> selection of command line utilities that you have put together under `Case_Files`

is the most accurate description, I guess I'm just not sure how that should be reflected in the documentation. Could you clarify / advise?

Thanks for all of your help!
