### 2023-03-14 07e3746
Logging in updates in LLC development
Helmholtz projection seems to produce Psi and Phi reasonably well
but the associated velocities are complete garbage

Also, subsampled grids from LLC grid gives very noisy Psi and Phi

First priority is to figure out why the velocities are nonsense
