+++ Case_Files/LLC_Helmholtz_projections.cpp
@@ -1,0 +1,241 @@
+#include <fenv.h>
+#include <stdio.h>
+#include <string.h>
+#include <stdlib.h>
+#include <algorithm>
+#include <math.h>
+#include <vector>
+#include <mpi.h>
+#include <omp.h>
+#include <cassert>
+
+#include "../netcdf_io.hpp"
+#include "../functions.hpp"
+#include "../constants.hpp"
+#include "../preprocess.hpp"
+
+int main(int argc, char *argv[]) {
+    
+    // PERIODIC_Y implies UNIFORM_LAT_GRID
+    static_assert( (constants::UNIFORM_LAT_GRID) or (not(constants::PERIODIC_Y)),
+            "PERIODIC_Y requires UNIFORM_LAT_GRID.\n"
+            "Please update constants.hpp accordingly.\n");
+
+    // Currently cannot be Cartesian
+    static_assert( not(constants::CARTESIAN),
+            "Toroidal projection not set to handle Cartesian coordinates.\n"
+            );
+
+    // Enable all floating point exceptions but FE_INEXACT
+    //feenableexcept(FE_ALL_EXCEPT & ~FE_INEXACT);
+
+    // Specify the number of OpenMP threads
+    //   and initialize the MPI world
+    int thread_safety_provided;
+    MPI_Init_thread(&argc, &argv, MPI_THREAD_MULTIPLE, &thread_safety_provided);
+    //MPI_Comm_set_errhandler(MPI_COMM_WORLD, MPI::ERRORS_THROW_EXCEPTIONS);
+
+    int wRank=-1, wSize=-1;
+    MPI_Comm_rank( MPI_COMM_WORLD, &wRank );
+    MPI_Comm_size( MPI_COMM_WORLD, &wSize );
+
+    //
+    //// Parse command-line arguments
+    //
+    InputParser input(argc, argv);
+    if(input.cmdOptionExists("--version")){
+        if (wRank == 0) { print_compile_info(NULL); } 
+        return 0;
+    }
+    const bool asked_help = input.cmdOptionExists("--help");
+    if (asked_help) {
+        fprintf( stdout, "\033[1;4mThe command-line input arguments [and default values] are:\033[0m\n" );
+    }
+
+    // first argument is the flag, second argument is default value (for when flag is not present)
+    const std::string   &input_fname      = input.getCmdOption("--input_file",      
+                                                               "input.nc",                 
+                                                               asked_help,
+                                                               "netCDF file containing vector field to Helmholtz decompose"),
+                        &output_fname     = input.getCmdOption("--output_file",     
+                                                               "projection_Helmholtz.nc",  
+                                                               asked_help,
+                                                               "Filename for the output (netCDF file of Helmholtz scalars)"),
+                        &seed_fname       = input.getCmdOption("--seed_file",       
+                                                               "zero",                  
+                                                               asked_help,
+                                                               "netCDF file containing initial guesses for Helmholtz scalars.\nUse 'zero' (the default) for no seed.");
+
+    const std::string   &time_dim_name      = input.getCmdOption("--time",        
+                                                                 "time",       
+                                                                 asked_help,
+                                                                 "Name of 'time' dimension in netCDF input file."),
+                        &depth_dim_name     = input.getCmdOption("--depth",       
+                                                                 "depth",      
+                                                                 asked_help,
+                                                                 "Name of 'depth' dimension in netCDF input file."),
+                        &latitude_dim_name  = input.getCmdOption("--latitude",    
+                                                                 "latitude",   
+                                                                 asked_help,
+                                                                 "Name of 'latitude' dimension in netCDF input file."),
+                        &longitude_dim_name = input.getCmdOption("--longitude",   
[…]
