+++ Case_Files/LLC_interpolator_simple.cpp
@@ -9,51 +9,10 @@
 #include <omp.h>
 #include <cassert>
 
-#include "../ALGLIB/stdafx.h"
-#include "../ALGLIB/interpolation.h"
-
 #include "../netcdf_io.hpp"
 #include "../functions.hpp"
 #include "../constants.hpp"
 #include "../preprocess.hpp"
-
-
-// Snyder, J. P. (1987). Map Projections: A Working Manual. https://doi.org/10.3133/pp1395
-// page 173
-void near_sided_project( 
-       double & proj_x,
-       double & proj_y,
-       const double & pt_lon,
-       const double & pt_lat,
-       const double & centre_lon,
-       const double & centre_lat,
-       const double & viewing_elev
-        ) {
-
-    const double    cos_phi = cos( pt_lat ),
-                    sin_phi = sin( pt_lat ),
-                    cos_phi0 = cos( centre_lat ),
-                    sin_phi0 = sin( centre_lat ),
-                    sin_lam = sin( pt_lon - centre_lon ),
-                    cos_lam = cos( pt_lon - centre_lon );
-
-    const double    P = 1 + viewing_elev / constants::R_earth;
-
-    const double    cos_c = sin_phi0 * sin_phi + cos_phi0 * cos_phi * cos_lam;
-
-    if (cos_c < 1/P ) {
-        proj_x = 1e100;
-        proj_y = 1e100;
-        return;
-    }
-    
-    const double k_prime = ( P - 1 ) / ( P - cos_c );
-
-    proj_x = constants::R_earth * k_prime * cos_phi * sin_lam;
-    proj_y = constants::R_earth * k_prime * ( cos_phi0 * sin_phi - sin_phi0 * cos_phi * cos_lam );
-
-};
-
 
 int main(int argc, char *argv[]) {
     
@@ -184,10 +143,6 @@ int main(int argc, char *argv[]) {
         bool near_pole;
         int Ngrid_spline;
 
-        alglib::real_2d_array xyf;
-        alglib::spline2dbuilder builder;
-        alglib::spline2dinterpolant spline;
-        alglib::spline2dfitreport rep;
         double lambdav = 0.000;
 
         size_t LL_ind, LR_ind, UL_ind, UR_ind;
@@ -255,7 +210,7 @@ int main(int argc, char *argv[]) {
                                     );
                             local_dist = sqrt( proj_x*proj_x + proj_y*proj_y );
 
-                            if ( (proj_x > 0) and (proj_y > 0) and ( local_dist < UR_dist ) ) {
+                            if ( (proj_x >= 0) and (proj_y > 0) and ( local_dist < UR_dist ) ) {
                                 UR_x = proj_x;
                                 UR_y = proj_y;
                                 UR_ind = in_index;
@@ -265,7 +220,7 @@ int main(int argc, char *argv[]) {
                                 #else
                                 UR_f = orig_data.variables["to_interp"][ in_index % Nlatlon_in ];
                                 #endif
-                            } else if ( (proj_x > 0) and (proj_y < 0) and ( local_dist < UL_dist ) ) {
+                            } else if ( (proj_x > 0) and (proj_y <= 0) and ( local_dist < UL_dist ) ) {
                                 UL_x = proj_x;
                                 UL_y = proj_y;
                                 UL_ind = in_index;
@@ -275,7 +230,7 @@ int main(int argc, char *argv[]) {
                                 #else
                                 UL_f = orig_data.variables["to_interp"][ in_index % Nlatlon_in ];
                                 #endif
-                            } else if ( (proj_x < 0) and (proj_y > 0) and ( local_dist < LR_dist ) ) {
+                            } else if ( (proj_x < 0) and (proj_y >= 0) and ( local_dist < LR_dist ) ) {
                                 LR_x = proj_x;
[…]
