### 2023-03-15 8c625ef
Adding in functionality to load in adjacency info
Still a work in progress
