+++ Case_Files/LLC_build_adjacency.cpp
@@ -1,0 +1,109 @@
+#include <fenv.h>
+#include <stdio.h>
+#include <string.h>
+#include <stdlib.h>
+#include <algorithm>
+#include <math.h>
+#include <vector>
+#include <mpi.h>
+#include <omp.h>
+#include <cassert>
+
+//#include "../ALGLIB/stdafx.h"
+//#include "../ALGLIB/interpolation.h"
+
+#include "../netcdf_io.hpp"
+#include "../functions.hpp"
+#include "../constants.hpp"
+#include "../preprocess.hpp"
+#include "../differentiation_tools.hpp"
+
+// Pragma-magic to allow reduction over vector operator
+//      thanks to: https://stackoverflow.com/questions/43168661/openmp-and-reduction-on-stdvector
+#pragma omp declare reduction(vec_double_plus : std::vector<double> : \
+                              std::transform(omp_out.begin(), omp_out.end(), omp_in.begin(), omp_out.begin(), std::plus<double>())) \
+                    initializer(omp_priv = decltype(omp_orig)(omp_orig.size()))
+
+
+int main(int argc, char *argv[]) {
+    
+    // PERIODIC_Y implies UNIFORM_LAT_GRID
+    static_assert( (constants::UNIFORM_LAT_GRID) or (not(constants::PERIODIC_Y)),
+            "PERIODIC_Y requires UNIFORM_LAT_GRID.\n"
+            "Please update constants.hpp accordingly.\n");
+    static_assert( not(constants::CARTESIAN),
+            "Toroidal projection now set to handle Cartesian coordinates.\n"
+            );
+
+    // Specify the number of OpenMP threads
+    //   and initialize the MPI world
+    int thread_safety_provided;
+    MPI_Init_thread(&argc, &argv, MPI_THREAD_MULTIPLE, &thread_safety_provided);
+    //MPI_Comm_set_errhandler(MPI_COMM_WORLD, MPI::ERRORS_THROW_EXCEPTIONS);
+
+    int wRank=-1, wSize=-1;
+    MPI_Comm_rank( MPI_COMM_WORLD, &wRank );
+    MPI_Comm_size( MPI_COMM_WORLD, &wSize );
+
+    //
+    //// Parse command-line arguments
+    //
+    InputParser input(argc, argv);
+    if(input.cmdOptionExists("--version")){
+        if (wRank == 0) { print_compile_info(NULL); } 
+        return 0;
+    }
+
+    // first argument is the flag, second argument is default value (for when flag is not present)
+    const std::string   &input_fname    = input.getCmdOption("--input_file",    "input.nc"),
+                        &output_fname   = input.getCmdOption("--output_file",   "adjacency.nc");
+
+    const std::string   &time_dim_name      = input.getCmdOption("--time",        "time"),
+                        &depth_dim_name     = input.getCmdOption("--depth",       "depth"),
+                        &latitude_dim_name  = input.getCmdOption("--latitude",    "latitude"),
+                        &longitude_dim_name = input.getCmdOption("--longitude",   "longitude");
+
+    const std::string &latlon_in_degrees  = input.getCmdOption("--is_degrees",   "true");
+
+    // Print processor assignments
+    const int max_threads = omp_get_max_threads();
+    omp_set_num_threads( max_threads );
+
+    // Print some header info, depending on debug level
+    print_header_info();
+
+    // Initialize dataset class instance
+    dataset source_data;
+
+    // Read in source data / get size information
+    #if DEBUG >= 1
+    if (wRank == 0) { fprintf(stdout, "Reading in source data.\n\n"); }
+    #endif
+
+    // Read in the grid coordinates
[…]
