+++ Functions/dataset_class.cpp
@@ -366,6 +366,8 @@ void dataset::write_adjacency(
     vars_to_write.push_back("adjacency_distances");
     vars_to_write.push_back("adjacency_ddlon_weights");
     vars_to_write.push_back("adjacency_ddlat_weights");
+    vars_to_write.push_back("adjacency_d2dlon2_weights");
+    vars_to_write.push_back("adjacency_d2dlat2_weights");
 
     initialize_adjacency_file( *this, vars_to_write, filename.c_str() );
 
@@ -376,28 +378,36 @@ void dataset::write_adjacency(
 
     size_t starts[2] = { 0,   0  };
     size_t counts[2] = { Npts, Nneighbours };
+
+    // Indices
     var_to_write.resize( Npts*Nneighbours );
     for ( size_t II = 0; II < Npts; II++ ) { for ( size_t JJ = 0; JJ < Nneighbours; JJ++ ) {
             var_to_write.at( Nneighbours*II + JJ ) = (double) adjacency_indices[II][JJ];
     } }
     write_field_to_output( var_to_write, "adjacency_indices", starts, counts, filename.c_str() );
 
+    // Projected x
     for ( size_t II = 0; II < Npts; II++ ) { for ( size_t JJ = 0; JJ < Nneighbours; JJ++ ) {
             var_to_write.at( Nneighbours*II + JJ ) = adjacency_projected_x[II][JJ];
     } }
     write_field_to_output( var_to_write, "adjacency_proj_x", starts, counts, filename.c_str() );
 
+    // Projected y
     for ( size_t II = 0; II < Npts; II++ ) { for ( size_t JJ = 0; JJ < Nneighbours; JJ++ ) {
             var_to_write.at( Nneighbours*II + JJ ) = adjacency_projected_y[II][JJ];
     } }
     write_field_to_output( var_to_write, "adjacency_proj_y", starts, counts, filename.c_str() );
 
+    // Distances
     for ( size_t II = 0; II < Npts; II++ ) { for ( size_t JJ = 0; JJ < Nneighbours; JJ++ ) {
             var_to_write.at( Nneighbours*II + JJ ) = adjacency_distances[II][JJ];
     } }
     write_field_to_output( var_to_write, "adjacency_distances", starts, counts, filename.c_str() );
 
+
     counts[1] = Nneighbours+1;
+
+    // 1st lon deriv weights
     var_to_write.resize( Npts*(Nneighbours+1) );
     for ( size_t II = 0; II < Npts; II++ ) { for ( size_t JJ = 0; JJ < Nneighbours+1; JJ++ ) {
             var_to_write.at( (Nneighbours+1)*II + JJ ) = adjacency_ddlon_weights[II][JJ];
@@ -405,10 +415,134 @@ void dataset::write_adjacency(
     write_field_to_output( var_to_write, "adjacency_ddlon_weights", 
             starts, counts, filename.c_str() );
 
+    // 1st lat deriv weights
     for ( size_t II = 0; II < Npts; II++ ) { for ( size_t JJ = 0; JJ < Nneighbours+1; JJ++ ) {
             var_to_write.at( (Nneighbours+1)*II + JJ ) = adjacency_ddlat_weights[II][JJ];
     } }
     write_field_to_output( var_to_write, "adjacency_ddlat_weights", 
             starts, counts, filename.c_str() );
+
+    // 2nd lon deriv weights
+    for ( size_t II = 0; II < Npts; II++ ) { for ( size_t JJ = 0; JJ < Nneighbours+1; JJ++ ) {
+            var_to_write.at( (Nneighbours+1)*II + JJ ) = adjacency_d2dlon2_weights[II][JJ];
+    } }
[…]

[…]
