### 2023-03-17 0e90617
LLC Helmholtz loads pre-computed adjacency data
