+++ Case_Files/LLC_Helmholtz_projections.cpp
@@ -183,8 +183,7 @@ int main(int argc, char *argv[]) {
     // Build the adjacency matrix and other adjacency-adjacent arrays
     //  down the road, just load in a pre-built one, but for right now
     //  this is easier.
-    source_data.build_adjacency();
-    //source_data.load_adjacency( adjacency_fname );
+    source_data.load_adjacency( adjacency_fname );
 
     // Read in the velocity fields
     source_data.load_variable( "u_lon", zonal_vel_name, input_fname, true, true );
