### 2023-03-17 b9b3897
Adjusting adjacency IO
