### 2023-03-17 bf2bd97
Adjusting fill values for double outputs
