### 2023-03-17 65b6f5e
Adding fill value for double outputs
