### 2023-03-17 c077c5d
Tidying up adjacency files a bit
