### 2023-03-24 cdcda86
Adding a local `constants.hpp` for each tutorial case
Re Issue #28
