+++ Tutorial/Spherical_Demo/Helmholtz/postprocessing/constants.hpp
@@ -1,0 +1,421 @@
+#ifndef CONSTANTS_HPP
+#define CONSTANTS_HPP 1
+
+#include <map>
+
+/*!
+ * \param DEBUG
+ * \brief Compile-time debug setting, specified in Makefile
+ *
+ * This is an integer than generally controls how much is printed
+ * during run-time. Higher values mean more print statements.
+ *
+ * @ingroup constants
+ */
+#ifndef DEBUG
+    #define DEBUG 1
+#endif
+
+/*!
+ * \file
+ * \brief Provide namespace for global constants (physical and computational).
+ *
+ * This header file provides a namespace to give a consistent source of constants.
+ *
+ * Usage:
+ * @code
+ * #include "constants.hpp"
+ * ...
+ * double R_earth = constants::R_earth;
+ * @endcode
+ */
+
+/*!
+ *  \addtogroup constants
+ *  @{
+ * \brief Provide namespace for global constants (physical and computational).
+ */
+
+/*! 
+ * \namespace constants 
+ * \brief Provide namespace for global constants (physical and computational).
+ */
+namespace constants
+{
+
+    /*!
+     * \param R_earth
+     * \brief Mean radius of the Earth
+     * @ingroup constants
+     */
+    const double R_earth = 6371e3;
+
+    /*!
+     * \param USE_HIGH_PRECISION_DISTANCE
+     * \brief If spatial resolution is very high on sphere (less than ~50 metres or so), 
+     * and high presision is needed in distance calculations, turn this on.
+     *
+     * The default spherical distance caluclation, spherical law of cosines, has floating point
+     * errors for small distances (a couple metres or so). However, so long as your grid is very high resolution,
+     * this shouldn't be an issue. This is particularly true if: you are also using a continuous kernel and the filtering scales
+     * themselves also aren't very short (couple of metres or so).
+     *
+     * The test routine (Tests/distance_formulas.cpp) uses both methods on a specified grid and outputs the result.
+     * If you think you might need the high precision scheme, test it out there first.
+     *
+     * @ingroup constants
+     */
+    const bool USE_HIGH_PRECISION_DISTANCE = false;
+
+    /*!
+     * \param rho0
+     * \brief Mean fluid density
+     * @ingroup constants
+     */
+    const double rho0 = 1025;
+
+    /*!
+     * \param g
+     * \brief (constant) acceleration due to gravity.
+     * @ingroup constants
+     */
+    const double g = 9.81;
+
+    /*!
+     * \param DiffOrd
+     * \brief Differentiation order for finite differencing (currently must be 2, 4, or 6).
+     * @ingroup constants
+     */
+    const int DiffOrd = 4;
+
+    /*!
+     * \param fill_value
+     * \brief Fill value used to indicate land values in output files.
+     * @ingroup constants
+     */
+    const double fill_value = -1e8;
+
+    /*!
+     * \param fill_value_s
+     * \brief Fill value used to indicate land values in output files (signed short)
+     * @ingroup constants
+     */
+    const signed short fill_value_s = -32767;
+
+    /*!
+     * \param DEFORM_AROUND_LAND
+     * \brief Boolean indicating whether or not the kernel should deform around land.
+     *
+     * If true, land is given zero weight (i.e. not factored into local average).
+     * If false, land is included in local average (treated as zero velocity
+     * @ingroup constants
+     */
+    const bool DEFORM_AROUND_LAND = false;
+
+    /*!
+     * \param FILTER_OVER_LAND
+     * \brief Boolean to indicate whether or not land values should be filled in with coarse-grained results.
+     *
[…]
