### 2023-03-24 cb890fe
Fixing issue that required a 'regions' file for scalar filtering
