### 2023-03-24 be2c290
Adding some usage expectation statements.
Re https://github.com/openjournals/joss-reviews/issues/4277#issuecomment-1475667430
