### 2023-03-24 #4277 @bastorer
Thanks for clarifying! I've added some notes to the documentation. Specifically, on the documentation landing page I added

> ## Usage
>
> * **Users** can expect to compile the executables in 'Case Files' and use them as command-line utilities to process existing netCDF-4 data. The tutorials illustrate the steps required for this usage, as well     as highlighting the kind of outputs / analysis that can be obtained.
> * **Developers** can use FlowSieve as a C++ library and develop additional diagnostic / analysis routines using the FlowSieve codebase.
