### 2023-03-25 husseinaluie/FlowSieve#29 @ashwinvis
40 processors is a bit much, so I have to get a supercomputer to test it. Would it be possible to reduce the requirements such that even with 8 processors you get something in a few minutes?
