### 2023-03-28 41f9cd3
More adjacency testing...
