+++ Functions/build_adjacency.cpp
@@ -55,7 +55,8 @@ void dataset::build_adjacency(
     //       sqrt( area per grid point ) ~ side length of grid, i.e. spacing between points
     // Is this perfect? Nope. But it'll get us in the ballpark
     const double    typical_spacing = sqrt( 4 * M_PI / num_pts ) * constants::R_earth,
-                    max_distance = 10 * typical_spacing;
+                    max_distance = 10 * typical_spacing,
+                    angle_per_neighbour = 360. / num_neighbours;
 
     std::vector<double> neighbour_x, neighbour_y, neighbour_dist;
     std::vector<size_t> neighbour_ind;
@@ -64,8 +65,11 @@ void dataset::build_adjacency(
     double pt_lon, pt_lat, search_lon, search_lat, poleward_lat, del_lon_lim,
            proj_x, proj_y, proj_theta, local_dist, denom, denom_sign, stencil_count,
            stencil_min_x, stencil_min_y, stencil_max_x, stencil_max_y;
-    bool near_pole;
+    bool near_pole, similar_angle;
     int Istencil, II, JJ, angle_ind;
+
+    double max_neighbour_dist = -1, ref_angle, curr_angle;
+    size_t furthest_neighbour_ind = 0 ;
 
     std::vector< double > xi, yi, ddx_coeffs, ddy_coeffs;
 
@@ -85,9 +89,10 @@ void dataset::build_adjacency(
              Ineighbour, neighbour_x, neighbour_y, neighbour_dist, neighbour_ind, stencil_set, \
              Istencil, II, JJ, xi, yi, denom, denom_sign, stencil_count, ddx_coeffs, ddy_coeffs, \
              stencil_min_x, stencil_min_y, stencil_max_x, stencil_max_y, \
-             LHS, report, alglib_info, LHS_vec, solve, LHS_list \
+             LHS, report, alglib_info, LHS_vec, solve, LHS_list, \
+             max_neighbour_dist, furthest_neighbour_ind, similar_angle, ref_angle, curr_angle \
            ) \
-    firstprivate( num_pts, max_distance, stdout, alglib::xdefault )
+    firstprivate( num_pts, max_distance, stdout, alglib::xdefault, angle_per_neighbour )
     {
         LHS_vec.resize( (num_neighbours+1)*(num_neighbours+1), 0. );
         LHS.setlength( num_neighbours+1, num_neighbours+1 );
@@ -109,6 +114,9 @@ void dataset::build_adjacency(
             neighbour_y.resize( num_neighbours, 0. );
             neighbour_dist.resize( num_neighbours, 1e100 );
             neighbour_ind.resize( num_neighbours, 0 );
+
+            max_neighbour_dist = 1e100;
+            furthest_neighbour_ind = 0;
 
             #if DEBUG > 0
             pt_lon = longitude.at( pt_index );
@@ -135,7 +143,7 @@ void dataset::build_adjacency(
                     continue;
                 }
 
-                // Next, a slightly less laxy pruning by longitude
+                // Next, a slightly less lazy pruning by longitude
                 poleward_lat = fabs(pt_lat) + (max_distance / constants::R_earth);
                 near_pole = poleward_lat >= (M_PI * 89.8 / 180);
                 del_lon_lim = std::fmin( M_PI, max_distance / ( cos(poleward_lat) * constants::R_earth ) );
@@ -145,14 +153,27 @@ void dataset::build_adjacency(
 
                 near_sided_project( proj_x, proj_y, search_lon, search_lat, pt_lon, pt_lat, 2e5 );
                 local_dist = distance( search_lon, search_lat, pt_lon, pt_lat );
-                if (local_dist < 1e-0) { continue; } // discard points that are erroneously close
+                if (local_dist < 1e-4) { continue; } // discard points that are erroneously close
[…]
