### 2023-03-28 ece2972
Testing out alternate second derivative schemes
