### 2023-03-31 8b679d5
Printing the number of string inputs along with the list
