### 2023-04-06 husseinaluie/FlowSieve#29 @bastorer
For something that runs in ~5 minutes on 8 processors, setting the number of vertical levels to 8 (last number on line 13 of `generate_data_sphere.py`, change from 48 to 8) should do the trick.
