### 2023-04-06 husseinaluie/FlowSieve#29 @ashwinvis
Managed to run with 8 vertical levels and 4 processors in a laptop for ~30 minutes.
