### 2023-04-06 #4277 @ashwinvis
@kthyng I have finished the review and I conclude that FlowSieve is a high-performance and potentially reusable tool for downscaling and / or analysing data from various geophysical applications. All the claims in the manuscript has been verified against the software implementation and I **recommend** that the article may be **accepted** into JOSS.

Congrats @bastorer
