+++ paper.md
@@ -14,9 +14,9 @@ authors:
     affiliation: 1
 
 affiliations:
- - name: University of Rochester, USA
+ - name: Department of Mechanical Engingeering, University of Rochester, USA
    index: 1
-date: 8 March 2022
+date: 11 April 2023
 bibliography: paper.bib
 
 ---
