+++ README.md
@@ -1,6 +1,7 @@
 # FlowSieve
 
 [![Documentation Status](https://readthedocs.org/projects/flowsieve/badge/?version=latest)](https://flowsieve.readthedocs.io/en/latest/?badge=latest)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7818192.svg)](https://doi.org/10.5281/zenodo.7818192)
 
 # About FlowSieve
 
