### 2023-04-11 #4277 @bastorer
# Additional Author Tasks After Review is Complete
  * Release version set to **v3.3.0**
  * DOI: 10.5281/zenodo.7818192    ( https://zenodo.org/record/7818192#.ZDV61S3MKwI )

@kthyng Great thanks! I think I've finished everything on my list :-)
