### 2023-04-11 #4277 @bastorer
> @bastorer it's not required, but you might consider creating a release or tag for your software on Github to match your official version.

Good idea. I've added a release + tag pointing to the Zenodo version
