### 2023-04-11 #4277 @bastorer
> Oops, I didn't even look carefully at the results of my editorialbot command!!

Looks like timings worked out perfect, the paper at the DOI destination has the correct reference!
