### 2023-06-08 #4277 @bastorer
Hi @kthyng

Sorry to resurrect this issue, but my PI asked me to include some additional funding acknowledgements in the JOSS paper (similar to https://github.com/openjournals/joss/issues/485). I've updated the paper.md file (relevant commit: https://github.com/husseinaluie/FlowSieve/commit/00bcbe0c47a043b464f67b71f27b56ad8ed95ccc).

Is it possible to update the JOSS version of the paper with the additional funding acknowledgements?

Thanks!
Ben
